% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{add_column_to_dataframe}
\alias{add_column_to_dataframe}
\title{Add a column to a data frame}
\usage{
add_column_to_dataframe(df, position, column_to_insert)
}
\arguments{
\item{df}{data frame to add the column to}

\item{position}{location in data frame where to insert the column.
Position can also be a name of a column}

\item{column_to_insert}{the elements of the column to insert. If
the column has a name, this name will be copied into the data frame.
Id is does not have a name, it will get the name "nc".}
}
\value{
A data frame with the column inserted
}
\description{
Add a column to a data frame
}
