% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_data.R
\docType{data}
\name{archipelago_data}
\alias{archipelago_data}
\title{data on 41 archipelagos}
\format{
A dataframe containing information on archipelago name, area, age and distance from the mainland
}
\description{
A dataframe with in subsequent columns the name of the archipelago (Archipelago)
the area of the archipelago (Area), the age (Age) and the distance from the mainland (Distance)
}
\keyword{datasets}
