% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_data.R
\docType{data}
\name{archipelagos41}
\alias{archipelagos41}
\title{DAISIE datalist object including bird phylogenetic data and
physical data for 41 archipelagos}
\format{
A datalist containing data on the 41 archipelagos studied in Valente et
al 2020 (Main Dataset D1). Contains colonisation and branching times for bird species
in each of the archipelagos. It also contains information on archipelago name, area,
age and distance from the nearest mainland.
}
\source{
Valente L, Phillimore AB, Melo M, Warren BH, Clegg SM, Havenstein K,
 Tiedemann R, Illera JC, Thébaud C, Aschenbach T, Etienne RS. A simple dynamic model
 explains island bird diversity worldwide (2020) Nature, 579, 92-96 \cr
}
\description{
A datalist with 41 items representing the 41 archipelagos.
Each archipelago can be called separately using archipelagos41[[x]] with x being a
number between 1 and 41. Using archipelagos41[[x]][[1]] will show just the
top part of the archipelago item where the archipelago name and physical features
are displayed. The structure of each of the archipelagos is the same as regular DAISIE
datalist generated using DAISIE_dataprep.
}
\keyword{datasets}
