% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_area_params.R
\name{are_area_pars}
\alias{are_area_pars}
\title{Test if list has area parameters}
\usage{
are_area_pars(area_pars)
}
\arguments{
\item{area_pars}{object to be tested if conforms to area parameters}
}
\value{
Boolean that indicates if list conforms to expected area parameters
as created by \link{create_area_pars}
}
\description{
Test if list has area parameters
}
\examples{
testit::assert(DAISIE::are_area_pars(
  create_area_pars(
    max_area = 10,
    proportional_peak_t = 0.5,
    peak_sharpness = 1,
    total_island_age = 5)) == TRUE)
}
