% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{are_constant_rate_area_pars}
\alias{are_constant_rate_area_pars}
\title{Test if list has area parameters}
\usage{
are_constant_rate_area_pars(area_pars)
}
\arguments{
\item{area_pars}{object to be tested if conforms to area parameters}
}
\value{
Boolean that indicates if list conforms to expected area parameters
as created by \link{create_area_pars}
}
\description{
Test if list has area parameters
}
\examples{
testit::assert(DAISIE::are_constant_rate_area_pars(area_pars = 1))
}
\author{
Richel J.C Bilderbeek, Joshua Lambert, Pedro Neves
}
