% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_check_format.R
\name{are_DAISIE_create_sim_pars}
\alias{are_DAISIE_create_sim_pars}
\title{Checks if parameters are valid}
\usage{
are_DAISIE_create_sim_pars(pars)
}
\arguments{
\item{pars}{DAISIE simulation parameters in the form of a named list.}
}
\value{
A boolean stating whether checks are TRUE
}
\description{
Checks if parameters are valid
}
