% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{are_hyper_pars}
\alias{are_hyper_pars}
\title{Test if a list has hyperparameters}
\usage{
are_hyper_pars(hyper_pars)
}
\arguments{
\item{hyper_pars}{A named list of numeric hyperparameters for the rate
calculations as returned by \code{\link{create_hyper_pars}()}:
\itemize{
  \item{[1]: is d the scaling parameter for exponent for calculating
  cladogenesis rate}
  \item{[2]: is x the exponent for calculating extinction rate}
}}
}
\value{
\code{TRUE} if list contains hyperparameters, \code{FALSE} otherwise.
}
\description{
Test if a list has hyperparameters
}
\examples{
testit::assert(
  DAISIE:::are_hyper_pars(create_hyper_pars(d = 0.027, x = 0.15))
)
}
\author{
Pedro Neves, Joshua Lambert
}
\keyword{internal}
