% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_check_format.R
\name{are_max_rates}
\alias{are_max_rates}
\title{Determines if object x are rates}
\usage{
are_max_rates(x)
}
\arguments{
\item{x}{object to be determined to be rates}
}
\value{
\code{TRUE} if object x is a list of rates, \code{FALSE} otherwise.
}
\description{
Determines if object x are rates
}
