% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_check_format.R
\name{are_rates}
\alias{are_rates}
\title{Determines if object rates are rates}
\usage{
are_rates(rates)
}
\arguments{
\item{rates}{named list of numeric rates as returned by
\code{\link{update_rates}}.}
}
\value{
TRUE if object rates is a list of rates
}
\description{
Determines if object rates are rates
}
