% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_area_params.R
\name{are_sea_level_pars}
\alias{are_sea_level_pars}
\title{Test if list has sea level parameters}
\usage{
are_sea_level_pars(sea_level_pars)
}
\arguments{
\item{sea_level_pars}{object to be tested if conforms to area parameters}
}
\value{
Boolean that indicates if list conforms to expected sea level
parameters as created by \link{create_sea_level_pars}
}
\description{
Test if list has sea level parameters
}
\examples{
testit::assert(DAISIE::are_sea_level_pars(
  create_sea_level_pars(
    amplitude = 10,
    frequency = 0.5,
    phase = 1)) == TRUE)
}
