% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trait_state_params.R
\name{are_trait_state_params}
\alias{are_trait_state_params}
\title{Test if list has trait state parameters}
\usage{
are_trait_state_params(trait_state_params)
}
\arguments{
\item{trait_state_params}{object to be tested if conforms to trait state parameters}
}
\value{
Boolean that indicates if list conforms to expected area parameters
as created by \link{create_trait_state_params}
}
\description{
Test if list has trait state parameters
}
\examples{
testit::assert(DAISIE:::are_trait_state_params(
  create_trait_state_params(
    trans_rate = 0.5,
    immig_rate2 = 0.1,
    ext_rate2 = 0.2,
    ana_rate2 = 0.3,
    clado_rate2 = 0.4,
    trans_rate2 = 0.5,
    M2 = 1000)) == TRUE)
}
