% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{calc_Abeta}
\alias{calc_Abeta}
\title{Calculates the area at a point in time from a beta function}
\usage{
calc_Abeta(proptime, proptime_max, peak, Amax)
}
\arguments{
\item{Amax}{A numeric value for maximum island area during the simulation.}
}
\value{
Numeric
}
\description{
Calculates the area at a point in time from a beta function
}
