% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{calc_next_timeval_shift}
\alias{calc_next_timeval_shift}
\title{Calculates when the next timestep will be, and if a shift has occured.}
\usage{
calc_next_timeval_shift(max_rates, timeval, dynamic_shift_times, total_time)
}
\arguments{
\item{max_rates}{named list of max rates as returned by
\code{\link{update_rates}}.}

\item{timeval}{current time of simulation}

\item{dynamic_shift_times}{numeric vector of times of rate shifts.}
}
\value{
named list with numeric vector containing the time of the next
timestep and the change in time.
}
\description{
Calculates when the next timestep will be, and if a shift has occured.
}
\author{
Joshua Lambert, Pedro Neves, Shu Xie
}
\keyword{internal}
