% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_compare.R
\name{compare_diff}
\alias{compare_diff}
\title{Comparing simulated results from DAISIE and TRASIE}
\usage{
compare_diff(DAISIE_output, TRASIE_output1, TRASIE_output2, TRASIE_output3,
  Tpars1, Tpars2, Tpars3)
}
\arguments{
\item{DAISIE_output}{List with simulation output from general DAISIE
\code{DAISIE_sim}}

\item{TRASIE_output1}{List with simulation output using TRASIE pars
\code{DAISIE_sim}}

\item{TRASIE_output2}{List with simulation output using TRASIE pars
\code{DAISIE_sim}}

\item{TRASIE_output3}{List with simulation output using TRASIE pars
\code{DAISIE_sim}}

\item{Tpars1}{A named list containing diversification rates considering two trait states:
\itemize{
  \item{[1]:A numeric with the per capita transition rate with state1}
  \item{[2]:A numeric with the per capita immigration rate with state2}
  \item{[3]:A numeric with the per capita extinction rate with state2}
  \item{[4]:A numeric with the per capita anagenesis rate with state2}
  \item{[5]:A numeric with the per capita cladogenesis rate with state2}
  \item{[6]:A numeric with the per capita transition rate with state2}
  \item{[7]:A numeric with the number of species with trait state 2 on mainland}
}}

\item{Tpars2}{A named list containing diversification rates considering two trait states:}

\item{Tpars3}{A named list containing diversification rates considering two trait states:}
}
\description{
Comparing simulated results from DAISIE and TRASIE
}
\seealso{
\code{\link{DAISIE_sim}},
  \code{\link{DAISIE_convert_to_classic_plot}}

Other plotting: \code{\link{DAISIE_plot_comparison_stts}}
}
\concept{plotting}
