% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{countspecies}
\alias{countspecies}
\title{Count the number of species}
\usage{
countspecies(datalistelement)
}
\arguments{
\item{datalistelement}{something}
}
\value{
A numeric value
}
\description{
Count the number of species
}
\keyword{internal}
