% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{counttype1}
\alias{counttype1}
\title{Counts the number of type 1 species}
\usage{
counttype1(datalistelement)
}
\arguments{
\item{datalistelement}{stub}
}
\value{
something
}
\description{
Counts the number of type 1 species
}
