% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{create_daisie_params}
\alias{create_daisie_params}
\title{Create a full-blown DAISIE parameter structure}
\usage{
create_daisie_params(time, M, pars, replicates)
}
\arguments{
\item{time}{something}

\item{M}{something}

\item{pars}{something}

\item{replicates}{something}
}
\description{
Create a full-blown DAISIE parameter structure
}
