% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_dist_pars}
\alias{create_dist_pars}
\title{Create list of distance pars}
\usage{
create_dist_pars(D)
}
\arguments{
\item{dist_pars}{A numeric vector for distance parameters for the rate
calculations:
\itemize{
\item{[1]: is D distance from the mainland}
}}
}
\value{
Named list with distance parameters
}
\description{
Create list of distance pars
}
\examples{
dist_pars <- create_dist_pars(D = 2500)
}
