% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_CS_full_stt.R
\name{create_full_CS_stt}
\alias{create_full_CS_stt}
\title{Unsampled CS full STT}
\usage{
create_full_CS_stt(stt_list, stac_vec, totaltime, trait_pars = NULL)
}
\arguments{
\item{stt_list}{List of full stt tables as
returned by DAISIE_sim_core functions}

\item{stac_vec}{Vector with status of species on island.}

\item{totaltime}{Numeric double with total time of simulation.}

\item{trait_pars}{A named list containing diversification rates considering 
two trait states created by \code{\link{create_trait_pars}}:
\itemize{
  \item{[1]:A numeric with the per capita transition rate with state1}
  \item{[2]:A numeric with the per capita immigration rate with state2}
  \item{[3]:A numeric with the per capita extinction rate with state2}
  \item{[4]:A numeric with the per capita anagenesis rate with state2}
  \item{[5]:A numeric with the per capita cladogenesis rate with state2}
  \item{[6]:A numeric with the per capita transition rate with state2} 
  \item{[7]:A numeric with the number of species with trait state 2 on mainland} 
}}
}
\value{
1 complete, unsampled STT table from all clades in an island of a
CS model as generated by DAISIE_sim_core functions.
}
\description{
Unsampled CS full STT
}
\author{
Pedro Neves, Joshua Lambert, Shu Xie, Giovanni Laudanno
}
