% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_CS_full_stt_SR.R
\name{create_full_CS_stt_SR}
\alias{create_full_CS_stt_SR}
\title{Unsampled CS full STT}
\usage{
create_full_CS_stt_SR(stt_list, stac_vec, totaltime)
}
\arguments{
\item{stt_list}{List of full stt tables as
returned by DAISIE_sim_core functions}

\item{stac_vec}{Vector with status of species on island.}

\item{totaltime}{Numeric double with total time of simulation.}
}
\value{
1 complete, unsampled STT table from all clades in an island of a
CS model as generated by DAISIE_sim_core functions.
}
\description{
Unsampled CS full STT
}
\author{
Pedro Neves, Joshua Lambert, Shu Xie, Giovanni Laudanno
}
