% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_hyper_pars}
\alias{create_hyper_pars}
\title{Create list of hyperparameters}
\usage{
create_hyper_pars(d_0, x, alpha, beta)
}
\arguments{
\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations:
\itemize{
\item{[1]: is d_0 the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: is x the exponent for calculating extinction rate}
\item{[3]: is alpha, the exponent for calculating the immigration rate}
\item{[4]: is beta the exponent for calculating the anagenesis rate.}
}}
}
\value{
Named list with hyperparameters
}
\description{
Create list of hyperparameters
}
\examples{
hyper_pars <- create_hyper_pars(d_0 = 0.027, x = 0.15, 0.294, 0.383)
}
