% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_pars}
\alias{create_pars}
\title{Create vector of model parameters}
\usage{
create_pars(clado_rate, ext_rate, carr_cap, immig_rate, ana_rate)
}
\arguments{
\item{clado_rate}{Numeric rate of cladogenesis}

\item{ext_rate}{Numeric rate of extinction}

\item{carr_cap}{Numeric carrying capacity}

\item{immig_rate}{Numeric rate of immigration}

\item{ana_rate}{Numeric rate of anagenesis}
}
\value{
Numeric vector with 5 elements
}
\description{
Create vector of model parameters
}
\examples{
create_pars(clado_rate = 1,
            ext_rate = 1,
            carr_cap = 10,
            immig_rate = 0.1,
            ana_rate = 1)
}
