% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_relaxed_pars.R
\name{create_relaxed_pars}
\alias{create_relaxed_pars}
\title{Create named list of parameters for sampling parameters}
\usage{
create_relaxed_pars(
  clado_rate_scale = NULL,
  ext_rate_scale = NULL,
  carry_cap_scale = NULL,
  immig_rate_scale = NULL,
  ana_rate_scale = NULL,
  clado_rate_shape = 1,
  ext_rate_shape = 1,
  carry_cap_shape = 1,
  immig_rate_shape = 1,
  ana_rate_shape = 1
)
}
\arguments{
\item{clado_rate_scale}{scale parameter for a gamma distribution for
cladogenesis rate}

\item{ext_rate_scale}{scale parameter for a gamma distribution for
extinction rate}

\item{carry_cap_scale}{scale parameter for a gamma distribution for
the carrying capacity}

\item{immig_rate_scale}{scale parameter for a gamma distribution for
immigration rate}

\item{ana_rate_scale}{scale parameter for a gamma distribution for
anagenesis rate}

\item{clado_rate_shape}{shape parameter for a gamma distribution for
cladogenesis rate}

\item{ext_rate_shape}{shape parameter for a gamma distribution for
extinction rate}

\item{carry_cap_shape}{shape parameter for a gamma distribution for
the carrying capacity}

\item{immig_rate_shape}{shape parameter for a gamma distribution for
immigration rate}

\item{ana_rate_shape}{shape parameter for a gamma distribution for
anagenesis rate}
}
\value{
named list of numerical values containing scale and shape parameters
for sampling parameters
}
\description{
Create named list of parameters for sampling parameters
}
