% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{create_relaxed_rate_pars}
\alias{create_relaxed_rate_pars}
\title{Creates the list required for the relaxed-rate parameters for a DAISIE
simulation using \code{DAISIE_sim}}
\usage{
create_relaxed_rate_pars(mean, sd)
}
\arguments{
\item{mean}{The mean of the gamma distribution}

\item{sd}{The standard deviation (sd) of the gamma distribution}
}
\value{
A list of two elements
}
\description{
Creates the list required for the relaxed-rate parameters for a DAISIE
simulation using \code{DAISIE_sim}
}
