% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_area_params.R
\name{create_sea_level_pars}
\alias{create_sea_level_pars}
\title{Create named list of sea-level parameters}
\usage{
create_sea_level_pars(amplitude, frequency, phase)
}
\arguments{
\item{amplitude}{the amplitude of the sine function}

\item{frequency}{the frequency of the sine function}

\item{phase}{the phase of the sine function}
}
\value{
list of numerical values containing parameters
for sea-level simulation
}
\description{
Create named list of sea-level parameters
}
\examples{
create_sea_level_pars(
  amplitude = 50,
  frequency = 2,
  phase = 0)
}
