% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{create_singleton_phylo}
\alias{create_singleton_phylo}
\title{Create an empty phylogeny}
\usage{
create_singleton_phylo(age)
}
\arguments{
\item{age}{Time of phylogeny}
}
\description{
Create an empty phylogeny
}
\author{
Giovanni Laudanno
}
