% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{create_test_daisie_params}
\alias{create_test_daisie_params}
\title{Create a sunction to test full-blown DAISIE parameter structure}
\usage{
create_test_daisie_params()
}
\description{
Create a sunction to test full-blown DAISIE parameter structure
}
