% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_simulation_output.R
\name{create_test_simulation_outputs}
\alias{create_test_simulation_outputs}
\title{Creates the output of a as-short-as-possible
simulation run}
\usage{
create_test_simulation_outputs(island_ontogeny = NULL)
}
\arguments{
\item{island_ontogeny}{Boolean indicating if an ontogeny or
no ontogeny scenario should be generated. Default is \code{FALSE}.}
}
\value{
simulation_outputs
}
\description{
Creates the output of a as-short-as-possible
simulation run
}
\examples{
\donttest{
standard_simulation <- create_test_simulation_outputs()

island_ontogeny_simulation <- create_test_simulation_outputs(
  island_ontogeny = "beta"
)

}
}
\author{
Pedro Neves, Richel J.C. Bilderbeek
}
