% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_trait_pars}
\alias{create_trait_pars}
\title{Create named list of trait state parameters}
\usage{
create_trait_pars(
  trans_rate,
  immig_rate2,
  ext_rate2,
  ana_rate2,
  clado_rate2,
  trans_rate2,
  M2
)
}
\arguments{
\item{trans_rate}{A numeric with the per capita transition rate with state1}

\item{immig_rate2}{A numeric with the per capita immigration rate with state2}

\item{ext_rate2}{A numeric with the per capita extinction rate with state2}

\item{ana_rate2}{A numeric with the per capita anagenesis rate with state2}

\item{clado_rate2}{A numeric with the per capita cladogenesis rate with state2}

\item{trans_rate2}{A numeric with the per capita transition rate with state2}

\item{M2}{A numeric with the number of species with trait state 2 on mainland}
}
\value{
list of numerical values containing trait state parameters
}
\description{
Create named list of trait state parameters
}
\examples{
testit::assert(DAISIE:::are_trait_pars(
  create_trait_pars(
    trans_rate = 0.5,
    immig_rate2 = 0.1,
    ext_rate2 = 0.2,
    ana_rate2 = 0.3,
    clado_rate2 = 0.4,
    trans_rate2 = 0.5,
    M2 = 1000)) == TRUE)
}
