% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_integrate.R
\name{cum_rho}
\alias{cum_rho}
\title{Cumulative Gamma distribution parameterised with mean and standard deviation}
\usage{
cum_rho(par_upper_bound, DAISIE_dist_pars)
}
\arguments{
\item{DAISIE_dist_pars}{A numeric vector of two elements, first is the mean
and second the standard deviation of the distribution.}
}
\value{
Numeric
}
\description{
Cumulative Gamma distribution parameterised with mean and standard deviation
}
\keyword{internal}
