% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_CS.R
\name{DAISIE_abm_factor}
\alias{DAISIE_abm_factor}
\title{Sets or retrieves the factor to calculate the step-size used by the odeint::adams_bashforth[_moulton] solvers.}
\usage{
DAISIE_abm_factor(factor)
}
\arguments{
\item{steps}{\code{factor}: sets step-size to \code{factor * (t1 - t0)}. \cr}
}
\value{
current factor
}
\description{
Sets or retrieves the factor to calculate the step-size used by the odeint::adams_bashforth[_moulton] solvers.
}
