% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_half_life.R
\name{DAISIE_avg_half_life}
\alias{DAISIE_avg_half_life}
\title{Measures relaxation average half-life from output 
from DAISIE_sim}
\usage{
DAISIE_avg_half_life(sim)
}
\arguments{
\item{sim}{output from DAISIE_sim}
}
\value{
an average half life of the island
}
\description{
Measures relaxation average half-life from output 
from DAISIE_sim
}
\author{
Joshua Lambert
}
