% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{DAISIE_calc_clade_ext_rate}
\alias{DAISIE_calc_clade_ext_rate}
\title{Calculate the clade-wide extinction rate}
\usage{
DAISIE_calc_clade_ext_rate(ps_ext_rate, n_species)
}
\arguments{
\item{ps_ext_rate}{per species extinction rate}

\item{n_species}{number of species in that clade}
}
\value{
the clade's extinction rate
}
\description{
Calculate the clade-wide extinction rate
}
\author{
Richel J.C. Bilderbeek
}
