% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_calc_half_life.R
\name{DAISIE_calc_half_life}
\alias{DAISIE_calc_half_life}
\title{Calculates the exact half-life from a DAISIE simulation that has reached
growth or decayed to half way between the initial conditions and the 
equilibrium}
\usage{
DAISIE_calc_half_life(island_replicates, pars, divdepmodel)
}
\arguments{
\item{island_replicates}{Island replicates in DAISIE format (produced in
\code{\link{DAISIE_sim}} with \code{format = TRUE} option). Minimally, this must be 
a list, that has as much elements as replicates. Each element must be a
list with the elements \code{island_age}, \code{not_present} 
and \code{stt_all}. \code{stt_all} must be a data frame with
the column names \code{Time}, \code{nI}, \code{nA}, \code{nC} 
and \code{present}.}

\item{divdepmodel}{Option divdepmodel = 'CS' runs a model with clade-specific 
carrying capacity, where diversity-dependence operates only within single 
clades, i.e. only among species originating from the same mainland colonist.
Option divdepmodel = 'IW' runs a model with island-wide carrying capacity,
where diversity-dependence operates within and among clades.}
}
\value{
a list of half-lives
}
\description{
Calculates the exact half-life from a DAISIE simulation that has reached
growth or decayed to half way between the initial conditions and the 
equilibrium
}
