% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_calc_half_life.R
\name{DAISIE_calc_half_life_exp_model}
\alias{DAISIE_calc_half_life_exp_model}
\title{Calculates the half-life from a DAISIE simulation using the exponential 
model of Diamond (1972)}
\usage{
DAISIE_calc_half_life_exp_model(island_replicates, time, pars)
}
\arguments{
\item{island_replicates}{output from DAISIE_sim_core}

\item{time}{time for simulation}

\item{pars}{the parameters used for the simulation}
}
\value{
a list of half-lives
}
\description{
Calculates the half-life from a DAISIE simulation using the exponential 
model of Diamond (1972)
}
