% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_create_island_mainland_ex.R
\name{DAISIE_create_island_mainland_ex}
\alias{DAISIE_create_island_mainland_ex}
\title{Converts simulation output into island output}
\usage{
DAISIE_create_island_mainland_ex(stt_table, totaltime, island_spec, mainland)
}
\arguments{
\item{stt_table}{Matrix with number of species at each time step.}

\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
list with the island information, composed stt table,
branching times of extant species, status of species on
the island and number of missing species.
}
\description{
Converts simulation output into island output
}
\keyword{internal}
