% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_CS.R
\name{DAISIE_CS_max_steps}
\alias{DAISIE_CS_max_steps}
\title{CS iteration control}
\usage{
DAISIE_CS_max_steps(max_steps)
}
\arguments{
\item{max_steps}{\code{max_steps}: sets max. iterations to \code{max_steps}. \cr}
}
\value{
current max. iterations
}
\description{
Sets or retrieves the max. number of iterations used by the odeint solver.
}
