% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpEC.R
\name{DAISIE_DE_logpEC}
\alias{DAISIE_DE_logpEC}
\title{Function to calculate the likelihood of observing an endemic lineage
with the colonization time at t1.}
\usage{
DAISIE_DE_logpEC(brts, missnumspec, pars1, methode, reltolint, abstolint)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic lineage
with the colonization time at t1.
\item{logLkb}{ The log-likelihood value computed based on a system of differential equations. }
}
\description{
This function calculates the log-likelihood of observing an endemic lineage
with the colonization time at t1.
}
\examples{

# Select a dataset from a DAISIE package

data(Galapagos_datalist)
datalist <- Galapagos_datalist

# Select an endemic lineage in the dataset
i <- 4
# Define example parameters
pars1 <- c(0.2, 0.1, 0.05, 0.02, 0.03)

# choose the method to solve the system of differential equations
log_likelihood <- DAISIE_DE_logpEC(brts, missnumspec, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)

print(log_likelihood)

}
