% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpEC_mainland.R
\name{DAISIE_DE_logpEC_mainland}
\alias{DAISIE_DE_logpEC_mainland}
\title{Function to calculate the likelihood of observing an endemic lineage on the island
that coexists with its mainland ancestors.}
\usage{
DAISIE_DE_logpEC_mainland(datalist, i, pars1, methode, reltolint, abstolint)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic lineage
that coexists with its mainland ancestors.
\item{logLkb}{ The log-likelihood value is computed based on a system of differential equations.}
}
\description{
This function calculates the log-likelihood of observing an endemic lineage on an island
for which the exact colonization time is known, and that coexists with its mainland ancestors.
}
