% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpEC_max_age_coltime.R
\name{DAISIE_DE_logpEC_max_age_coltime}
\alias{DAISIE_DE_logpEC_max_age_coltime}
\title{Function to calculate the likelihood of observing an endemic lineage on the island
with information on the maximum age of colonization}
\usage{
DAISIE_DE_logpEC_max_age_coltime(
  datalist,
  i,
  pars1,
  methode,
  reltolint,
  abstolint
)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic lineage
with its mainland ancestors
\item{logLkb}{ The log-likelihood value computed based on a system of differential equations.}
}
\description{
This function calculates the log-likelihood of observing an endemic lineage on an island
for which the exact colonization time is unknown, but the maximum age of colonization is known.
}
