% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpEC_max_age_coltime_and_mainland.R
\name{DAISIE_DE_logpEC_max_age_coltime_and_mainland}
\alias{DAISIE_DE_logpEC_max_age_coltime_and_mainland}
\title{Function to calculate the likelihood of observing an endemic lineage on the island
with maximum age of colonization, and that coexists with its mainland ancestors.}
\usage{
DAISIE_DE_logpEC_max_age_coltime_and_mainland(
  brts,
  missnumspec,
  pars1,
  methode,
  reltolint,
  abstolint
)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic lineage
with maximum age of colonization that coexists with its mainland ancestors
\item{logLkb}{ The log-likelihood value is computed based on a system of differential equations.}
}
\description{
This function calculates the log-likelihood of observing an endemic lineage on an island
with maximum age of colonization, and that coexists with its mainland ancestors.
}
