% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpEC_unknown_coltime.R
\name{DAISIE_DE_logpEC_unknown_coltime}
\alias{DAISIE_DE_logpEC_unknown_coltime}
\title{Function to calculate the likelihood of observing an endemic lineage on the island
with unknown colonization time}
\usage{
DAISIE_DE_logpEC_unknown_coltime(brts, missnumspec, pars1, methode, rtol, atol)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic lineage
with an unknown colonization time
\item{logLkb}{ The log-likelihood value computed based on a system of differential equations.}
}
\description{
This function calculates the log-likelihood of observing an endemic lineage on an island
for which the exact colonization time is unknown.
}
