% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpES.R
\name{DAISIE_DE_logpES}
\alias{DAISIE_DE_logpES}
\title{Function to calculate the likelihood of observing an endemic singleton lineage
with the colonization time at t1.}
\usage{
DAISIE_DE_logpES(datalist, i, pars1, methode, reltolint, abstolint)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic singleton lineage
with the colonization time at t1.
\item{logL1b}{ The log-likelihood value computed based on a system of differential equations. }
}
\description{
This function calculates the log-likelihood of observing an endemic singleton lineage
with the colonization time at t1.
}
\examples{

# Select a dataset from a DAISIE package

data(Galapagos_datalist)
datalist <- Galapagos_datalist

# Select an endemic singleton lineage in the dataset
i <- 6
# Define example parameters
pars1 <- c(0.2, 0.1, 0.05, 0.02, 0.03)

# choose the method to solve the system of differential equations
log_likelihood <- DAISIE_DE_logpES(datalist, i, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)

print(log_likelihood)

}
