% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpES_mainland.R
\name{DAISIE_DE_logpES_mainland}
\alias{DAISIE_DE_logpES_mainland}
\title{Function to calculate the likelihood of observing an endemic singleton lineage on the island
that coexists with its mainland ancestor.}
\usage{
<<<<<<< HEAD
DAISIE_DE_logpES_mainland(
  brts,
  missnumspec,
  pars1,
  methode,
  reltolint,
  abstolint
)
=======
DAISIE_DE_logpES_mainland(datalist, i, pars1, methode, reltolint, abstolint)
>>>>>>> d0885cf4232f775a0460a4d70d01bbca65d98758
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic singleton lineage
on the island with its mainland ancestor.
\item{logL1b}{ The log-likelihood value computed based on a system of differential equations.}
}
\description{
This function calculates the log-likelihood of observing an endemic singleton lineage on an island
 that coexists with its mainland ancestor
}
