% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpES_max_age_coltime_and_mainland.R
\name{DAISIE_DE_logpES_max_age_coltime_and_mainland}
\alias{DAISIE_DE_logpES_max_age_coltime_and_mainland}
\title{Function to calculate the likelihood of observing an endemic singleton lineage on the island
with maximum age of colonization, and that coexists on the island with its mainland ancestor.}
\usage{
DAISIE_DE_logpES_max_age_coltime_and_mainland(
  datalist,
  i,
  pars1,
  methode,
  reltolint,
  abstolint
)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic singleton lineage
with maximum age of colonization, and its mainland ancestor.
\item{logL1b}{ The log-likelihood value computed based on a system of differential equations.}
}
\description{
This function calculates the log-likelihood of observing an endemic singleton lineage on an island
for which the exact colonization time is unknown, but the maximum of colonization is known, and that coexists on the island
with its mainland ancestors
}
