% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpES_max_min_age_coltime.R
\name{DAISIE_DE_logpES_max_min_age_coltime}
\alias{DAISIE_DE_logpES_max_min_age_coltime}
\title{Function to compute the likelihood of observing an endemic singleton lineage
on the island given the minimum and maximum colonization ages.}
\usage{
DAISIE_DE_logpES_max_min_age_coltime(
  datalist,
  i,
  pars1,
  methode,
  reltolint,
  abstolint
)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic singleton lineage
for which the minimum and maximum ages of colonization are given.
\item{logL1b}{ The log-likelihood value computed based on a system of differential equations.}
}
\description{
This function calculates the log-likelihood of observing an endemic singleton lineage on an island
for which the exact colonization time is unknown, but the maximum and minimum ages of colonization are given.
}
