% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpNE.R
\name{DAISIE_DE_logpNE}
\alias{DAISIE_DE_logpNE}
\title{Function to calculate the likelihood of observing a non-endemic lineage
with the colonization time at t1.}
\usage{
<<<<<<< HEAD
DAISIE_DE_logpNE(brts, missnumspec, pars1, methode, reltolint, abstolint)
=======
DAISIE_DE_logpNE(datalist, i, pars1, methode, reltolint, abstolint)
>>>>>>> d0885cf4232f775a0460a4d70d01bbca65d98758
}
\value{
The output is a numeric value representing the log-likelihood of observing a non-endemic lineage
with the colonization time at t1.
\item{logLMb}{ The log-likelihood value computed based on the differential equation system. }
}
\description{
This function calculates the log-likelihood of observing a non-endemic lineage
with the colonization time at t1.
}
\examples{

# Select a dataset from a DAISIE package

data(Galapagos_datalist)
datalist <- Galapagos_datalist

# Select a non-endemic lineage in the dataset
i <- 3
# Define example parameters
pars1 <- c(0.2, 0.1, 0.05, 0.02, 0.03)

# choose the method to solve the system of differential equations
<<<<<<< HEAD
log_likelihood <- DAISIE_DE_logpNE(brts, missnumspec, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)
=======
log_likelihood <- DAISIE_DE_logpNE(datalist, i, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)
>>>>>>> d0885cf4232f775a0460a4d70d01bbca65d98758

print(log_likelihood)

}
