% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpNE_max_min_age_coltime.R
\name{DAISIE_DE_logpNE_max_min_age_coltime}
\alias{DAISIE_DE_logpNE_max_min_age_coltime}
\title{Function to calculate the likelihood of observing a non-endemic lineage on the island
with minimum and maximum ages of colonization}
\usage{
DAISIE_DE_logpNE_max_min_age_coltime(
  datalist,
  i,
  pars1,
  methode,
  reltolint,
  abstolint
)
}
\value{
The output is a numeric value representing the log-likelihood of observing a non-endemic singleton lineage
for which the minimum and maximum ages of colonization are given.
\item{logL1b}{ The log-likelihood value computed based on the differential equation system.}
}
\description{
This function calculates the log-likelihood of observing a non-endemic lineage on an island
for which the exact colonization time is unknown, but the maximum and minimum ages of colonization are known.
}
