% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpNE_unknown_coltime.R
\name{DAISIE_DE_logpNE_unknown_coltime}
\alias{DAISIE_DE_logpNE_unknown_coltime}
\title{Function to calculate the likelihood of observing a non-endemic lineage on the island
with unknown colonization time.}
\usage{
DAISIE_DE_logpNE_unknown_coltime(brts, pars1, methode, reltolint, abstolint)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic singleton lineage
with unknown colonization time.
\item{logL0b}{ The log-likelihood value computed based on system of differential equations.}
#' @examples

# Select a dataset from a DAISIE package

data(Galapagos_datalist)
datalist <- Galapagos_datalist

# Select a non-endemic lineage in the dataset
i <- 2
# Define example parameters
pars1 <- c(0.2, 0.1, 0.05, 0.02, 0.03)

# choose the method to solve the system of differential equations
log_likelihood <- DAISIE_DE_logpNE(brts, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)

print(log_likelihood)
}
\description{
This function calculates the log-likelihood of observing a non-endemic lineage on an island
for which the exact colonization time is unknown.
}
