% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_calc_half_life.R
\name{DAISIE_estimate_half_life}
\alias{DAISIE_estimate_half_life}
\title{Estimates half-life based on a DAISIE simulation}
\usage{
DAISIE_estimate_half_life(simulation, pars)
}
\arguments{
\item{simulation}{DAISIE_sim output}

\item{pars}{Numeric vector with simulation parameters}
}
\value{
List of numerics with half-lives
}
\description{
Estimates half-life based on a DAISIE simulation
}
\author{
Joshua Lambert
}
