% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ExpEIN.R
\name{DAISIE_ExpEIN2}
\alias{DAISIE_ExpEIN2}
\title{The expected number of endemics and non-endemics under the DAISIE model}
\usage{
DAISIE_ExpEIN2(
  t,
  pars1,
  M,
  initEI = c(0, 0),
  pars2 = list(res = 1000, ddep = 11, methode = "odeint::runge_kutta_fehlberg78",
    reltolint = 1e-16, abstolint = 1e-16)
)
}
\arguments{
\item{t}{The time at which the expectations need to be computed.}

\item{pars1}{Vector of model parameters: \cr \cr \code{pars1[1]} corresponds
to lambda^c (cladogenesis rate) \cr \code{pars1[2]} corresponds to mu
(extinction rate) \cr \code{pars1[3]} corresponds to K (clade-level
carrying capacity) \cr \code{pars1[4]} corresponds to gamma
(immigration rate) \cr \code{pars1[5]} corresponds to lambda^a
(anagenesis rate).}

\item{M}{Numeric defining the size of mainland pool, i.e. the number of
species that can potentially colonize the island.}

\item{initEI}{The initial values for the number of endemics and
  non-endemics. In \code{\link{DAISIE_probdist}()} or
\code{\link{DAISIE_margprobdist}()} either this or initprobs must be NULL. In
\code{\link{DAISIE_numcol}()} when it is NULL, it is assumed that the island
  is empty.}

\item{pars2}{list of settings
\itemize{
\item{\code{res}: the number of equations}
\item{\code{ddep}: the model of diversity-dependence}
\item{\code{methode}: the method used to integrate the ODE system}
\item{\code{reltolint}: the relative tolerance in integration}
\item{\code{abstolint}: the absolute tolerance in integration}
}}
}
\value{
\item{out}{The output is a list with three elements: \cr \cr
\code{ExpE} The number of endemic species \cr \code{ExpI} The number of
non-endemic species \cr \code{ExpN} The sum of the number of endemics and
non-endemics }
}
\description{
This function calculates the expected number of endemics, non-endemics and
the sum of these for a given set of parameter values, a given mainland
species pool size and a given time, where there can be diversity-dependence
}
\author{
Rampal S. Etienne
}
