% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_CS.R
\name{DAISIE_format_CS}
\alias{DAISIE_format_CS}
\title{Wrapper function around for the correct DAISIE_format_CS}
\usage{
DAISIE_format_CS(island_replicates, time, M, sample_freq = 25, verbose = TRUE)
}
\arguments{
\item{island_replicates}{DAISIE_sim_core simulation output}

\item{time}{Numeric double with total time of simulation.}

\item{M}{Int stating number of mainland species.}

\item{sample_freq}{Int stating how often results are
sampled for plotting.}

\item{verbose}{Logical controling if progress is printed to console.}
}
\value{
List with CS DAISIE simulation output
}
\description{
Wrapper function around for the correct DAISIE_format_CS
}
