% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_CS_full_stt.R
\name{DAISIE_format_CS_full_stt}
\alias{DAISIE_format_CS_full_stt}
\title{Formats clade-specific simulation output into standard
DAISIE list output}
\usage{
DAISIE_format_CS_full_stt(
  island_replicates,
  time,
  M,
  sample_freq,
  island_type = "oceanic",
  verbose = TRUE
)
}
\arguments{
\item{island_replicates}{Int stating number of replicates.}

\item{time}{Numeric double with total time of simulation.}

\item{M}{Int stating number of mainland species.}

\item{sample_freq}{Int stating how often results are
sampled for plotting.}

\item{island_type}{type of island for simulation.}

\item{verbose}{Logical controling if progress is printed to console.}
}
\value{
List with CS DAISIE simulation output
}
\description{
Formats clade-specific simulation output into standard
DAISIE list output
}
