% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_GW.R
\name{DAISIE_format_GW}
\alias{DAISIE_format_GW}
\title{Formats guild-wide simulation output into standard DAISIE
list output}
\usage{
DAISIE_format_GW(
  island_replicates,
  time,
  M,
  sample_freq,
  num_guilds,
  verbose = TRUE
)
}
\arguments{
\item{island_replicates}{List output from DAISIE_sim_core functions.}

\item{time}{Numeric double with total time of simulation.}

\item{M}{Int stating number of mainland species.}

\item{sample_freq}{Int stating how often results are
sampled for plotting.}

\item{num_guilds}{number of guilds on the mainland.}

\item{verbose}{Logical controling if progress is printed to console.}
}
\value{
List with GW DAISIE simulation output
}
\description{
Formats guild-wide simulation output into standard DAISIE
list output
}
