% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_IW.R
\name{DAISIE_format_IW}
\alias{DAISIE_format_IW}
\title{Formats island-wide simulation output into 
standard DAISIE list output}
\usage{
DAISIE_format_IW(island_replicates, time, M, sample_freq,
  verbose = FALSE, Tpars = NULL, island_type = island_type)
}
\value{
List with CS DAISIE simulation output
}
\description{
Formats island-wide simulation output into 
standard DAISIE list output
}
\examples{
island_replicates[[rep]] <- DAISIE_sim_core(time = 10,
                                            mainland_n = 1000,
                                            pars = c(2, 2, 40, 0.1, 1),
                                            ddmodel_sim = 11,
                                            island_type = "oceanic",
                                            nonoceanic = NULL,
                                            island_ontogeny = NULL,
                                            Apars = NULL,
                                            Epars = NULL,
                                            Tpars = NULL,
                                            keep_final_state = FALSE,
                                            island_spec = NULL)
DAISIE_format_IW(island_replicates = island_replicates,
                 time = 10,
                 M = 1000,
                 sample_freq = 25,
                 verbose = FALSE,
                 Tpars = NULL,
                 island_type = "oceanic")
}
