% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_half_life.R
\name{DAISIE_half_life}
\alias{DAISIE_half_life}
\title{Measures relaxation half-life from output 
from DAISIE_sim_core}
\usage{
DAISIE_half_life(sim_core, pars)
}
\arguments{
\item{sim_core}{output from DAISIE_sim_core}

\item{pars}{parameter used to simulate data}
}
\value{
a half life of the island
}
\description{
Measures relaxation half-life from output 
from DAISIE_sim_core
}
\author{
Joshua Lambert
}
