% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DAISIE_IW_num_threads}
\alias{DAISIE_IW_num_threads}
\title{IW concurrency control}
\usage{
DAISIE_IW_num_threads(rnumthreads = NULL)
}
\arguments{
\item{num_threads}{\code{num_threads < 0, omitted}: retrieves number of threads. \cr
\code{num_threads = 0}: sets the number of threads to the number of available cores. \cr
\code{num_threads = 1}: single-threaded execution. \cr
\code{num_threads > 1}: sets the number of threads to \code{num_threads}.}
}
\value{
number of threads
}
\description{
Sets or retrieves the number of threads used by the odeint solver.
}
\note{
The maximum number of threads is limited to the value of the C++
standard library function \code{std::thread::hardware_concurrency()}.
This is also the default number of threads upon library load.
Multithreading incurs some overhead. Therefore, single-threaded execution
might be faster for small systems.
}
\examples{
DAISIE_IW_num_threads(as.integer(Sys.getenv("OMP_NUM_THREADS"))

}
