% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ML3.R
\name{DAISIE_loglik_all_choosepar3}
\alias{DAISIE_loglik_all_choosepar3}
\title{ML stub}
\usage{
DAISIE_loglik_all_choosepar3(trparsopt, trparsfix, idparsopt, idparsfix,
  pars2, datalist, methode, CS_version = 1, abstolint = 1e-16,
  reltolint = 1e-10)
}
\arguments{
\item{trparsopt}{stub}

\item{trparsfix}{stub}

\item{idparsopt}{stub}

\item{idparsfix}{stub}

\item{pars2}{stub}

\item{datalist}{stub}

\item{methode}{stub}

\item{CS_version}{stub}

\item{abstolint}{stub}

\item{reltolint}{stub}
}
\value{
stub
}
\description{
ML stub
}
