% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_all_choosepar_relaxed_rate.R
\name{DAISIE_loglik_all_choosepar_relaxed_rate}
\alias{DAISIE_loglik_all_choosepar_relaxed_rate}
\title{Find suitable title}
\usage{
DAISIE_loglik_all_choosepar_relaxed_rate(
  trparsopt,
  trparsfix,
  idparsopt,
  idparsfix,
  relaxed_dist_pars,
  idparsrelaxed,
  idparsnoshift,
  idparseq,
  pars2,
  datalist,
  methode,
  CS_version = 1,
  abstolint = 1e-16,
  reltolint = 1e-10
)
}
\arguments{
\item{trparsopt}{stub}

\item{trparsfix}{stub}

\item{idparsopt}{The ids of the parameters that must be optimized. The ids
are defined as follows: \cr \cr id = 1 corresponds to lambda^c (cladogenesis
rate) \cr id = 2 corresponds to mu (extinction rate) \cr id = 3 corresponds
to K (clade-level carrying capacity) \cr id = 4 corresponds to gamma
(immigration rate) \cr id = 5 corresponds to lambda^a (anagenesis rate) \cr
id = 6 corresponds to lambda^c (cladogenesis rate) for an optional subset of
the species \cr id = 7 corresponds to mu (extinction rate) for an optional
subset of the species\cr id = 8 corresponds to K (clade-level carrying
capacity) for an optional subset of the species\cr id = 9 corresponds to
gamma (immigration rate) for an optional subset of the species\cr id = 10
corresponds to lambda^a (anagenesis rate) for an optional subset of the
species\cr id = 11 corresponds to p_f (fraction of mainland species that
belongs to the second subset of species.}

\item{idparsfix}{The ids of the parameters that should not be optimized,
e.g. c(1,3) if lambda^c and K should not be optimized.}

\item{relaxed_dist_pars}{parameters for relaxing parameters in ML}

\item{idparsrelaxed}{the id of the parameters to be relaxed in ML}

\item{idparsnoshift}{For datatype = 'single' only: The ids of the parameters
that should not be different between two groups of species; This can only
apply to ids 6:10, e.g. idparsnoshift = c(6,7) means that lambda^c and mu
have the same values for both groups.}

\item{idparseq}{stub}

\item{pars2}{Vector of settings: \cr \cr \code{pars2[1]} corresponds to res,
the maximum number of endemics or non-endemics for which the ODE system is
solved; this must be much larger than the actual number for which the
probability needs to be calculated.) \cr \code{pars2[2]} corresponds to M,
size of the mainland pool, i.e the number of species that can potentially
colonize the island.}

\item{datalist}{Data object containing information on colonisation and
branching times. This object can be generated using the DAISIE_dataprep
function, which converts a user-specified data table into a data object, but
the object can of course also be entered directly. It is an R list object
with the following elements.\cr The first element of the list has two or
three components: \cr \cr \code{$island_age} - the island age \cr Then,
depending on whether a distinction between types is made, we have:\cr
\code{$not_present} - the number of mainland lineages that are not present
on the island \cr or:\cr \code{$not_present_type1} - the number of mainland
lineages of type 1 that are not present on the island \cr
\code{$not_present_type2} - the number of mainland lineages of type 2 that
are not present on the island \cr \cr The remaining elements of the list
each contains information on a single colonist lineage on the island and has
5 components:\cr \cr \code{$colonist_name} - the name of the species or
clade that colonized the island \cr \code{$branching_times} - island age and
stem age of the population/species in the case of Non-endemic,
Non-endemic_MaxAge and Endemic anagenetic species. For cladogenetic species
these should be island age and branching times of the radiation including
the stem age of the radiation.\cr \code{$stac} - the status of the colonist
\cr \cr * Non_endemic_MaxAge: 1 \cr * Endemic: 2 \cr * Endemic&Non_Endemic:
3 \cr * Non_Endemic: 4 \cr * Endemic_Singleton_MaxAge: 5 \cr *
Endemic_Clade_MaxAge: 6 \cr * Endemic&Non_Endemic_Clade_MaxAge: 7 \cr \cr
\code{$missing_species} - number of island species that were not sampled for
particular clade (only applicable for endemic clades) \cr \code{$type1or2} -
whether the colonist belongs to type 1 or type 2 \cr}

\item{methode}{Method of the ODE-solver. See package deSolve for details.
Default is "lsodes".}

\item{CS_version}{For internal testing purposes only. Default is 1, the
original DAISIE code.}

\item{abstolint}{stub}

\item{reltolint}{stub}
}
\value{
loglikelihood
}
\description{
Find suitable title
}
