% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ML_IW.R
\name{DAISIE_loglik_IW_choosepar}
\alias{DAISIE_loglik_IW_choosepar}
\title{Title}
\usage{
DAISIE_loglik_IW_choosepar(trparsopt, trparsfix, idparsopt, idparsfix,
  pars2, M, datalist, methode = "ode45", abstolint = 1e-16,
  reltolint = 1e-14)
}
\arguments{
\item{trparsopt}{stub}

\item{trparsfix}{stub}

\item{idparsopt}{stub}

\item{idparsfix}{stub}

\item{pars2}{stub}

\item{datalist}{stub}

\item{methode}{stub}

\item{abstolint}{stub}

\item{reltolint}{stub}
}
\value{
stub
}
\description{
Title
}
