% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_LR.R
\name{DAISIE_LR}
\alias{DAISIE_LR}
\title{Bootstrap likelihood ratio test}
\usage{
DAISIE_LR(
  datalist,
  datatype = "single",
  initparsopt_dd,
  initparsopt_di,
  idparsnoshift = 6:10,
  idparsmat = NULL,
  endmc = 1000,
  seed = 17,
  alpha = 0.05,
  res = 100,
  ddmodel = 0,
  cond = 0,
  island_ontogeny = NA,
  eqmodel = 0,
  x_E = 0.95,
  x_I = 0.98,
  tol = c(1e-04, 1e-05, 1e-07),
  maxiter = 1000 * round((1.25)^length(idparsopt)),
  methode = "lsodes",
  optimmethod = "subplex",
  CS_version = 1,
  verbose = TRUE,
  tolint = c(1e-16, 1e-10)
)
}
\arguments{
\item{datalist}{stub}

\item{datatype}{stub}

\item{initparsopt_dd}{stub}

\item{initparsopt_di}{stub}

\item{idparsnoshift}{stub}

\item{idparsmat}{stub}

\item{endmc}{stub}

\item{seed}{stub}

\item{alpha}{stub}

\item{res}{stub}

\item{ddmodel}{stub}

\item{cond}{stub}

\item{island_ontogeny}{stub}

\item{eqmodel}{stub}

\item{x_E}{stub}

\item{x_I}{stub}

\item{tol}{stub}

\item{maxiter}{stub}

\item{methode}{stub}

\item{optimmethod}{stub}

\item{CS_version}{stub}

\item{verbose}{stub}

\item{tolint}{stub}
}
\value{
stub
}
\description{
Bootstrap likelihood ratio test
}
\examples{
stub
}
