% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_make_archipelago.R
\name{DAISIE_make_global}
\alias{DAISIE_make_global}
\title{Title}
\usage{
DAISIE_make_global(archipelago_list, M, phylo_data, archipelago_data)
}
\arguments{
\item{archipelago_list}{}

\item{M}{}

\item{phylo_data}{}

\item{archipelago_data}{}
}
\description{
Title
}
