% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ML2.R
\name{DAISIE_ML2}
\alias{DAISIE_ML2}
\title{Computes MLE for two sets of species under a clade specific scenario}
\usage{
DAISIE_ML2(
  datalist,
  initparsopt,
  idparsopt,
  parsfix,
  idparsfix,
  idparsmat,
  res = 100,
  ddmodel = 0,
  cond = 0,
  island_ontogeny = NA,
  tol = c(1e-04, 1e-05, 1e-07),
  maxiter = 1000 * round((1.25)^length(idparsopt)),
  methode = "lsodes",
  optimmethod = "subplex",
  verbose = 0,
  tolint = c(1e-16, 1e-10)
)
}
\arguments{
\item{island_ontogeny}{In \code{\link{DAISIE_sim_time_dependent}} a string
describing the type of island ontogeny. Can be \code{"const"}, \code{"beta"}
for a beta function describing area through time. String checked by
\code{\link{is_island_ontogeny_input}}.
\cr In all other functions a numeric describing the type of island ontogeny.
Can be \code{0} for constant, \code{1} for a beta function describing area
through time.}

\item{verbose}{Logical, \code{Default = TRUE} Give intermediate output,
also if everything goes ok.}
}
\value{
The output is a dataframe containing estimated parameters and
maximum loglikelihood.  \item{lambda_c}{ gives the maximum likelihood
estimate of lambda^c, the rate of cladogenesis} \item{mu}{ gives the maximum
likelihood estimate of mu, the extinction rate} \item{K}{ gives the maximum
likelihood estimate of K, the carrying-capacity} \item{gamma}{ gives the
maximum likelihood estimate of gamma, the immigration rate }
\item{lambda_a}{ gives the maximum likelihood estimate of lambda^a, the rate
of anagenesis} \item{lambda_c2}{ gives the maximum likelihood estimate of
lambda^c2, the rate of cladogenesis for the second group of
species} \item{mu2}{ gives the maximum likelihood estimate of mu2, the
extinction rate for the second group of species} \item{K2}{ gives
the maximum likelihood estimate of K2, the carrying-capacity for the
 second group of species} \item{gamma2}{ gives the maximum
likelihood estimate of gamma2, the immigration rate for the second
group of species} \item{lambda_a2}{ gives the maximum likelihood estimate of
lambda^a2, the rate of anagenesis for the second group of species}
\item{loglik}{ gives the maximum loglikelihood} \item{df}{ gives the number
of estimated parameters, i.e. degrees of feedom} \item{conv}{ gives a
message on convergence of optimization; conv = 0 means convergence}
}
\description{
Computes MLE for two sets of species under a clade specific scenario
}
