% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ML3.R
\name{DAISIE_ML3}
\alias{DAISIE_ML3}
\title{Computes MLE for single type species under a clade specific scenario with 
ontogeny}
\usage{
DAISIE_ML3(datalist, initparsopt, idparsopt, parsfix, idparsfix,
  res = 100, ddmodel = 0, cond = 0, island_ontogeny, tol = c(1e-04,
  1e-05, 1e-07), maxiter = 1000 * round((1.25)^length(idparsopt)),
  methode = "lsodes", optimmethod = "subplex", CS_version = 1,
  verbose = 0, tolint = c(1e-16, 1e-10))
}
\value{
The output is a dataframe containing estimated parameters and
maximum loglikelihood.  \item{lambda_c}{ gives the maximum likelihood
estimate of lambda^c, the rate of cladogenesis} \item{mu}{ gives the maximum
likelihood estimate of mu, the extinction rate} \item{K}{ gives the maximum
likelihood estimate of K, the carrying-capacity} \item{gamma}{ gives the
maximum likelihood estimate of gamma, the immigration rate }
\item{lambda_a}{ gives the maximum likelihood estimate of lambda^a, the rate
of anagenesis} \item{loglik}{ gives the maximum loglikelihood} 
\item{df}{ gives the number of estimated parameters, i.e. degrees of feedom}
\item{conv}{ gives a message on convergence of optimization;
conv = 0 means convergence}
}
\description{
Computes MLE for single type species under a clade specific scenario with 
ontogeny
}
