% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_MW_ML.R
\name{DAISIE_MW_loglik_choosepar}
\alias{DAISIE_MW_loglik_choosepar}
\title{Title}
\usage{
DAISIE_MW_loglik_choosepar(trparsopt, trparsfix, idparsopt, idparsfix,
  pars2, datalist, methode = "lsodes", CS_version = 1,
  abstolint = 1e-16, reltolint = 1e-14, distance_type = "continent",
  distance_dep = "power", parallel = "local", cpus = 3)
}
\arguments{
\item{trparsopt}{stub}

\item{trparsfix}{stub}

\item{idparsopt}{stub}

\item{idparsfix}{stub}

\item{pars2}{stub}

\item{datalist}{stub}

\item{methode}{stub}

\item{CS_version}{stub}

\item{abstolint}{stub}

\item{reltolint}{stub}
}
\value{
stub
}
\description{
Title
}
