% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{DAISIE_nonoceanic_spec}
\alias{DAISIE_nonoceanic_spec}
\title{Calculates the species on the island initially when \code{nonoceanic_pars[1]
!= 0}}
\usage{
DAISIE_nonoceanic_spec(prob_samp, prob_nonend, mainland_n)
}
\arguments{
\item{prob_samp}{probability of a species being sampled
from the mainland pool}

\item{prob_nonend}{probability of a species sampled being non-endemic}

\item{mainland_n}{number of species in the mainland pool}
}
\value{
A list of non-endemic species, endemic species and the new
mainland species pool
}
\description{
Calculates the species on the island initially when \code{nonoceanic_pars[1]
!= 0}
}
\examples{
DAISIE_nonoceanic_spec(
prob_samp = 0.1,
prob_nonend = 0.9,
mainland_n = 1000)
}
