% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_params.R
\name{DAISIE_plot_area}
\alias{DAISIE_plot_area}
\title{Plots island area function through time}
\usage{
DAISIE_plot_area(
  total_time,
  area_pars,
  peak,
  island_ontogeny = "beta",
  resolution,
  sea_level = "const"
)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{area_pars}{A named list containing area and sea level parameters as
created by \code{\link{create_area_pars}()}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: current area}
  \item{[3]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
  \item{[7]: angle of the slope of the island}
}}

\item{peak}{A numeric value specifying the peakiness (or shaprness) of the
ontogeny curve. Higher values imply peakier ontogeny. This value is
internally calculated by \code{\link{calc_peak}()} given the area at the
present and the \code{area_pars}.}

\item{island_ontogeny}{In \code{\link{DAISIE_sim_time_dep}()},
\code{\link{DAISIE_ML_CS}} and plotting a string describing the type of
island ontogeny. Can be \code{"const"}, \code{"beta"} for a beta function
describing area through time. \cr In all other functions a
numeric describing the type of island ontogeny. Can be \code{0} for
constant, \code{1} for a beta function describing area through time. In ML
functions \code{island_ontogeny = NA} assumes constant ontogeny. Time
dependent estimation is not yet available as development is still ongoing.
Will return an error if called in that case.}

\item{resolution}{numeric indicating resolution of plot. Should be < 0.}

\item{sea_level}{In \code{\link{DAISIE_sim_time_dep}()} and plotting a
string describing the type of sea level. Can be \code{"const"} or
\code{"sine"} for a sine function describing area through time.
\cr In all other functions a numeric describing the type of sea level. Can
be \code{0} for constant, \code{1} for a sine function describing area
through time.}
}
\value{
a plot with the area size through time
}
\description{
Plots island area function through time
}
\seealso{
Other rate calculations:
\code{\link[=get_ana_rate]{get_ana_rate()}},
\code{\link[=get_ext_rate]{get_ext_rate()}},
\code{\link[=get_immig_rate]{get_immig_rate()}},
\code{\link[=island_area]{island_area()}},
\code{\link[=island_area_vector]{island_area_vector()}},
\code{\link[=update_max_rates]{update_max_rates()}},
\code{\link[=update_rates]{update_rates()}}
}
\concept{rate calculations}
\keyword{internal}
