% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_params.R
\name{DAISIE_plot_cladogenesis}
\alias{DAISIE_plot_cladogenesis}
\title{Plot cladogenesis rate through time}
\usage{
DAISIE_plot_cladogenesis(
  totaltime,
  K,
  area_pars,
  lac,
  island_ontogeny = "beta",
  hyper_pars = NULL,
  dist_pars = NULL,
  removed_timepoints,
  resolution
)
}
\arguments{
\item{totaltime}{total time of simulation}

\item{K}{K (clade-level carrying capacity)}

\item{area_pars}{a named list containing area parameters as created by create_area_pars:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{lac}{minimum per capita cladogenesis rate}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{NULL},
\code{beta} for a beta function describing area through time.}

\item{removed_timepoints}{starting position of time vector}

\item{resolution}{resolution of time axis}
}
\value{
a plot with per capita cladogenesis rate through time and dataframe with immigration
at corresponding time
}
\description{
Plot cladogenesis rate through time
}
\author{
Pedro Neves
}
