% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_comparison_stts.R
\name{DAISIE_plot_comparison_stts}
\alias{DAISIE_plot_comparison_stts}
\title{Plot STT and overlay additional STT curves}
\usage{
DAISIE_plot_comparison_stts(time, plot_lists_simulations,
  plot_lists_simulations_MLE, type = type)
}
\arguments{
\item{time}{the time span simulated}

\item{plot_lists_simulations}{}

\item{plot_lists_simulations_MLE}{}

\item{type}{}
}
\value{
Standard \code{\link{DAISIE_plot_stt}} with overlaid additional
STT curves for comparison.
}
\description{
Plot STT and overlay additional STT curves
}
\seealso{
\code{\link{DAISIE_plot_sims}}, \code{\link{DAISIE_plot_stt}}, 
\code{\link{DAISIE_prepare_data_plotting}}
}
