% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_pc_rates.R
\name{DAISIE_plot_pc_rates}
\alias{DAISIE_plot_pc_rates}
\title{Plot per capita STT and per-capita rate change thourgh simulation}
\usage{
DAISIE_plot_pc_rates(
  simulation_outputs,
  totaltime,
  pars,
  island_ontogeny,
  area_pars,
  ext_pars,
  mainland_n = 1000,
  resolution = 0.001
)
}
\arguments{
\item{simulation_outputs}{Set of simulation outputs as created by
\code{\link{DAISIE_sim_constant_rate}},
\code{\link{DAISIE_sim_time_dependent}},
or \code{\link{DAISIE_sim_constant_rate_shift}}.}

\item{totaltime}{Numeric; total time to simulate}

\item{pars}{Contains the model parameters: \cr \cr \code{pars[1]}
corresponds to lambda^c (cladogenesis rate) \cr \code{pars[2]} corresponds
to mu (extinction rate) \cr \code{pars[3]} corresponds to K (clade-level
carrying capacity). Set K=Inf for non-diversity dependence.\cr
\code{pars[4]} corresponds to gamma (immigration rate) \cr \code{pars[5]}
corresponds to lambda^a (anagenesis rate) \cr \code{pars[6]} corresponds to
lambda^c (cladogenesis rate) for type 2 species \cr \code{pars[7]}
corresponds to mu (extinction rate) for type 2 species\cr \code{pars[8]}
corresponds to K (clade-level carrying capacity) for type 2 species.  Set
K=Inf for non-diversity dependence.\cr \code{pars[9]} corresponds to gamma
(immigration rate) for type 2 species\cr \code{pars[10]} corresponds to
lambda^a (anagenesis rate) for type 2 species\cr The elements 6:10 are
optional and are required only when type 2 species are included.}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{"const"},
\code{"beta"} for a beta function describing area through time.}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{ext_pars}{a numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{mainland_n}{The size of the mainland pool, i.e the number of species that can
potentially colonize the island}

\item{resolution}{numeric indicating resolution of plot. Should be < 0.}
}
\value{
A composite plot containing the STT plot, and variation in area,
extinction rate, cladogenesis rate and immigration rate
}
\description{
Plot per capita STT and per-capita rate change thourgh simulation
}
\author{
Pedro Neves
}
