% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_stt.R
\name{DAISIE_prepare_data_plotting}
\alias{DAISIE_prepare_data_plotting}
\title{Prepare input for DAISIE_stt}
\usage{
DAISIE_prepare_data_plotting(island_replicates)
}
\arguments{
\item{island_replicates}{Island replicates in DAISIE format (produced in
\code{\link{DAISIE_sim}} with \code{format = TRUE} option). Minimally, this must be 
a list, that has as much elements as replicates. Each element must be a
list with the elements \code{island_age}, \code{not_present} 
and \code{stt_all}. \code{stt_all} must be a data frame with
the column names \code{Time}, \code{nI}, \code{nA}, \code{nC} 
and \code{present}.}
}
\value{
a list with wrangled data to be used for plotting STT plots with
DAISIE_plot_stt
}
\description{
Prepare input for DAISIE_stt
}
\seealso{
\code{\link{DAISIE_plot_stt}}, \code{\link{DAISIE_plot_sims}}
}
