% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sample_event.R
\name{DAISIE_sample_event}
\alias{DAISIE_sample_event}
\title{Samples what event to happen next}
\usage{
DAISIE_sample_event(rates, island_ontogeny = NULL, Tpars = NULL)
}
\arguments{
\item{rates}{numeric list with probability rates for each event. In the
ontogeny case it also contains the maximum possible probability for the
event at each timestep.}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{NULL},
\code{"beta"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{Tpars}{A named list containing diversification rates considering two trait states:
\itemize{
  \item{[1]:A numeric with the per capita transition rate with state1}
  \item{[2]:A numeric with the per capita immigration rate with state2}
  \item{[3]:A numeric with the per capita extinction rate with state2}
  \item{[4]:A numeric with the per capita anagenesis rate with state2}
  \item{[5]:A numeric with the per capita cladogenesis rate with state2}
  \item{[6]:A numeric with the per capita transition rate with state2}
  \item{[7]:A numeric with the number of species with trait state 2 on mainland}
}}
}
\value{
numeric indicating what event will happen, or a supposed event that would
happen in some timesteps of the ontogeny algorithm.
\itemize{
  \item{[1]: immigration event}
  \item{[2]: extinction event}
  \item{[3]: cladogenesis event}
  \item{[4]: anagenesis event}
  \item{[5]: proposed extinction that will not happen or transition event}
  \item{[6]: proposed immigration that will not happen or immigration event with state2}
  \item{[7]: proposed cladogenesis that will not happen or extinction event with state2}
  \item{[8]: cladogenesis event with state2}
  \item{[9]: anagenesis event with state2}
  \item{[10]: transition event with state2}
}
}
\description{
Samples what event to happen next
}
\author{
Pedro Neves
}
