% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sample_event_constant_rate.R
\name{DAISIE_sample_event_constant_rate}
\alias{DAISIE_sample_event_constant_rate}
\title{Samples what event to happen next}
\usage{
DAISIE_sample_event_constant_rate(rates)
}
\arguments{
\item{rates}{numeric list with probability rates for each event. In the
ontogeny case it also contains the maximum possible probability for the
event at each timestep.
\code{\link{update_rates}}.}
}
\value{
numeric indicating what event will happen, or a supposed event that
would happen in some timesteps of the ontogeny algorithm.
\itemize{
  \item{[1]: immigration event}
  \item{[2]: extinction event}
  \item{[3]: cladogenesis event}
  \item{[4]: anagenesis event}
}
}
\description{
Samples what event to happen next
}
\author{
Pedro Neves
}
