% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sample_event_time_dependent.R
\name{DAISIE_sample_event_time_dependent}
\alias{DAISIE_sample_event_time_dependent}
\title{Samples what event to happen next}
\usage{
DAISIE_sample_event_time_dependent(max_rates)
}
\arguments{
\item{max_rates}{named list of numeric max rates as returned by
\code{\link{update_rates}}.}
}
\value{
numeric indicating what event will happen, or a supposed event that
would happen in some timesteps of the ontogeny algorithm.
\itemize{
  \item{[1]: immigration event}
  \item{[2]: extinction event}
  \item{[3]: cladogenesis event}
  \item{[4]: anagenesis event}
}
}
\description{
Samples what event to happen next
}
\author{
Pedro Neves
}
