% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sample_event_trait_dependent.R
\name{DAISIE_sample_event_trait_dependent}
\alias{DAISIE_sample_event_trait_dependent}
\title{Samples what event to happen next}
\usage{
DAISIE_sample_event_trait_dependent(rates)
}
\arguments{
\item{rates}{named list of numeric rates as returned by
\code{\link{update_rates}}.}
}
\value{
numeric indicating what event will happen, or a supposed event that
would happen in some timesteps of the ontogeny algorithm.
\itemize{
  \item{[1]: immigration event with trait1}
  \item{[2]: extinction event with trait1}
  \item{[3]: cladogenesis event with trait1}
  \item{[4]: anagenesis event with trait1}
  \item{[5]: transition event with trait1}
  \item{[6]: immigration event with trait2}
  \item{[7]: extinction event with trait2}
  \item{[8]: cladogenesis event with trait2}
  \item{[9]: anagenesis event with trait2}
  \item{[10]: transition event with trait2}
}
}
\description{
Samples what event to happen next
}
\author{
Pedro Neves
}
