% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_half_life.R
\name{DAISIE_sim_avg_half_life}
\alias{DAISIE_sim_avg_half_life}
\title{Title}
\usage{
DAISIE_sim_avg_half_life(time, M, pars, replicates, nonoceanic,
  divdepmodel = "CS", prop_type2_pool = NA,
  replicates_apply_type2 = TRUE, sample_freq = 10000,
  plot_sims = TRUE)
}
\arguments{
\item{time}{}

\item{M}{}

\item{pars}{}

\item{replicates}{}

\item{nonoceanic}{}

\item{divdepmodel}{}

\item{prop_type2_pool}{}

\item{replicates_apply_type2}{}

\item{sample_freq}{}

\item{plot_sims}{}
}
\description{
Title
}
