% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{DAISIE_sim_core}
\alias{DAISIE_sim_core}
\title{Internal function of the DAISIE simulation}
\usage{
DAISIE_sim_core(time, mainland_n, pars, ddmodel_sim = 11,
  island_type = "oceanic", nonoceanic_params = NULL,
  k_dist_params = NULL, island_ontogeny = 0, Apars = NULL,
  Epars = NULL, ext_multiplier = 0.5, keep_final_state = FALSE,
  island_spec = NULL)
}
\arguments{
\item{time}{Simulated amount of time}

\item{mainland_n}{A numeric stating the number of mainland species, that
is the number of species that can potentially colonize the island.
If \code{\link{DAISIE_sim}} uses a clade-specific diversity dependence,
this value is set to 1.
If \code{\link{DAISIE_sim}} uses an island-wide diversity dependence,
this value is set to the number of mainland species.}

\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}

\item{ddmodel_sim}{The a numeric vector to determined which parameters should
be diversity dependent.}

\item{island_type}{Option island_type = 'oceanic' is a model equal to Valente
et al., 2015. island_type = 'nonoceanic' is a nonoceanic model where initial
species richness is non-zero determined by the nonoceanic parameters.}

\item{nonoceanic_params}{A vector of length three with: the island area as a
proportion of the mainland, the probability of native species being
nonendemic and the size of the mainland pool.}

\item{k_dist_params}{Vector of two elements to define gamma
distribution for sampling carrying capacities. \code{k_dist_params[1]}
is the shape parameter, \code{k_dist_params[2]} is the rate parameter}

\item{island_ontogeny}{a numeric describing the type of island ontogeny.
Can be \code{0} for constant, \code{1} for a linear function through time}

\item{Apars}{A named list containing area parameters as created by
create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{ext_multiplier}{or \code{2} for a beta function describing area.}

\item{keep_final_state}{Logical indicating if final state of simulation
should be returned. Default is \code{FALSE}.}

\item{island_spec}{A matrix with species on island (state of system}
}
\description{
Internal function of the DAISIE simulation
}
