% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core_checked.R
\name{DAISIE_sim_core_checked}
\alias{DAISIE_sim_core_checked}
\title{Runs one DAISIE simulation with a clade-specific carrying capacity.
Version of \code{DAISIE_sim_core} that checks all its inputs
and uses descriptively named arguments}
\usage{
DAISIE_sim_core_checked(sim_time, n_mainland_species, clado_rate, ext_rate,
  carr_cap, imm_rate, ana_rate)
}
\arguments{
\item{sim_time}{length of the simulated time}

\item{n_mainland_species}{number of mainland species}

\item{clado_rate}{cladogenesis rate}

\item{ext_rate}{extinction rate}

\item{carr_cap}{carrying capacity}

\item{imm_rate}{immigration rate}

\item{ana_rate}{anagenesis rate}
}
\value{
a list with these elements:
\describe{
  \item{stt_table}{a species-through-time table}
  \item{branching_times}{branching times}
  \item{stac}{
    the status of the colonist
    \itemize{
      \item{1: \code{Non_endemic_MaxAge} (?immigrant is present but has not formed an extant clade)}
      \item{2: \code{Endemic} (?immigrant is not present but has formed an extant clade)}
      \item{3: \code{Endemic&Non_Endemic} (?immigrant is present and has formed an extant clade)}
      \item{4: \code{Non_endemic} (?immigrant is present but has not formed an extant clade, and it is known when it immigrated)}
    }
  }
  \item{missing_species}{number of missing species}
  \item{other_clades_same_ancestor}{(not always present) ?no idea}
}
}
\description{
Runs one DAISIE simulation with a clade-specific carrying capacity.
Version of \code{DAISIE_sim_core} that checks all its inputs
and uses descriptively named arguments
}
\author{
Richel J.C. Bilderbeek
}
