% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core_constant_rate_shift.R
\name{DAISIE_sim_core_constant_rate_shift}
\alias{DAISIE_sim_core_constant_rate_shift}
\title{Internal function of the DAISIE simulation}
\usage{
DAISIE_sim_core_constant_rate_shift(
  time,
  mainland_n,
  pars,
  nonoceanic_pars = c(0, 0),
  hyper_pars = NULL,
  area_pars = NULL,
  dist_pars = NULL,
  shift_times
)
}
\arguments{
\item{time}{Simulated amount of time}

\item{mainland_n}{A numeric stating the number of mainland species, that
is the number of species that can potentially colonize the island.
If \code{\link{DAISIE_sim_constant_rate_shift}} uses a clade-specific
diversity dependence, this value is set to 1.
If \code{\link{DAISIE_sim_constant_rate_shift}} uses an island-wide
diversity dependence, this value is set to the number of mainland species.}

\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}

\item{nonoceanic_pars}{A vector of length three with: the island area as a
proportion of the mainland, the probability of native species being
nonendemic and the size of the mainland pool.}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations:
\itemize{
\item{[1]: is d_0 the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: is x the exponent for calculating extinction rate}
\item{[3]: is alpha, the exponent for calculating the immigration rate}
\item{[4]: is beta the exponent for calculating the anagenesis rate.}
}}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{dist_pars}{a numeric for the distance from the mainland.}

\item{shift_times}{a numeric vector specifying when the rate shifts occur
before the present.}
}
\description{
Internal function of the DAISIE simulation
}
