% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_half_life.R
\name{DAISIE_sim_half_life}
\alias{DAISIE_sim_half_life}
\title{Title}
\usage{
DAISIE_sim_half_life(time, mainland_n, pars, nonoceanic)
}
\arguments{
\item{time}{}

\item{mainland_n}{}

\item{pars}{}

\item{nonoceanic}{}
}
\description{
Title
}
