% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_island_ontogeny.R
\name{DAISIE_sim_island_ontogeny}
\alias{DAISIE_sim_island_ontogeny}
\title{Simulate islands with given parameters.}
\usage{
DAISIE_sim_island_ontogeny(time, M, pars, replicates, divdepmodel = "CS",
  prop_type2_pool = NA, replicates_apply_type2 = TRUE, sample_freq = 25,
  plot_sims = TRUE, island_ontogeny = NULL, Apars = NULL, Epars = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{time}{simulated amount of time}

\item{pars}{a numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{Apars}{a numeric vector:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{a numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{mainland_n}{number of mainland species, that
is, the number of species that can potentially colonize the island.
If \code{\link{DAISIE_sim}} uses a clade-specific diversity dependence,
this value is set to 1. 
If \code{\link{DAISIE_sim}} uses an island-specific diversity dependence,
this value is set to the number of mainland species.}
}
\description{
This function simulates islands with given cladogenesis,
 extinction, Kprime, immigration and anagenesis parameters.
  If a single parameter set is provided (5 parameters) it simulates islands
   where all species have the same macro-evolutionary process.
    If two paramater sets (10 parameters) are provided, it simulates islands
     where two different macro-evolutionary processes operate, one applying
      to type 1 species and other to type 2 species.
}
