% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_mainland.R
\name{DAISIE_sim_mainland}
\alias{DAISIE_sim_mainland}
\title{Simulates mainland extinction as a pure-death process with replacement
via the speciation of an exisiting mainland species.}
\usage{
DAISIE_sim_mainland(time, M, mainland_ext)
}
\arguments{
\item{time}{Numeric defining the length of the simulation in time units.
For example, if an island is known to be 4 million years old, setting
time = 4 will simulate the entire life span of the island; setting time = 2
will stop the simulation at the mid-life of the island.}

\item{M}{Numeric defining the size of mainland pool, i.e. the number of
species that can potentially colonize the island.}

\item{mainland_ext}{Numeric extinction rate on the mainland}
}
\value{
a list where each element is a mainland lineage, this could be a
single or multiple lineages. Each element of the list is a matrix with nine
columns. The columns contain
\enumerate{
    \item Species identity
    \item Mainland ancestor identity
    \item Colonisation time
    \item Species type
    \item Branching code
    \item Branching time (forwards in time from the start of the simulation)
    \item Anagenetic origin
    \item Species origination time
    \item Species extinction time
}
}
\description{
Simulates mainland extinction as a pure-death process with replacement
via the speciation of an exisiting mainland species.
}
\keyword{internal}
