% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_mainland_ex.R
\name{DAISIE_sim_mainland_ex}
\alias{DAISIE_sim_mainland_ex}
\title{Simulate oceanic islands with mainland extinction given
parameters under time-constant rates}
\usage{
DAISIE_sim_mainland_ex(
  time,
  M,
  island_pars,
  mainland_ext,
  replicates,
  sample_freq = 25,
  plot_sims = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{time}{Numeric defining the length of the simulation in time units.
For example, if an island is known to be 4 million years old, setting
time = 4 will simulate the entire life span of the island; setting time = 2
will stop the simulation at the mid-life of the island.}

\item{M}{Numeric defining the size of mainland pool, i.e. the number of
species that can potentially colonize the island.}

\item{island_pars}{A numeric vector containing the model parameters:
\itemize{
  \item{\code{pars[1]}: lambda^c (cladogenesis rate)}
  \item{\code{pars[2]}: mu (extinction rate)}
  \item{\code{pars[3]}: K (carrying capacity), set K=Inf for diversity
  independence.}
  \item{\code{pars[4]}: gamma (immigration rate)}
  \item{\code{pars[5]}: lambda^a (anagenesis rate)}
  }}

\item{mainland_ext}{Numeric extinction rate on the mainland}

\item{replicates}{Integer specifying number of island replicates to be
simulated.}

\item{sample_freq}{Numeric specifing the number of units times should be
divided by for plotting purposes. Larger values will lead to plots with
higher resolution, but will also run slower.}

\item{plot_sims}{\code{Default = TRUE} plots species-through-time (STT)
plots. It detects how many types of species are present. If only one type
of species is present, STT is plotted for all species. If two types are
present, three plots are produced: STT for all, STT for type 1 and STT for
type 2.}

\item{verbose}{In simulation and dataprep functions a logical,
\code{Default = TRUE} gives intermediate output should be printed.
For ML functions a numeric determining if intermediate output should be
printed, \code{Default = 0} does not print, \code{verbose = 1} prints
intermediate output of the parameters and loglikelihood, \code{verbose = 2}
means also intermediate progress during loglikelihood computation is shown.}

\item{...}{Any arguments to pass on to plotting functions.}
}
\value{
A list. The highest level of the least corresponds to each individual
replicate. The first element of each replicate is composed of island
information containing:
\itemize{
  \item{\code{$island_age}: A numeric with the island age.}
  \item{\code{$not_present}: the number of mainland lineages that are not
    present on the island.}
  \item{\code{$stt_all}: STT table for all species on the island
    (nI - number of non-endemic species; nA - number of anagenetic species,
    nC - number of cladogenetic species, present - number of independent
    colonisations present)}
}
The subsequent elements of the list pertaining to each replcate contain
information on a single colonist lineage on the island and have 4 components:
\itemize{
  \item{\code{$branching_times}: island age and stem age of the
    population/species in the case of Non-endemic, Non-endemic_MaxAge and
    Endemic anagenetic species.

    For cladogenetic species these should
    be island age and branching times of the radiation including the
    stem age of the radiation.}
  \item{\code{$stac}: An integer ranging from 1 to 4
  indicating the status of the colonist:}
  \enumerate{
    \item Non_endemic_MaxAge
    \item Endemic
    \item Endemic&Non_Endemic
    \item Non_endemic_MaxAge
}
\item{\code{$missing_species}: number of island species that were
not sampled for particular clade (only applicable for endemic clades)}
\item{\code{$type_1or2}: whether the colonist belongs to type 1 or type 2}
}
}
\description{
This function simulates islands with given cladogenesis,
extinction, Kprime, immigration and anagenesis parameters, and simulates
the mainland given an extinction rate parameter, all of which are modelled as
time-constant parameters.
}
\examples{
## Simulate 2 islands (replicates) for 1 million years, with a mainland
## extinction rate of 1 (SpMy^-1). Pool size 100.

set.seed(1)
island_replicates <- DAISIE_sim_mainland_ex(
  time = 1,
  M = 100,
  island_pars = c(1, 1, 10, 0.1, 1),
  mainland_ext = 1,
  replicates = 2,
  plot_sims = FALSE,
  verbose = FALSE
)

}
\seealso{
\code{\link{DAISIE_plot_sims}()} for plotting STT of simulation
outputs.

Other simulation models: 
\code{\link{DAISIE_sim_constant_rate_shift}()},
\code{\link{DAISIE_sim_relaxed_rate}()},
\code{\link{DAISIE_sim_time_dependent}()},
\code{\link{DAISIE_sim}()}
}
\author{
Luis Valente, Albert Phillimore, Joshua Lambert, Shu Xie, Pedro
Neves, Richèl J. C. Bilderbeek, Rampal Etienne
}
\concept{simulation models}
\keyword{models}
