% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daisie_sim_rcpp}
\alias{daisie_sim_rcpp}
\title{Driver for the daisie sim time loop}
\value{
List with the stt_table and the island_spec matrix
}
\description{
This is the rcpp function to run DAISIE simulations
}
