% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_update_state.R
\name{DAISIE_sim_update_state}
\alias{DAISIE_sim_update_state}
\title{Updates state of island given sampled event}
\usage{
DAISIE_sim_update_state(timeval, totaltime, possible_event, maxspecID,
  mainland_spec, island_spec, stt_table, Tpars)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{simulated amount of time}

\item{possible_event}{numeric indicating what event will happen.}

\item{maxspecID}{current species IDs}

\item{mainland_spec}{number of mainland species}

\item{island_spec}{A matrix with species on island (state of system at each time point)}

\item{stt_table}{A species-through-time table}

\item{Tpars}{A named list containing diversification rates considering two trait states:
\itemize{
  \item{[1]:A numeric with the per capita transition rate with state1}
  \item{[2]:A numeric with the per capita immigration rate with state2}
  \item{[3]:A numeric with the per capita extinction rate with state2}
  \item{[4]:A numeric with the per capita anagenesis rate with state2}
  \item{[5]:A numeric with the per capita cladogenesis rate with state2}
  \item{[6]:A numeric with the per capita transition rate with state2}
  \item{[7]:A numeric with the number of species with trait state 2 on mainland}
}}
}
\description{
Makes the event happen by updating island species matrix and species IDs.
What event happens is determined by the sampling in the algorithm.
}
\seealso{
\link{DAISIE_sim_core}
}
