% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_update_state_constant_rate.R
\name{DAISIE_sim_update_state_constant_rate}
\alias{DAISIE_sim_update_state_constant_rate}
\title{Updates state of island given sampled event for a constant rate case.}
\usage{
DAISIE_sim_update_state_constant_rate(
  timeval,
  totaltime,
  possible_event,
  maxspecID,
  mainland_spec,
  island_spec,
  stt_table
)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{simulated amount of time}

\item{possible_event}{numeric indicating what event will happen.}

\item{maxspecID}{current species IDs}

\item{mainland_spec}{number of mainland species}

\item{island_spec}{A matrix with species on island (state of system at each time point)}

\item{stt_table}{A species-through-time table}
}
\description{
Makes the event happen by updating island species matrix and species IDs.
What event happens is determined by the sampling in the algorithm.
}
\seealso{
\link{DAISIE_sim_core_constant_rate},
\link{DAISIE_sim_update_state_constant_rate}
}
