% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_update_state_trait.R
\name{DAISIE_sim_update_state_shu}
\alias{DAISIE_sim_update_state_shu}
\title{Updates state of island given sampled event}
\usage{
DAISIE_sim_update_state_shu(timeval, totaltime, possible_event, maxspecID,
  mainland_spec0, mainland_spec1, island_spec, stt_table)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{simulated amount of time}

\item{possible_event}{numeric indicating what event will happen.}

\item{maxspecID}{current species IDs}

\item{mainland_spec0}{number of mainland species with trait state 1}

\item{island_spec}{A matrix with species on island (state of system at each time point)}

\item{stt_table}{A species-through-time table}

\item{mainland_spec}{number of mainland species with trait state 0}
}
\description{
Makes the event happen by updating island species matrix and species IDs.
What event happens is determined by the sampling in the algorithm.
}
\seealso{
\link{DAISIE_sim_core}
}
