% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_update_state_time_dependent.R
\name{DAISIE_sim_update_state_time_dependent}
\alias{DAISIE_sim_update_state_time_dependent}
\title{Updates state of island given sampled event}
\usage{
DAISIE_sim_update_state_time_dependent(
  timeval,
  totaltime,
  possible_event,
  maxspecID,
  mainland_spec,
  island_spec,
  stt_table,
  rates,
  max_rates
)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{simulated amount of time}

\item{possible_event}{numeric indicating what event will happen.}

\item{maxspecID}{current species IDs}

\item{mainland_spec}{number of mainland species}

\item{island_spec}{A matrix with species on island (state of system at each time point)}

\item{stt_table}{A species-through-time table}

\item{rates}{numeric list with probability rates for each event. In the
ontogeny case it also contains the maximum possible probability for the
event at each timestep.}

\item{max_rates}{named list of max rates as returned by
\code{\link{update_rates}}.}
}
\description{
Makes the event happen by updating island species matrix and species IDs.
What event happens is determined by the sampling in the algorithm.
}
\seealso{
\link{DAISIE_sim_core_time_dependent}
}
