% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_with_ontogeny.R
\name{DAISIE_sim_with_ontogeny}
\alias{DAISIE_sim_with_ontogeny}
\title{Internal function of the DAISIE simulation}
\usage{
DAISIE_sim_with_ontogeny(time, mainland_n, pars, Apars = NULL,
  Epars = NULL, island_ontogeny = NULL)
}
\arguments{
\item{time}{Simulated amount of time}

\item{mainland_n}{A numeric stating the number of mainland species, that
is, the number of species that can potentially colonize the island.
If \code{\link{DAISIE_sim}} uses a clade-specific diversity dependence,
this value is set to 1. 
If \code{\link{DAISIE_sim}} uses an island-specific diversity dependence,
this value is set to the number of mainland species.}

\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}

\item{Apars}{A named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{island_ontogeny}{A string describing the type of island ontogeny. Can be \code{NULL},
\code{quadratic} for a beta function describing area through time,
 or \code{linear} for a linear function}
}
\description{
Internal function of the DAISIE simulation
}
