% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_single_branch.R
\name{DAISIE_single_branch}
\alias{DAISIE_single_branch}
\title{Single-branch phylogenetic tree}
\usage{
DAISIE_single_branch(tip.label = "t1", edge.length = 1)
}
\arguments{
\item{tip.label}{Name of the unique tip}

\item{edge.length}{Length of the unique edge}
}
\value{
A phylogenetic tree in `phylo` format.
}
\description{
Single-branch phylogenetic tree
}
\examples{

# A tree with one species with a 15Myr-old stem
DAISIE:::DAISIE_single_branch("some species", edge.length = 15)

}
\author{
Raphael Scherrer (github.com/rscherrer)
}
\keyword{internal}
