% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{DAISIE_spec_tables}
\alias{DAISIE_spec_tables}
\title{Title}
\usage{
DAISIE_spec_tables(
  stt_table,
  totaltime,
  timeval,
  nonoceanic_sample,
  island_spec
)
}
\arguments{
\item{island_spec}{}
}
\value{

}
\description{
Title
}
