% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_SR_sim.R
\name{DAISIE_SR_sim}
\alias{DAISIE_SR_sim}
\title{Depracated function for single-shift simulation model. Please use
DAISIE_sim_constant_rate_shift.}
\usage{
DAISIE_SR_sim(
  time,
  M,
  pars,
  replicates,
  sample_freq = 25,
  plot_sims = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{time}{Numeric defining the length of the simulation in time units.
For example, if an island is known to be 4 million years old, setting
time = 4 will simulate the entire life span of the island; setting time = 2
will stop the simulation at the mid-life of the island.}

\item{M}{Numeric defining the size of mainland pool, i.e. the number of
species that can potentially colonize the island.}

\item{pars}{A numeric vector containing the model parameters:
\itemize{
  \item{pars[1]: lambda^c (cladogenesis rate)}
  \item{pars[2]: mu (extinction rate)}
  \item{pars[3]: K (carrying capacity), set K=Inf for diversity
  independence.}
  \item{pars[4]: gamma (immigration rate)}
  \item{pars[5]: lambda^a (anagenesis rate)}
  \item{pars[6]: lambda^c (cladogenesis rate) for either type 2 species
  or rate set 2 in rate shift model}
  \item{pars[7]: mu (extinction rate) for either type 2 species or rate
  set 2 in rate shift model}
  \item{pars[8]: K (carrying capacity) for either type 2 species or rate
  set 2 in rate shift model, set K=Inf for diversity independence.}
  \item{pars[9]: gamma (immigration rate) for either type 2 species or
  rate set 2 in rate shift model}
  \item{pars[10]: lambda^a (anagenesis rate) for either type 2 species
  or rate set 2 in rate shift model}
  \item{pars[11]: time of rate shift before the present}
}}

\item{replicates}{Integer specifying number of island replicates to be
simulated.}

\item{sample_freq}{Numeric specifing the number of units times should be
divided by for plotting purposes. Larger values will lead to plots with
higher resolution, but will also run slower.}

\item{plot_sims}{\code{Default = TRUE} plots species-through-time (STT)
plots. It detects how many types of species are present. If only one type
of species is present, STT is plotted for all species. If two types are
present, three plots are produced: STT for all, STT for type 1 and STT for
type 2.}

\item{verbose}{In simulation and dataprep functions a logical,
\code{Default = TRUE} gives intermediate output should be printed.
For ML functions a numeric determining if intermediate output should be
printed, \code{Default = 0} does not print, \code{verbose = 1} prints
intermediate output of the parameters and loglikelihood, \code{verbose = 2}
means also intermediate progress during loglikelihood computation is shown.}
}
\value{
simulation output
}
\description{
Depracated function for single-shift simulation model. Please use
DAISIE_sim_constant_rate_shift.
}
