% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_test.R
\name{DAISIE_test}
\alias{DAISIE_test}
\title{Tests DAISIE for data sets included in the paper by Valente et al}
\usage{
DAISIE_test()
}
\description{
Tests DAISIE for data sets included in the paper by Valente et al.
}
\keyword{biogeography}
\keyword{island}
\keyword{model}
