% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sample_event.R
\name{determine_event}
\alias{determine_event}
\title{Samples what event to happen next}
\usage{
determine_event(rates, island_ontogeny)
}
\arguments{
\item{rates}{numeric list with probability rates for each event. In the 
ontogeny case it also contains the maximum possible probability for the 
event at each timestep.}
}
\value{
numeric indicating what event will happen, or a supposed event that would
happen in some timesteps of the ontogeny algorithm.
#' \itemize{
  \item{[1]: immigration event}
  \item{[2]: extinction event}
  \item{[3]: cladogenesis event}
  \item{[4]: anagenesis event}
  \item{[5]: proposed extinction that will not happen}
  \item{[6]: proposed immigration that will not happen}
  \item{[7]: proposed cladogenesis that will not happen}
}
}
\description{
Samples what event to happen next
}
\author{
Pedro Neves
}
