% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_dynamic_t_hor}
\alias{get_dynamic_t_hor}
\title{Dynamically update horizon time}
\usage{
get_dynamic_t_hor(t_hor, timeval, totaltime, ext, area_pars,
  island_ontogeny, sea_level, ext_multiplier, interval_min = NULL,
  interval_max = NULL)
}
\arguments{
\item{t_hor}{time of horizon for max extinction.}

\item{timeval}{current time of simulation.}

\item{totaltime}{total time of simulation.}

\item{ext}{effective extinction rate at timeval.}

\item{area_pars}{a named list containing area parameters as created
by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{island_ontogeny}{a numeric describing the type of island ontogeny.
Can be \code{0} for constant, \code{1} for a beta function describing area.
  describing area through time.}

\item{sea_level}{a numeric describing the type of sea level.}

\item{ext_multiplier}{reduces or increases distance of horizon to current
simulation time.}

\item{interval_min}{Lower bound of interval for which maximum area should be
determined. Temporarily hard coded to 0 (see note).}

\item{interval_max}{Upper bound of interval for which maximum area should be
determined. Temporarily hard coded to \code{totaltime} (see note).}
}
\value{
Numeric value with updated t_hor
}
\description{
Update horizon time according to the dynamic maxima of an area/sea-level
function.
}
\note{
At the moment sea-level is set to 0 and only global maximum of function
is calculated.
}
\examples{
t_hor <- 5
timeval <- 1
totaltime <- 10
ext <- 0.5
area_pars <- DAISIE::create_area_pars(
  max_area = 5000,
  proportional_peak_t = 0.5,
  peak_sharpness = 1,
  total_island_age = 15,
  sea_level_amplitude = 0,
  sea_level_frequency = 0
)
island_ontogeny <- 1
sea_level <- 0
ext_multiplier <- 1000

testthat::expect_silent(
  dynamic_t_hor <- DAISIE:::get_dynamic_t_hor(
    t_hor = t_hor,
    timeval = timeval,
    totaltime = totaltime,
    ext = ext,
    area_pars = area_pars,
    island_ontogeny = island_ontogeny,
    sea_level = sea_level,
    ext_multiplier = ext_multiplier
  )
)

}
\seealso{
\code{\link{get_t_hor}} for details on
algorithm. Relies on \code{\link[stats]{optimize}} to determine maxima.

Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ana_rate}}, \code{\link{get_ext_rate}},
  \code{\link{get_immig_rate}}, \code{\link{get_t_hor}},
  \code{\link{island_area_vector}},
  \code{\link{island_area}}, \code{\link{update_rates}}
}
\author{
Pedro Neves, Joshua Lambert
}
\concept{rates calculation}
