% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_max_rates.R
\name{get_global_max_area}
\alias{get_global_max_area}
\title{Get the maximum area}
\usage{
get_global_max_area(total_time, area_pars, peak, island_ontogeny, sea_level)
}
\arguments{
\item{total_time}{Numeric defining the length of the simulation in time
units.}

\item{area_pars}{A named list containing area and sea level parameters as
created by \code{\link{create_area_pars}()}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: current area}
  \item{[3]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
  \item{[7]: angle of the slope of the island}
}}

\item{peak}{A numeric value specifying the peakiness (or shaprness) of the
ontogeny curve. Higher values imply peakier ontogeny. This value is
internally calculated by \code{\link{calc_peak}()} given the area at the
present and the \code{area_pars}.}

\item{island_ontogeny}{In \code{\link{DAISIE_sim_time_dep}()},
\code{\link{DAISIE_ML_CS}} and plotting a string describing the type of
island ontogeny. Can be \code{"const"}, \code{"beta"} for a beta function
describing area through time. String checked by
\code{\link{is_island_ontogeny_input}()}. \cr In all other functions a
numeric describing the type of island ontogeny. Can be \code{0} for
constant, \code{1} for a beta function describing area through time. In ML
functions \code{island_ontogeny = NA} assumes constant ontogeny. Time
dependent estimation is not yet available as development is still ongoing.
Will return an error if called in that case.}

\item{sea_level}{In \code{\link{DAISIE_sim_time_dep}()} and plotting a
string describing the type of sea level. Can be \code{"const"} or
\code{"sine"} for a sine function describing area through time. String
checked by \code{\link{is_sea_level_input}()}.
\cr In all other functions a numeric describing the type of sea level. Can
be \code{0} for constant, \code{1} for a sine function describing area
through time.}
}
\value{
Numeric maximum area during the simulation.
}
\description{
Get the maximum area
}
\examples{
timeval <- 1
total_time <- 10
area_pars <- DAISIE::create_area_pars(
  max_area = 5000,
  current_area = 4000,
  proportional_peak_t = 0.5,
  total_island_age = 15,
  sea_level_amplitude = 0,
  sea_level_frequency = 0,
  island_gradient_angle = 0
)
island_ontogeny <- 1
sea_level <- 0
peak <- DAISIE:::calc_peak(total_time = total_time, area_pars = area_pars)
testthat::expect_silent(
  global_max_area_time <- DAISIE:::get_global_max_area(
    total_time = total_time,
    peak = peak,
    area_pars = area_pars,
    island_ontogeny = island_ontogeny,
    sea_level = sea_level
  )
)

}
\author{
Pedro Neves, Joshua Lambert, Shu Xie
}
\keyword{internal}
