% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_max_rates.R
\name{get_global_max_area_time}
\alias{get_global_max_area_time}
\title{Title}
\usage{
get_global_max_area_time(totaltime, area_pars, island_ontogeny, sea_level)
}
\arguments{
\item{totaltime}{stub}

\item{area_pars}{stub}

\item{island_ontogeny}{stub}

\item{sea_level}{stub}
}
\value{

}
\description{
Title
}
\examples{
timeval <- 1
totaltime <- 10
area_pars <- DAISIE::create_area_pars(
  max_area = 5000,
  proportional_peak_t = 0.5,
  peak_sharpness = 1,
  total_island_age = 15,
  sea_level_amplitude = 0,
  sea_level_frequency = 0
)
island_ontogeny <- 1
sea_level <- 0

testthat::expect_silent(
  global_max_area_time <- DAISIE:::get_global_max_area_time(
    timeval = timeval,
    totaltime = totaltime,
    area_pars = area_pars,
    island_ontogeny = island_ontogeny,
    sea_level = sea_level
  )
)

}
\author{
Pedro Neves, Joshua Lambert, Shu Xie
}
