% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_thor_half}
\alias{get_thor_half}
\title{Creates the horizon time for the cladogenesis and immigration
rate.}
\usage{
get_thor_half(timeval, totaltime, Apars, ext_multiplier, island_ontogeny,
  thor_c_i)
}
\arguments{
\item{timeval}{}

\item{totaltime}{}

\item{Apars}{area parameters, 
as created by \link{create_area_params}}

\item{ext_multiplier}{}

\item{island_ontogeny}{}

\item{thor_c_i}{time horizon clagogesis and immigration}
}
\description{
Creates the horizon time for the cladogenesis and immigration
rate.
}
