% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{is.odd}
\alias{is.odd}
\title{Checks whether an input is odd}
\usage{
is.odd(x)
}
\arguments{
\item{x}{Object to determine}
}
\value{
A boolean indicating if object is odd
}
\description{
Checks whether an input is odd
}
\examples{
DAISIE::is.odd(5)
}
