% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_island_ontogeny.R
\name{is_island_ontogeny_runtime}
\alias{is_island_ontogeny_runtime}
\title{Check if island_ontogeny is correct during runtime (i.e. numeric)}
\usage{
is_island_ontogeny_runtime(island_ontogeny)
}
\arguments{
\item{island_ontogeny}{Character string that can be \code{"const"},
\code{"linear"} or \code{"beta"} depending on type of island ontogeny desired}
}
\value{
Boolean stating if island_ontogeny is correct.
}
\description{
Check if island_ontogeny is correct during runtime (i.e. numeric)
}
\seealso{
is_island_ontogeny_runtime
}
