% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{is_odd}
\alias{is_odd}
\title{Checks whether an input is odd}
\usage{
is_odd(x)
}
\arguments{
\item{x}{Object to determine}
}
\value{
A boolean indicating if object is odd
}
\description{
Checks whether an input is odd
}
\examples{
oddness <- DAISIE:::is_odd(5)
}
