% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_check_format.R
\name{is_sea_level_input}
\alias{is_sea_level_input}
\title{Check if sea_level is correct after user input}
\usage{
is_sea_level_input(sea_level)
}
\arguments{
\item{sea_level}{Character string that can be \code{"const"} or
\code{"sine"} depending on if sea-level is speified}
}
\value{
Boolean stating if sea_level is correct.
}
\description{
Check if sea_level is correct after user input
}
\seealso{
is_sea_level_runtime
}
