% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_check_format.R
\name{is_sea_level_runtime}
\alias{is_sea_level_runtime}
\title{Check if sea_level is correct during runtime (i.e. numeric)}
\usage{
is_sea_level_runtime(sea_level)
}
\arguments{
\item{sea_level}{Character string that can be \code{"const"},
or \code{"sine"} depending on type of island ontogeny desired}
}
\value{
Boolean stating if sea_level is correct.
}
\description{
Check if sea_level is correct during runtime (i.e. numeric)
}
\seealso{
is_sea_level_input
}
