% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_simulation_outputs.R
\name{is_simulation_outputs}
\alias{is_simulation_outputs}
\title{Measures if the input is a valid collection of simulation
outputs.}
\usage{
is_simulation_outputs(simulation_outputs)
}
\arguments{
\item{simulation_outputs}{A list with matrices? of simulation produced by
DAISIE_sim.}
}
\value{
TRUE of the input is a valid collection of simulation
outputs.
}
\description{
Measures if the input is a valid collection of simulation
outputs.
}
\examples{
library(testthat)
 
expect_false(is_simulation_outputs("nonsense"))   
}
\author{
Richel J.C Bilderbeek, Pedro Neves
}
