% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{jitter_initparsopt}
\alias{jitter_initparsopt}
\title{Jitter initial ML parameters}
\usage{
jitter_initparsopt(initparsopt, optimmethod)
}
\arguments{
\item{initparsopt}{The initial values of the parameters that must be
optimized, they are all positive.}

\item{optimmethod}{Method used in likelihood optimization. Default is
"subplex" (see subplex package). Alternative is 'simplex' which was the
method in previous versions.}
}
\value{
A numeric vector, \code{initparsopt}, for which elements that are 1
  have been added 1e-5.
}
\description{
Jitter initial ML parameters
}
\examples{
initparsopt <- c(1, 0.5, 20, 0.001, 0.5)
suppressMessages(
  initparsopt <- DAISIE:::jitter_initparsopt(
    initparsopt = initparsopt,
    optimmethod = "subplex"
  )
)
testit::assert(initparsopt == c(1.00001, 0.5, 20, 0.001, 0.5))
}
\author{
Pedro Neves, Joshua Lamebert
}
