% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_integrate.R
\name{mcVectorize}
\alias{mcVectorize}
\title{Vectorize a function and allow multithreading}
\usage{
mcVectorize(
  FUN,
  vectorize.args = arg.names,
  SIMPLIFY = TRUE,
  USE.NAMES = TRUE,
  mc.preschedule = TRUE,
  mc.set.seed = TRUE,
  mc.silent = FALSE,
  mc.cores = getOptions("mc.cores", 2L),
  mc.cleanup = TRUE
)
}
\arguments{
\item{FUN}{function to apply, found via \code{\link[base]{match.fun}}.}

\item{vectorize.args}{a character vector of arguments which should be
    vectorized.  Defaults to all arguments of \code{FUN}.}

\item{SIMPLIFY}{logical or character string; attempt to reduce the
    result to a vector, matrix or higher dimensional array; see
    the \code{simplify} argument of \code{\link[base]{sapply}}.}

\item{USE.NAMES}{logical; use names if the first \dots argument has
    names, or if it is a character vector, use that character vector as
    the names.}

\item{mc.preschedule}{
    Ignored on Windows.}

\item{mc.set.seed}{
    Ignored on Windows.}

\item{mc.silent}{
    Ignored on Windows.}

\item{mc.cores}{Number of cores to use}

\item{mc.cleanup}{
    Ignored on Windows.}
}
\value{
A function with the same arguments as FUN wrapping
a call to \code{mcmapply}
}
\description{
Vectorize a function and allow multithreading
}
