% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{pick_timeval}
\alias{pick_timeval}
\title{Calculates when the next timestep will be.}
\usage{
pick_timeval(rates, timeval)
}
\arguments{
\item{rates}{list of numeric with probabilities of each event}

\item{timeval}{current time of simulation}
}
\description{
Calculates when the next timestep will be.
}
