% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_elements.R
\name{remove_empty_elements}
\alias{remove_empty_elements}
\title{Remove superfluous lines in empty simulations}
\usage{
remove_empty_elements(sims)
}
\arguments{
\item{sims}{List with output of DAISIE_sim}
}
\value{
Output of DAISIE_sim without emptu elements
}
\description{
Remove superfluous lines in empty simulations
}
\keyword{internal}
