% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_integrate.R
\name{rho}
\alias{rho}
\title{Gamma distribution density parameterised with mean and standard deviation}
\usage{
rho(DAISIE_par, DAISIE_dist_pars)
}
\arguments{
\item{DAISIE_par}{A numeric parameter to evaluate the integral of the
function.}

\item{DAISIE_dist_pars}{A numeric vector of two elements, first is the mean
and second the standard deviation of the distribution.}
}
\value{
Numeric
}
\description{
Gamma distribution density parameterised with mean and standard deviation
}
