% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{sample_relaxed_dist}
\alias{sample_relaxed_dist}
\title{Samples from distribution when parameter is relaxed}
\usage{
sample_relaxed_dist(pars, relaxed_rate_dist, relaxed_rate_pars)
}
\arguments{
\item{pars}{A numeric vector containing the model parameters:
\itemize{
  \item{pars[1]: lambda^c (cladogenesis rate)}
  \item{pars[2]: mu (extinction rate)}
  \item{pars[3]: K (carrying capacity), set K=Inf for diversity
  independence.}
  \item{pars[4]: gamma (immigration rate)}
  \item{pars[5]: lambda^a (anagenesis rate)}
  \item{pars[6]: lambda^c (cladogenesis rate) for either type 2 species
  or rate set 2 in rate shift model}
  \item{pars[7]: mu (extinction rate) for either type 2 species or rate
  set 2 in rate shift model}
  \item{pars[8]: K (carrying capacity) for either type 2 species or rate
  set 2 in rate shift model, set K=Inf for diversity independence.}
  \item{pars[9]: gamma (immigration rate) for either type 2 species or
  rate set 2 in rate shift model}
  \item{pars[10]: lambda^a (anagenesis rate) for either type 2 species
  or rate set 2 in rate shift model}
}
The elements 6:10 are required only when type 2 species are included
or in the rate shift model.}

\item{relaxed_rate_dist}{string defining which distribution to sample from,
options are gamma or normal for the null model.}

\item{relaxed_rate_pars}{Named list of parameters for parameterising the
distribution to be sampled.}
}
\value{
A vector of parameters.
}
\description{
Samples from distribution when parameter is relaxed
}
