% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{sample_relaxed_rate}
\alias{sample_relaxed_rate}
\title{Samples from distribution when parameter is relaxed}
\usage{
sample_relaxed_rate(pars, relaxed_par, relaxed_rate_pars)
}
\arguments{
\item{pars}{A vector of 5 elements with the model parameters}

\item{relaxed_par}{A string determining which parameter is relaxed}

\item{relaxed_rate_pars}{A list of two numbers, element one is the
distribution mean, element two is the distribution standard deviation (sd)}
}
\value{
A vector of parameters.
}
\description{
Samples from distribution when parameter is relaxed
}
\keyword{internal}
