% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{set_default_pars}
\alias{set_default_pars}
\title{Sets standard metaparameters to defaults when NULL}
\usage{
set_default_pars(
  island_ontogeny,
  sea_level,
  hyper_pars,
  dist_pars,
  ext_pars,
  totaltime,
  pars
)
}
\arguments{
\item{island_ontogeny}{a numeric describing the type of island ontogeny.
Can be \code{0} for constant, \code{1} for a beta function describing area.}

\item{sea_level}{a string describing the type of sea level.
Can be \code{"const"} or \code{"sine"} for a sine function describing area}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations:
\itemize{
\item{[1]: is d_0 the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: is x the exponent for calculating extinction rate}
\item{[3]: is alpha, the exponent for calculating the immigration rate}
\item{[4]: is beta the exponent for calculating the anagenesis rate.}
}}

\item{dist_pars}{a numeric for the distance from the mainland.}

\item{ext_pars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}
}
\value{
List with standard metaparameters
}
\description{
Sets standard metaparameters to defaults when NULL
}
