% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{sim_mainland}
\alias{sim_mainland}
\title{Simulates mainland extinction as a pure-death process with replacement
either via dispersal of a new species into the mainland species pool or via
the speciation of an exisiting mainland species.}
\usage{
sim_mainland(time, M, mu_m, replacement)
}
\arguments{
\item{time}{numeric determing duration of the simulation in million years}

\item{M}{numeric determining size of the mainland species pool}

\item{mu_m}{numeric determining mainland extinction rate in extinction
events per lineage per million years.}

\item{replacement}{string of either "dispersal" or "speciation".}
}
\value{
a list where each element is a mainland lineage, this could be a
single lineage in the case of replacement by dispersal or a single or
multiple lineages in the case of replacement by speciation. Each element of
the list is a matrix with nine columns. The columns contain
\enumerate{
    \item Species identity
    \item Mainland ancestor identity
    \item Colonisation time
    \item Species type
    \item Branching code
    \item Branching time (forwards in time from the start of the simulation)
    \item Anagenetic origin
    \item Species origination time
    \item Species extinction time
}
}
\description{
Simulates mainland extinction as a pure-death process with replacement
either via dispersal of a new species into the mainland species pool or via
the speciation of an exisiting mainland species.
}
\keyword{internal}
