% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAISIE_loglik_CS.R
\name{TRAISIE_loglik_CS}
\alias{TRAISIE_loglik_CS}
\title{testing function, for comparison with DAISIE}
\usage{
TRAISIE_loglik_CS(
  parameter,
  datalist,
  methode = "odeint::runge_kutta_cash_karp54",
  atol = 1e-15,
  rtol = 1e-15,
  num_observed_states,
  num_hidden_states,
  cond = 1,
  num_threads = 1,
  verbose = FALSE
)
}
\arguments{
\item{datalist}{Data object containing information on colonisation and
  branching times. This object can be generated using the DAISIE_dataprep
  function, which converts a user-specified data table into a data object,
  but the object can of course also be entered directly.
  It is an R list object with the following elements.\cr
  The first element of the list has two or three components: \cr \cr
  \code{$island_age} - the island age \cr
  Then, depending on whether a distinction between types is
  made, we have:\cr
 \code{$not_present} - the number of mainland lineages
  that are not present on the island \cr
  or:\cr
 \code{$not_present_type1} -
  the number of mainland lineages of type 1 that are not present on the
  island \cr
 \code{$not_present_type2} - the number of mainland lineages of
  type 2 that are not present on the island \cr \cr
 The remaining elements of
  the list each contains information on a single colonist lineage on the
  island and has 5 components:\cr \cr
 \code{$colonist_name} - the name of the
  species or clade that colonized the island \cr
 \code{$branching_times} -
  island age followed by stem age of the population/species in the case of
  Non-endemic, Non-endemic_MaxAge species and Endemic species with no close relatives
  on the island. For endemic clades with more than one species on the island
  (cladogenetic clades/ radiations) these should be island age followed by the
  branching times of the island clade including the stem age of the clade\cr
 \code{$stac} - the status of the colonist \cr \cr
- Non_endemic_MaxAge: 1 \cr
- Endemic: 2 \cr
- Endemic&Non_Endemic: 3 \cr
- Non_Endemic: 4 \cr
- Endemic_Singleton_MaxAge: 5 \cr
- Endemic_Clade_MaxAge: 6 \cr
- Endemic&Non_Endemic_Clade_MaxAge: 7 \cr
- Non_endemic_MaxAge_MinAge: 8 \cr
- Endemic_Singleton_MaxAge_MinAge: 9 \cr \cr
 \code{$missing_species} - number of island species that were not
  sampled for particular clade (only applicable for endemic clades) \cr
 \code{$type1or2} - whether the colonist belongs to type 1 or type 2 \cr}

\item{methode}{Method of the ODE-solver. Supported Boost \code{ODEINT}
solvers (steppers) are:
\code{"odeint::runge_kutta_cash_karp54"}
\code{"odeint::runge_kutta_fehlberg78"}
\code{"odeint::runge_kutta_dopri5"}
\code{"odeint::bulirsch_stoer"}
without \code{odeint::}-prefix, \code{\link[deSolve]{ode}} method is
assumed. The default method overall is
\code{"lsodes"} for \code{\link{DAISIE_ML_CS}()}
and \code{"ode45"} from \code{\link[deSolve]{ode}()} for
\code{\link{DAISIE_ML_IW}()}.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{num_observed_states}{The number of observed trait states.}

\item{num_hidden_states}{The number of hidden trait states.}

\item{cond}{cond = 0 : conditioning on island age \cr cond = 1 :
conditioning on island age and non-extinction of the island biota \cr.
cond > 1 : conditioning on island age and having at least cond colonizations
on the island. This last option is not yet available for the IW model \cr}

\item{num_threads}{number of threads to be used. Default is one thread.}

\item{verbose}{A numeric vector of length 1, which in simulations and
`DAISIEdataprep()` can be `1` or `0`, where `1` gives intermediate output
should be printed.
For ML functions a numeric determining if intermediate output should be
printed. The default: `0` does not print, `1` prints the initial
likelihood and the settings that were selected (which parameters are
to be optimised, fixed or shifted), `2` prints the same as `1 and also the
intermediate output of the parameters and loglikelihood, while `3` the
same as `2` and prints intermediate progress during likelihood computation.}
}
\description{
This function compute the likelihood that all species that colonize the island
have gone extinct prior to the present.
}
