% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAISIE_logpES_max_min_age_hidden.R
\name{TRAISIE_logpES_max_min_age_hidden}
\alias{TRAISIE_logpES_max_min_age_hidden}
\title{testing function, for comparison with DAISIE}
\usage{
TRAISIE_logpES_max_min_age_hidden(
  datalist,
  brts,
  parameter,
  trait,
  num_observed_states,
  num_hidden_states,
  trait_mainland_ancestor = NA,
  status,
  sampling_fraction,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "odeint::runge_kutta_cash_karp54"
)
}
\arguments{
\item{datalist}{Data object containing information on colonisation and
  branching times. This object can be generated using the DAISIE_dataprep
  function, which converts a user-specified data table into a data object,
  but the object can of course also be entered directly.
  It is an R list object with the following elements.\cr
  The first element of the list has two or three components: \cr \cr
  \code{$island_age} - the island age \cr
  Then, depending on whether a distinction between types is
  made, we have:\cr
 \code{$not_present} - the number of mainland lineages
  that are not present on the island \cr
  or:\cr
 \code{$not_present_type1} -
  the number of mainland lineages of type 1 that are not present on the
  island \cr
 \code{$not_present_type2} - the number of mainland lineages of
  type 2 that are not present on the island \cr \cr
 The remaining elements of
  the list each contains information on a single colonist lineage on the
  island and has 5 components:\cr \cr
 \code{$colonist_name} - the name of the
  species or clade that colonized the island \cr
 \code{$branching_times} -
  island age followed by stem age of the population/species in the case of
  Non-endemic, Non-endemic_MaxAge species and Endemic species with no close relatives
  on the island. For endemic clades with more than one species on the island
  (cladogenetic clades/ radiations) these should be island age followed by the
  branching times of the island clade including the stem age of the clade\cr
 \code{$stac} - the status of the colonist \cr \cr
- Non_endemic_MaxAge: 1 \cr
- Endemic: 2 \cr
- Endemic&Non_Endemic: 3 \cr
- Non_Endemic: 4 \cr
- Endemic_Singleton_MaxAge: 5 \cr
- Endemic_Clade_MaxAge: 6 \cr
- Endemic&Non_Endemic_Clade_MaxAge: 7 \cr
- Non_endemic_MaxAge_MinAge: 8 \cr
- Endemic_Singleton_MaxAge_MinAge: 9 \cr \cr
 \code{$missing_species} - number of island species that were not
  sampled for particular clade (only applicable for endemic clades) \cr
 \code{$type1or2} - whether the colonist belongs to type 1 or type 2 \cr}

\item{brts}{Numeric vector of branching times}

\item{num_observed_states}{The number of observed trait states.}

\item{num_hidden_states}{The number of hidden trait states.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{Method of the ODE-solver. Supported Boost \code{ODEINT}
solvers (steppers) are:
\code{"odeint::runge_kutta_cash_karp54"}
\code{"odeint::runge_kutta_fehlberg78"}
\code{"odeint::runge_kutta_dopri5"}
\code{"odeint::bulirsch_stoer"}
without \code{odeint::}-prefix, \code{\link[deSolve]{ode}} method is
assumed. The default method overall is
\code{"lsodes"} for \code{\link{DAISIE_ML_CS}()}
and \code{"ode45"} from \code{\link[deSolve]{ode}()} for
\code{\link{DAISIE_ML_IW}()}.}
}
\description{
this function calculates the likelihood of observing a singleton endemic species on an island
with the trait state `i`, and for which only the estimated maximum and minimum ages of colonization are known.
}
\examples{
library(DAISIE)
data("Biwa_datalist")
datalist <- Biwa_datalist

datalist[[1]]$Mainland_pool_sizes <- c(500, 400)
datalist[[1]]$M <- 1000

parameter <- list(
  c(2.546591, 2.546591, 2.546591, 2.546591),
  c(2.678781, 2.678781, 2.678781, 2.678781),
  c(0.009326754, 0.009326754, 0.009326754, 0.009326754),
  c(1.008583, 1.008583, 1.008583, 1.008583),
  matrix(c(
    0,    0,    0,  0,
    0,    0,    0.00,0.00,
    rep(0, 8)
  ), nrow = 4, byrow = TRUE),
  1
)


parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    .001,    0.005,  0,
    0,    0,    0.002,0.005,
    0,    .1000,    0,  0,
    0,    0,    0.100,0.00
  ), nrow = 4, byrow = TRUE),
  1
)




TRAISIE_logpES_max_min_age_hidden(
  datalist              = datalist,
  brts                  = c(4, 3.9999, 0.001),
  trait                 = 0,
  status                = 9,
  parameter             = parameter,
  num_observed_states   = 2,
  num_hidden_states     = 2,
  atol                  = 1e-15,
  rtol                  = 1e-15,
  methode               = "ode45",
  trait_mainland_ancestor = c(1, 0),
  sampling_fraction     = c(1,1),
)
}
