% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_integrate.R
\name{transform_gamma_pars}
\alias{transform_gamma_pars}
\title{Transforms mean and standard deviation to shape and scale gamma parameters}
\usage{
transform_gamma_pars(par_mean, par_sd)
}
\arguments{
\item{par_mean}{mean of the relaxed parameter}

\item{par_sd}{standard deviation of the relaxed parameter}
}
\value{
list to shape and scale parameters
}
\description{
Transforms mean and standard deviation to shape and scale gamma parameters
}
