% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{translate_sea_level}
\alias{translate_sea_level}
\title{Translate user-friendly sea-level codes to numerics}
\usage{
translate_sea_level(sea_level)
}
\value{
Numeric, 0 for null-sea-level, 1 for sine function
}
\description{
Translate user-friendly sea-level codes to numerics
}
\examples{
translate_sea_level("const")
}
