% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{update_rates_mainland_ex}
\alias{update_rates_mainland_ex}
\title{Calculates algorithm rates for mainland extinction}
\usage{
update_rates_mainland_ex(
  timeval,
  totaltime,
  gam,
  laa,
  lac,
  mu,
  K,
  num_spec,
  num_immigrants,
  mainland_n,
  island_spec = NULL
)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{gam}{A numeric with the per capita immigration rate.}

\item{laa}{A numeric with the per capita anagenesis rate.}

\item{lac}{A numeric with the per capita cladogenesis rate.}

\item{mu}{A numeric with the per capita extinction rate.}

\item{K}{A numeric with carrying capacity.}

\item{num_spec}{A numeric with the current number of species.}

\item{num_immigrants}{A numeric with the current number of non-endemic
species (a.k.a non-endemic species).}

\item{mainland_n}{A numeric stating the number of mainland species, that
is the number of species that can potentially colonize the island.
If using a clade-specific diversity dependence, this value is set to 1.
If using an island-wide diversity dependence, this value is set to the
number of mainland species.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
a named list with the updated effective rates.
}
\description{
Internal function that updates the all the rates at time t.
}
\seealso{
Other rate calculations: 
\code{\link{DAISIE_plot_area}()},
\code{\link{get_ana_rate}()},
\code{\link{get_ext_rate}()},
\code{\link{get_immig_rate}()},
\code{\link{island_area_vector}()},
\code{\link{island_area}()},
\code{\link{update_max_rates}()},
\code{\link{update_rates}()}
}
\concept{rate calculations}
\keyword{internal}
