% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{update_thor_timeval}
\alias{update_thor_timeval}
\title{Realigns timeval with thor and updates thor}
\usage{
update_thor_timeval(timeval, totaltime, Apars, ext_multiplier, island_ontogeny,
  thor)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{simulated amount of time}

\item{Apars}{a named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{ext_multiplier}{reduces or increases distance of horizon to current
simulation time}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{NULL},
\code{quadratic} for a beta function describing area through time,
 or \code{linear} for a linear function}

\item{thor}{time of horizon for max extinction}
}
\value{
numeric list with updated thor and updated timeval
}
\description{
Realigns timeval with thor and updates thor
}
\author{
Pedro Neves
}
