% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_rect.R
\name{white_rect}
\alias{white_rect}
\title{Add white rectangles to a ggplot}
\usage{
white_rect(xmin, xmax, ymin, ymax, color = "white")
}
\arguments{
\item{xmin, xmax, ymin, ymax}{Coordinates of the rectangle's edges}

\item{color}{Filling and border color of the rectangle. Defaults to white.}
}
\value{
A `geom` to add to a `ggplot`
}
\description{
Add white rectangles to a ggplot
}
