% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_data.R
\docType{data}
\name{archipelago_data}
\alias{archipelago_data}
\title{Physical data on 41 archipelagos}
\format{
A dataframe containing information on archipelago name, area, age and distance from the mainland
}
\source{
Valente L, Phillimore AB, Melo M, Warren BH, Clegg SM, Havenstein K,
 Tiedemann R, Illera JC, Thébaud C, Aschenbach T, Etienne RS. A simple dynamic model
 explains island bird diversity worldwide (2020) Nature, 579, 92-96 \cr
}
\description{
A dataframe with in subsequent columns the name of the archipelago (Archipelago)
the area of the archipelago (Area), the age (Age) and the distance from the mainland (Distance)
}
\keyword{datasets}
