% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_rates.R
\name{are_rates}
\alias{are_rates}
\title{Determines if object x are rates}
\usage{
are_rates(x)
}
\arguments{
\item{x}{object to be determined to be rates}
}
\value{
TRUE if object x is a list of rates
}
\description{
Determines if object x are rates
}
