% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{calc_next_timeval}
\alias{calc_next_timeval}
\title{Calculates when the next timestep will be.}
\usage{
calc_next_timeval(rates, timeval)
}
\arguments{
\item{rates}{list of numeric with probabilities of each event}

\item{timeval}{current time of simulation}
}
\value{
named list with numeric vector containing the time of the next
timestep and the change in time.
}
\description{
Calculates when the next timestep will be.
}
\author{
Pedro Neves
}
