% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{check_island_state}
\alias{check_island_state}
\title{Checks if any changes on the mainland change the state of the island species}
\usage{
check_island_state(timeval, island_spec, mainland)
}
\arguments{
\item{timeval}{Numeric defining current time of simulation.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\value{
matrix
}
\description{
Checks if any changes on the mainland change the state of the island species
}
