% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{countspecies}
\alias{countspecies}
\title{Counts the number of species}
\usage{
countspecies(datalistelement)
}
\arguments{
\item{datalistelement}{stub}
}
\value{
A numeric
}
\description{
Counts the number of species
}
