% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{create_CS_version}
\alias{create_CS_version}
\title{Creates the list object for CS_version argument in DAISIE_ML_CS}
\usage{
create_CS_version(model = 1, relaxed_par = NULL)
}
\arguments{
\item{model}{the CS model to run, options are \code{1} for single rate
DAISIE model, \code{2} for multi-rate DAISIE, or \code{0} for IW test
model}

\item{relaxed_par}{the parameter to relax (integrate over). Options are
\code{"cladogenesis"}, \code{"extinction"}, \code{"carrying_capacity"},
\code{"immigration"}, or \code{"anagenesis"}}
}
\value{
A list of two elements
\itemize{
  \item{model: the CS model to run, options are \code{1} for single rate
  DAISIE model, \code{2} for multi-rate DAISIE, or \code{0} for IW test
  model}
  \item{relaxed_par: the parameter to relax (integrate over). Options are
\code{"cladogenesis"}, \code{"extinction"}, \code{"carrying_capacity"},
\code{"immigration"}, or \code{"anagenesis"}}
}
}
\description{
Creates the list object for CS_version argument in DAISIE_ML_CS
}
