% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_default_pars}
\alias{create_default_pars}
\title{Creates standard metaparameters to defaults when NULL}
\usage{
create_default_pars(
  island_ontogeny = 0,
  sea_level = 0,
  area_pars = NULL,
  hyper_pars = NULL,
  totaltime
)
}
\arguments{
\item{island_ontogeny}{In \code{\link{DAISIE_sim_time_dependent}},
\code{\link{DAISIE_ML_CS}} and plotting a string describing the type of
island ontogeny. Can be \code{"const"}, \code{"beta"} for a beta function
describing area through time. String checked by
\code{\link{is_island_ontogeny_input}}. \cr In all other functions a numeric
describing the type of island ontogeny. Can be \code{0} for constant,
\code{1} for a beta function describing area through time. In ML functions
\code{island_ontogeny = NA} assumes constant ontogeny.}

\item{sea_level}{In \code{\link{DAISIE_sim_time_dependent}} and plotting a
string describing the type of sea level. Can be \code{"const"} or
\code{"sine"} for a sine function describing area through time. String
checked by \code{\link{is_sea_level_input}}.
\cr In all other functions a numeric describing the type of sea level. Can
be \code{0} for constant, \code{1} for a sine function describing area
through time.}

\item{area_pars}{A named list containing area and sea level parameters as
created by \code{\link{create_area_pars}()}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: current area}
  \item{[3]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
  \item{[7]: angle of the slope of the island}
}}

\item{hyper_pars}{A named list of numeric hyperparameters for the rate
calculations as returned by \code{\link{create_hyper_pars}}:
\itemize{
\item{[1]: is d the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: is x the exponent for calculating extinction rate}
}}

\item{totaltime}{Numeric defining the length of the simulation in time units.}
}
\value{
Named list with standard parameters and metaparameters needed by
\code{DAISIE}.
}
\description{
Creates standard metaparameters to defaults when NULL
}
\examples{
default_pars <- create_default_pars(
  island_ontogeny = 0,
  sea_level = 0,
  area_pars = create_area_pars(
    max_area = 1000,
    proportional_peak_t = 0.1,
    total_island_age = 10,
    sea_level_amplitude = 1,
    sea_level_frequency = 10,
    island_gradient_angle = 0
  ),
  hyper_pars = create_hyper_pars(
    d = 0,
    x = 0
  ),
  totaltime = 15
)

}
\author{
Joshua Lambert, Pedro Neves
}
