% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_default_pars}
\alias{create_default_pars}
\title{Creates standard metaparameters to defaults when NULL}
\usage{
create_default_pars(
  island_ontogeny = 0,
  sea_level = 0,
  area_pars = NULL,
  hyper_pars = NULL,
  dist_pars = NULL,
  ext_pars = NULL,
  totaltime,
  pars
)
}
\arguments{
\item{island_ontogeny}{a numeric describing the type of island ontogeny.}

\item{sea_level}{a numeric describing the type of sea level.}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations:
\itemize{
\item{[1]: is d_0 the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: x, the exponent for calculating extinction rate}
\item{[3]: alpha, the exponent for calculating the immigration rate}
\item{[4]: beta, the exponent for calculating the anagenesis rate.}
}}

\item{dist_pars}{a numeric for the distance from the mainland.}

\item{ext_pars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{totaltime}{Length of the simulation in time units.}

\item{pars}{Contains the model parameters:
\itemize{
  \item{[1]: lambda^c (cladogenesis rate)}
  \item{[2]: mu (extinction rate)}
  \item{[3]: K (clade-level carrying capacity). Set K=Inf for non-diversity
  dependence}
  \item{[4]: gamma (immigration rate)}
  \item{[5]: lambda^a (anagenesis rate)}
}}
}
\value{
Named list with standard parameters and metaparameters needed by
\code{DAISIE}.
}
\description{
Creates standard metaparameters to defaults when NULL
}
\examples{
default_pars <- create_default_pars(
  island_ontogeny = 0,
  sea_level = 0,
  area_pars = create_area_pars(
    max_area = 1000,
    proportional_peak_t = 0.1,
    peak_sharpness = 1,
    total_island_age = 10,
    sea_level_amplitude = 1,
    sea_level_frequency = 10,
    island_gradient_angle = 0
  ),
  hyper_pars = create_hyper_pars(
    d_0 = 0,
    x = 0,
    alpha = 0,
    beta = 0
  ),
  dist_pars = create_dist_pars(D = 25000),
  ext_pars = c(5, 10),
  totaltime = 15,
  pars = c(2, 1, 0.05, 0.001, 1)
)

}
\author{
Joshua Lambert, Pedro Neves
}
