% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{create_default_pars}
\alias{create_default_pars}
\title{Creates standard metaparameters to defaults when NULL}
\usage{
create_default_pars(
  island_ontogeny,
  sea_level,
  area_pars,
  hyper_pars,
  dist_pars,
  ext_pars,
  totaltime,
  pars
)
}
\arguments{
\item{island_ontogeny}{a numeric describing the type of island ontogeny.
Can be \code{0} for constant, \code{1} for a beta function describing area.
rate for the calculation of t_hor. Default is 0.5.}

\item{sea_level}{a string describing the type of sea level.
Can be \code{"const"} or \code{"sine"} for a sine function describing area}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations:
\itemize{
\item{[1]: is d_0 the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: is x the exponent for calculating extinction rate}
\item{[3]: is alpha, the exponent for calculating the immigration rate}
\item{[4]: is beta the exponent for calculating the anagenesis rate.}
}}

\item{dist_pars}{a numeric for the distance from the mainland.}

\item{ext_pars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}
}
\value{
List with standard metaparameters
}
\description{
Creates standard metaparameters to defaults when NULL
}
