% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_dist_pars}
\alias{create_dist_pars}
\title{Create list of distance pars}
\usage{
create_dist_pars(D)
}
\arguments{
\item{D}{A numeric defining the distance parameters for the rate
calculations.}
}
\value{
Named list with distance parameters
}
\description{
Create list of distance pars
}
\examples{
dist_pars <- create_dist_pars(D = 2500)
}
\author{
Pedro Neves, Joshua Lambert
}
