% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pars.R
\name{create_hyper_pars}
\alias{create_hyper_pars}
\title{Create list of hyperparameters}
\usage{
create_hyper_pars(d, x)
}
\arguments{
\item{d}{Numeric defining the scaling parameter for exponent for
calculating cladogenesis rate.}

\item{x}{Numeric defining the exponent for calculating extinction rate.}
}
\value{
Named list with hyperparameters
}
\description{
Create list of hyperparameters
}
\examples{
hyper_pars <- create_hyper_pars(d = 0.027, x = 0.15)
}
\author{
Pedro Neves, Joshua Lambert
}
