% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rates.R
\name{create_rates}
\alias{create_rates}
\title{Create the rates per ?island}
\usage{
create_rates(immig_rate, ext_rate, ana_rate, clado_rate, ext_rate_max,
  immig_rate_max, clado_rate_max, immig_rate1, ext_rate1, ana_rate1,
  clado_rate1, trans_rate, trans_rate1)
}
\arguments{
\item{immig_rate}{immigration rate}

\item{ext_rate}{extinction rate}

\item{ana_rate}{anagenesis rate}

\item{clado_rate}{cladogenesis rate}

\item{ext_rate_max}{maximum extinction rate}

\item{immig_rate_max}{maximum immigration rate}

\item{clado_rate_max}{maximum cladogenesis rate}

\item{immig_rate1}{immigration rate with state 1}

\item{ext_rate1}{extinction rate with state 1}

\item{ana_rate1}{anagenesis rate with state 1}

\item{clado_rate1}{cladogenesis rate with state 1}

\item{trans_rate}{transition rate}

\item{trans_rate1}{transition rate with state 1}
}
\value{
the rates, as can be verified by \link{are_rates}
}
\description{
Create the rates per ?island
}
\examples{
testit::assert(
  DAISIE:::are_rates(
    create_rates(
      immig_rate = 0.1,
      ext_rate = 0.2,
      ana_rate = 0.3,
      clado_rate = 0.4,
      ext_rate_max = 0.5,
      immig_rate_max = 0.6,
      clado_rate_max = 0.7,
      immig_rate1 = 0.1,
      ext_rate1 = 0.2,
      ana_rate1 = 0.3,
      clado_rate1 = 0.4,
      trans_rate = 0.5,
      trans_rate1 = 0.5
    
  )
)
)
}
