% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rates.R
\name{create_rates}
\alias{create_rates}
\title{Create the rates per ?island}
\usage{
create_rates(immig_rate, ext_rate, ana_rate, clado_rate, ext_rate_max,
  immig_rate_max, clado_rate_max)
}
\arguments{
\item{immig_rate}{immigration rate}

\item{ext_rate}{extinction rate}

\item{ana_rate}{anagenesis rate}

\item{clado_rate}{cladogenesis rate}

\item{ext_rate_max}{maximum extinction rate}

\item{immig_rate_max}{maximum immigration rate}

\item{clado_rate_max}{maximum cladogenesis rate}
}
\value{
the rates, as can be verified by \link{are_rates}
}
\description{
Create the rates per ?island
}
\examples{
  testit::assert(
    are_rates(
      create_rates(
        immig_rate = 0.1,
        ext_rate = 0.2,
        ana_rate = 0.3,
        clado_rate = 0.4,
        ext_rate_max = 0.5,
        immig_rate_max = 0.6,
        clado_rate_max = 0.7
      )
    )
  )
}
