% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{create_test_daisie_pars}
\alias{create_test_daisie_pars}
\title{Create a function to test full-blown DAISIE parameter structure}
\usage{
create_test_daisie_pars()
}
\description{
Create a function to test full-blown DAISIE parameter structure
}
