% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_simulation_output.R
\name{create_test_simulation_outputs}
\alias{create_test_simulation_outputs}
\title{Creates the output of a as-short-as-possible
simulation run}
\usage{
create_test_simulation_outputs()
}
\value{
simulation_outputs
}
\description{
Creates the output of a as-short-as-possible
simulation run
}
\author{
Pedro Neves, Richel J.C. Bilderbeek
}
