\name{DAISIE-internal}
\alias{DAISIE_loglik_rhs}
\alias{DAISIE_loglik_rhs2}
\alias{DAISIE_loglik}
\alias{DAISIE_ML1}
\alias{DAISIE_ML2}
\alias{checkprobs}
\alias{checkprobs2}
\alias{divdepvec}
\alias{countspecies}
\alias{roundn}
\alias{counttype1}
\alias{countspeciestype1}
\alias{countimmi}
\alias{fconstr13}
\alias{fconstr15}
\alias{antidiagSums}
\alias{quantiles}
\alias{calcMN}
\alias{DAISIE_IC}
\alias{DAISIE_eq}
\alias{DAISIE_probdist_rhs}
\alias{DAISIE_numcol_dist}
\alias{DAISIE_format_CS}
\alias{DAISIE_format_IW}
\alias{DAISIE_KLdist}
\alias{DAISIE_loglik_all_choosepar}
\alias{DAISIE_loglik_all_choosepar2}
\alias{DAISIE_simplex}
\alias{DAISIE_sim_min_type2}
\alias{DAISIE_loglik_high_lambda}
\alias{Add_brt_table}
\title{Internal DAISIE Functions}
\description{Internal DAISIE functions}
\details{These are not to be called by the user.}
\keyword{internal}
