% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{DAISIE_calc_clade_ext_rate}
\alias{DAISIE_calc_clade_ext_rate}
\title{Calculate the clade-wide extinction rate}
\usage{
DAISIE_calc_clade_ext_rate(ps_ext_rate, n_species)
}
\arguments{
\item{ps_ext_rate}{per species extinction rate}

\item{n_species}{number of species in that clade}
}
\value{
the clade's extinction rate
}
\description{
Calculate the clade-wide extinction rate
}
\examples{
  testit::assert(
    DAISIE_calc_clade_ext_rate(
      ps_ext_rate = 0.2, 
      n_species = 4
    ) == 0.8
  )
}
\author{
Richel J.C. Bilderbeek
}
