% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_calc_half_life.R
\name{DAISIE_calc_half_life}
\alias{DAISIE_calc_half_life}
\title{Calculates the exact half-life from a DAISIE simulation that has reached
growth or decayed to half way between the initial conditions and the 
equilibrium}
\usage{
DAISIE_calc_half_life(island_replicates, pars, divdepmodel)
}
\arguments{
\item{island_replicates}{output from DAISIE_sim_core}

\item{pars}{the parameters used for the simulation}
}
\value{
a list of half-lives
}
\description{
Calculates the exact half-life from a DAISIE simulation that has reached
growth or decayed to half way between the initial conditions and the 
equilibrium
}
