% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sumstats_rates.R
\name{DAISIE_calc_sumstats_pcrates}
\alias{DAISIE_calc_sumstats_pcrates}
\title{Calculate summary statistics of per capita rates}
\usage{
DAISIE_calc_sumstats_pcrates(pars, Apars, Epars, totaltime,
  island_ontogeny = "quadratic", extcutoff = 1100, mainland_n = 1000,
  resol = 100)
}
\arguments{
\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}

\item{Apars}{a named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{a numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{totaltime}{A numeric indicating total time of simulation}

\item{island_ontogeny}{a string describing the type of island ontogeny. 
Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{extcutoff}{cutoff for extinction rate preventing it from being too 
large and slowing down simulation}

\item{mainland_n}{A numeric stating the number of mainland species, that
is, the number of species that can potentially colonize the island.
If \code{\link{DAISIE_sim}} uses a clade-specific diversity dependence,
this value is set to 1. 
If \code{\link{DAISIE_sim}} uses an island-specific diversity dependence,
this value is set to the number of mainland species.}

\item{resol}{A numeric indicating number of decimals per unit time.
Default is 100}
}
\value{
A named list with mean and median values of specified parameters
}
\description{
Calculates the summary statistics of per capita rates throught an ontogeny
simulation
}
\references{
Valente, Luis M., Rampal S. Etienne, and Albert B. Phillimore. 
"The effects of island ontogeny on species diversity and phylogeny." 
Proceedings of the Royal Society of London B: Biological Sciences 281.1784 (2014): 20133227.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ext_rate}

Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ana_rate}}, \code{\link{get_immig_rate}},
  \code{\link{get_t_hor}}, \code{\link{island_area}},
  \code{\link{update_rates}}
}
\author{
Pedro Neves
}
