% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_stt.R
\name{DAISIE_convert_to_classic_plot}
\alias{DAISIE_convert_to_classic_plot}
\title{Prepare input for DAISIE_stt}
\usage{
DAISIE_convert_to_classic_plot(simulation_outputs)
}
\arguments{
\item{simulation_outputs}{A list with matrices and vectors of simulation
produced by DAISIE_sim functions.}
}
\value{
a list with wrangled data to be used for plotting STT plots with
DAISIE_plot_stt
}
\description{
Prepare input for DAISIE_stt
}
\examples{
utils::data("islands_1type_1000reps", package = "DAISIE")
simulation_outuputs <- DAISIE::DAISIE_convert_to_classic_plot(
islands_1type_1000reps
)


}
\seealso{
\code{\link{DAISIE_plot_stt}}, \code{\link{DAISIE_plot_sims}}
}
