% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_create_island.R
\name{DAISIE_create_island}
\alias{DAISIE_create_island}
\title{Converts simulation output into island output}
\usage{
DAISIE_create_island(stt_table, totaltime, island_spec, mainland_n,
  keep_final_state = FALSE)
}
\arguments{
\item{stt_table}{a species-through-time table}

\item{totaltime}{simulated amount of time}

\item{island_spec}{matrix with species on island (state of system at each time point)}

\item{mainland_n}{number of mainland species}

\item{keep_final_state}{logical indicating if final state of simulation 
should be returned. Default is \code{FALSE}}
}
\value{
list with the island information, composed stt table, branching times of extant
species, status os species on the island and number of missing species.
}
\description{
Converts simulation output into island output
}
