% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_create_sim_params.R
\name{DAISIE_create_sim_params}
\alias{DAISIE_create_sim_params}
\title{Create parameters for DAISIE_sim}
\usage{
DAISIE_create_sim_params(time = 10, M = 1000, pars = c(2, 2, 40, 0.1,
  1), replicates = 100, mainland_params = NULL, divdepmodel = "CS",
  ddmodel = c(1, 0, 1), island_type = "oceanic", nonoceanic = NULL,
  prop_type2_pool = NA, replicates_apply_type2 = TRUE,
  sample_freq = 25, plot_sims = TRUE, island_ontogeny = "const",
  Apars = NULL, Epars = NULL, keep_final_state = FALSE,
  stored_data = NULL, verbose = TRUE)
}
\arguments{
\item{time}{Length of the simulation in time units.}

\item{M}{The size of mainland pool.}

\item{pars}{model parameters.}

\item{replicates}{number of simulation replicates.}

\item{mainland_params}{parameters for simulation mainland processes.}

\item{divdepmodel}{governing diversity-dependent model}

\item{ddmodel}{parameters to be diversity-dependent}

\item{island_type}{oceanic or non-oceanic island}

\item{nonoceanic}{parameters for non-oceanic island model}

\item{prop_type2_pool}{Fraction of mainland species that belongs to the
second subset of species.}

\item{replicates_apply_type2}{Default replicates_apply_type2 = TRUE runs 
simulations until the number of islands where a type 2 species has 
colonised is equal to the specified number of replicates.}

\item{sample_freq}{number of units times should be divided by.}

\item{plot_sims}{Default = TRUE plots species-through-time (STT) plots.}

\item{island_ontogeny}{a string describing the type of island ontogeny. 
Can be \code{"const"}, \code{"beta"}.}

\item{Apars}{A numeric vector:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: vale from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{keep_final_state}{logical indicating if final state of simulation 
should be returned. Default is \code{FALSE}.}

\item{stored_data}{output of DAISIE_sim function when run with 
keep_final_state. If not \code{NULL}.}

\item{verbose}{\code{Default = TRUE} Give intermediate output.}
}
\value{
list of parameters to run the model
}
\description{
Create parameters for DAISIE_sim
}
\examples{
DAISIE_create_sim_half_life_params(
time = 10,
M = 1000,
pars = c(2, 2, 40, 0.1, 1),
replicates = 100,
mainland_params = NULL,
divdepmodel = "CS",
ddmodel = c(1, 0, 1),
island_type = "oceanic",
nonoceanic = NULL,
prop_type2_pool = NA,
replicates_apply_type2 = TRUE,
sample_freq = 25,
plot_sims = TRUE,
island_ontogeny = "const",
Apars = NULL,
Epars = NULL,
keep_final_state = FALSE,
stored_data = NULL,
verbose = TRUE)
}
