% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_loglik_CS.R
\name{DAISIE_DE_loglik_CS}
\alias{DAISIE_DE_loglik_CS}
\title{Computes the loglikelihood of the DAISIE_DE model given data and a set
of model parameters}
\usage{
DAISIE_DE_loglik_CS(
  pars1,
  pars2,
  datalist,
  methode = "odeint::runge_kutta_cash_karp54",
  abstolint = 1e-15,
  reltolint = 1e-15,
  equal_extinction = TRUE,
  sampling = "n"
)
}
\arguments{
\item{pars1}{Contains the model parameters: \cr \cr
\code{pars1[1]} corresponds to lambda^c (cladogenesis rate) \cr
\code{pars1[2]} corresponds to mu (extinction rate of endemic species) \cr
\code{pars1[3]} corresponds to mu2 (extinction rate of non-endemic species) \cr
\code{pars1[4]} corresponds to gamma (immigration rate) \cr
\code{pars1[5]} corresponds to lambda^a (anagenesis rate) \cr}

\item{pars2}{Contains the model settings \cr \cr
\code{pars2[1]} irrelevant for DAISIE_DE \cr
\code{pars2[2]} irrelevant for DAISIE_DE \cr
\code{pars2[3]} corresponds to cond = setting of conditioning\cr \cr
cond = 0 : conditioning on island age \cr
cond = 1 : conditioning on island age and non-extinction of the island biota \cr \cr
cond > 1 : conditioning on island age and having at least cond colonizations on the island \cr \cr
\code{pars2[4]} sets the level of verbosity. When equal to 0, no output is generated. At higher values
(1 or 2) more output will be generated.}

\item{datalist}{Data object containing information on colonisation and
branching times. This object can be generated using the DAISIE_dataprep
function, which converts a user-specified data table into a data object, but
the object can of course also be entered directly. It is an R list object
with the following elements.\cr The first element of the list has two or
three components:
\cr \cr \code{$island_age} - the island age \cr
Then, depending on whether a distinction between types is made, we have:\cr
\code{$not_present} - the number of mainland lineages that are not present
on the island \cr
or:\cr
\code{$not_present_type1} - the number of mainland lineages of type 1 that are not present on the island \cr
\code{$not_present_type2} - the number of mainland lineages of type 2 that
are not present on the island \cr \cr
The remaining elements of the list
each contains information on a single colonist lineage on the island and has
5 components:\cr \cr
\code{$colonist_name} - the name of the species or
clade that colonized the island \cr
\code{$branching_times} - island age and
stem age of the population/species in the case of Non-endemic,
Non-endemic_MaxAge and Endemic anagenetic species. For cladogenetic species
these should be island age and branching times of the radiation including
the stem age of the radiation.\cr
\code{$stac} - the status of the colonist \cr \cr
- Non_endemic_MaxAge: 1 \cr
- Endemic: 2 \cr
- Endemic&Non_Endemic: 3 \cr
- Non_Endemic: 4 \cr
- Endemic_Singleton_MaxAge: 5 \cr
- Endemic_Clade_MaxAge: 6 \cr
- Endemic&Non_Endemic_Clade_MaxAge: 7 \cr
- Non_endemic_MaxAge_MinAge: 8 \cr
- Endemic_Singleton_MaxAge_MinAge: 9 \cr
\code{$missing_species} - number of island species that were not sampled for
particular clade (only applicable for endemic clades) \cr
\code{$type1or2} - whether the colonist belongs to type 1 or type 2. Currently
not implemented for DAISIE_DE \cr}

\item{methode}{Method of the ODE-solver. See package deSolve for details.
Default is "odeint::runge_kutta_cask_karp54"}

\item{abstolint}{Absolute tolerance of the integration}

\item{reltolint}{Relative tolerance of the integration}

\item{equal_extinction}{If FALSE the extinction rates of endemic and
non-endemic species are different, otherwise they are set equal in
optimization}

\item{sampling}{Determines whether n-sampling or rho-sampling should be used when
function_to_optimize = 'DAISIE_DE'.}
}
\value{
The loglikelihood
}
\description{
Computes the loglikelihood of the DAISIE_DE model given
colonization and branching times for lineages on an island, and a set of model
parameters. The output is a loglikelihood value
}
\references{
O.N. Dehayem et al. 2026. Preprint.
}
\seealso{
\code{\link{DAISIE_ML}}
}
\author{
Rampal S. Etienne & Bart Haegeman
}
