% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpES_unknown_coltime.R
\name{DAISIE_DE_logpES_unknown_coltime}
\alias{DAISIE_DE_logpES_unknown_coltime}
\title{Function to calculate the likelihood of observing an endemic singleton lineage on the island
with unknown colonization time.}
\usage{
DAISIE_DE_logpES_unknown_coltime(
  datalist,
  i,
  pars1,
  methode,
  reltolint,
  abstolint
)
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic singleton lineage
with unknown colonization time.
\item{logL1b}{ The log-likelihood value computed based on a system of differential equations.}
#'
}
\description{
This function calculates the log-likelihood of observing an endemic singleton lineage on an island
for which the exact colonization time is unknown.
}
\examples{

# Select a dataset from a DAISIE package

data(Biwa_datalist)
datalist <- Biwa_datalist

# Select an endemic lineage in the dataset
i <- 60
# Define example parameters
pars1 <- c(0.2, 0.1, 0.05, 0.02, 0.03)

# choose the method to solve the system of differential equations
<<<<<<< HEAD
log_likelihood <- DAISIE_DE_logpES_unknown_coltime(brts, missnumspec, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)
=======
log_likelihood <- DAISIE_DE_logpES_unknown_coltime(datalist, i, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)
>>>>>>> d0885cf4232f775a0460a4d70d01bbca65d98758

print(log_likelihood)
}
