% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpNE_max_age_coltime.R
\name{DAISIE_DE_logpNE_max_age_coltime}
\alias{DAISIE_DE_logpNE_max_age_coltime}
\title{Function to calculate the likelihood of observing a non-endemic lineage on the island
with a maximum age of colonization.}
\usage{
DAISIE_DE_logpNE_max_age_coltime(
  datalist,
  i,
  pars1,
  methode,
  reltolint,
  abstolint
)
}
\value{
The output is a numeric value representing the log-likelihood of observing a non-endemic singleton lineage
for which the maximum colonization time is given.
\item{logL1b}{ The log-likelihood value computed based on a system of differential equations. }
}
\description{
This function calculates the log-likelihood of observing a non-endemic lineage on an island
for which the exact colonization time is unknown, but the maximum age of colonization is given.
}
\examples{

# Select a dataset from a DAISIE package

data(Biwa_datalist)
datalist <- Biwa_datalist

# Select an endemic lineage in the dataset
i <- 40
# Define example parameters
pars1 <- c(0.2, 0.1, 0.05, 0.02, 0.03)

# choose the method to solve the system of differential equations
log_likelihood <- DAISIE_DE_logpNE_max_age_coltime(datalist, i, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)

print(log_likelihood)

}
