% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_DE_logpNE_unknown_coltime.R
\name{DAISIE_DE_logpNE_unknown_coltime}
\alias{DAISIE_DE_logpNE_unknown_coltime}
\title{Function to calculate the likelihood of observing a non-endemic lineage on the island
with unknown colonization time.}
\usage{
<<<<<<< HEAD
DAISIE_DE_logpNE_unknown_coltime(brts, missnumspec, pars1, methode, rtol, atol)
=======
DAISIE_DE_logpNE_unknown_coltime(datalist, i, pars1, methode, rtol, atol)
>>>>>>> d0885cf4232f775a0460a4d70d01bbca65d98758
}
\value{
The output is a numeric value representing the log-likelihood of observing an endemic singleton lineage
with unknown colonization time.
\item{logL0b}{ The log-likelihood value computed based on system of differential equations.}
#' @examples

# Select a dataset from a DAISIE package

data(Galapagos_datalist)
datalist <- Galapagos_datalist

# Select a non-endemic lineage in the dataset
i <- 2
# Define example parameters
pars1 <- c(0.2, 0.1, 0.05, 0.02, 0.03)

# choose the method to solve the system of differential equations
<<<<<<< HEAD
log_likelihood <- DAISIE_DE_logpNE(brts, missnumspec, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)
=======
log_likelihood <- DAISIE_DE_logpNE(datalist, i, pars1, methode = "lsodes", reltolint = 1e-16, abstolint = 1e-16)
>>>>>>> d0885cf4232f775a0460a4d70d01bbca65d98758

print(log_likelihood)
}
\description{
This function calculates the log-likelihood of observing a non-endemic lineage on an island
for which the exact colonization time is unknown.
}
