% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_extract_stt_median.R
\name{DAISIE_extract_stt_median}
\alias{DAISIE_extract_stt_median}
\title{Extract the STT median from the output of DAISIE_sim functions}
\usage{
DAISIE_extract_stt_median(island_replicates)
}
\arguments{
\item{island_replicates}{List output from
\code{\link{DAISIE_sim_core_constant_rate}},
\code{\link{DAISIE_sim_core_time_dependent}},
\code{\link{DAISIE_sim_core_constant_rate_shift}} or
\code{\link{DAISIE_sim_min_type2}} functions. Minimally, this must be a
list that has as many elements as replicates. Each element must be a list
with the elements \code{island_age}, \code{not_present} and \code{stt_all}.
\code{stt_all} must be a data frame with the column names \code{Time},
\code{nI}, \code{nA}, \code{nC} and \code{present}.}
}
\value{
a matrix (?)
}
\description{
Extract the STT median from the output of DAISIE_sim functions
}
