% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_extract_stt_median.R
\name{DAISIE_extract_stt_median}
\alias{DAISIE_extract_stt_median}
\title{Extract the STT median from the output of \code{\link{DAISIE_sim}}}
\usage{
DAISIE_extract_stt_median(island_replicates)
}
\arguments{
\item{island_replicates}{the result of \code{\link{DAISIE_sim}}}
}
\value{
a matrix (?)
}
\description{
Extract the STT median from the output of \code{\link{DAISIE_sim}}
}
