% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_GW.R
\name{DAISIE_format_GW}
\alias{DAISIE_format_GW}
\title{Formats guild-wide simulation output into standard DAISIE
list output}
\usage{
DAISIE_format_GW(island_replicates, time, M, sample_freq, island_type,
  num_guilds, start_midway = FALSE, verbose = TRUE)
}
\arguments{
\item{island_replicates}{List output from DAISIE_sim_core.}

\item{time}{Numeric double with total time of simulation.}

\item{M}{Int stating number of mainland species.}

\item{sample_freq}{Int stating how often results are
sampled for plotting.}

\item{island_type}{type of island for simulation.}

\item{num_guilds}{number of guilds on the mainland.}

\item{start_midway}{Logical stating if simulation starts at t > 0.}

\item{verbose}{Logical controling if progress is printed to console.}
}
\value{
List with GW DAISIE simulation output
}
\description{
Formats guild-wide simulation output into standard DAISIE
list output
}
