% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_IW.R
\name{DAISIE_format_IW}
\alias{DAISIE_format_IW}
\title{Formats island-wide simulation output into 
standard DAISIE list output}
\usage{
DAISIE_format_IW(island_replicates, time, M, sample_freq,
  verbose = FALSE, island_type = island_type)
}
\arguments{
\item{island_replicates}{Int stating number of replicates.}

\item{time}{Length of the simulation in time units. For examples, if an
island is known to be 4 million years old, setting time = 4 will simulate 
the entire life span of the island; setting time = 2 will stop the 
simulation at the mid-life of the island.}

\item{M}{The size of mainland pool, i.e. the number of species that can 
potentially colonize the island.}

\item{sample_freq}{Specifies the number of units times should be divided by
for plotting purposes. Larger values will lead to plots with higher 
resolution, but will also run slower.}

\item{verbose}{sets whether parameters and likelihood should be printed (1)
or not (0)}

\item{island_type}{oceanic or non-oceanic island}
}
\value{
List with CS DAISIE simulation output
}
\description{
Formats island-wide simulation output into 
standard DAISIE list output
}
