% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_format_IW.R
\name{DAISIE_format_IW}
\alias{DAISIE_format_IW}
\title{Formats island-wide simulation output into 
standard DAISIE list output}
\usage{
DAISIE_format_IW(island_replicates, time, M, sample_freq, island_type)
}
\arguments{
\item{island_replicates}{Int stating number of replicates.}

\item{time}{Numeric double with total time of simulation.}

\item{M}{Int stating number of mainland species.}

\item{sample_freq}{Int stating how often results are 
sampled for plotting.}

\item{island_type}{type of island for simulation.}
}
\value{
List with CS DAISIE simulation output
}
\description{
Formats island-wide simulation output into 
standard DAISIE list output
}
