% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_make_archipelago.R
\name{DAISIE_make_archipelago}
\alias{DAISIE_make_archipelago}
\title{Does somethings}
\usage{
DAISIE_make_archipelago(archipelago, M, phylo_data, archipelago_data)
}
\arguments{
\item{archipelago}{something}

\item{M}{Numeric defining the size of mainland pool, i.e. the number of
species that can potentially colonize the island.}

\item{phylo_data}{something}

\item{archipelago_data}{something}
}
\value{
something
}
\description{
Does somethings
}
