% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_make_archipelago.R
\name{DAISIE_make_global}
\alias{DAISIE_make_global}
\title{Does something}
\usage{
DAISIE_make_global(archipelago_list, M, phylo_data, archipelago_data)
}
\arguments{
\item{archipelago_list}{something}

\item{M}{something}

\item{phylo_data}{something}

\item{archipelago_data}{something}
}
\value{
something
}
\description{
Does something
}
