% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ML1.R
\name{DAISIE_ML1}
\alias{DAISIE_ML1}
\title{Computes MLE for single type species under a clade specific scenario}
\usage{
DAISIE_ML1(datalist, initparsopt, idparsopt, parsfix, idparsfix,
  idparsnoshift = 6:10, res = 100, ddmodel = 0, cond = 0,
  eqmodel = 0, x_E = 0.95, x_I = 0.98, tol = c(1e-04, 1e-05,
  1e-07), maxiter = 1000 * round((1.25)^length(idparsopt)),
  methode = "lsodes", optimmethod = "subplex", CS_version = 1,
  verbose = 0, tolint = c(1e-16, 1e-10), island_ontogeny = NA)
}
\value{
The output is a dataframe containing estimated parameters and
maximum loglikelihood.  \item{lambda_c}{ gives the maximum likelihood
estimate of lambda^c, the rate of cladogenesis} \item{mu}{ gives the maximum
likelihood estimate of mu, the extinction rate} \item{K}{ gives the maximum
likelihood estimate of K, the carrying-capacity} \item{gamma}{ gives the
maximum likelihood estimate of gamma, the immigration rate }
\item{lambda_a}{ gives the maximum likelihood estimate of lambda^a, the rate
of anagenesis} \item{loglik}{ gives the maximum loglikelihood} \item{df}{ gives the number
of estimated parameters, i.e. degrees of feedom} \item{conv}{ gives a
message on convergence of optimization; conv = 0 means convergence}
}
\description{
Computes MLE for single type species under a clade specific scenario
}
