% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_MW_ML.R
\name{DAISIE_MW_ML}
\alias{DAISIE_MW_ML}
\title{Maximization of the loglikelihood under the DAISIE model with clade-specific
diversity-dependence and explicit dependencies on island area and distance
from the mainland or nearest landmass as hypothesized by MacArthur & Wilson}
\usage{
DAISIE_MW_ML(datalist, initparsopt, idparsopt, parsfix, idparsfix,
  res = 100, ddmodel = 11, cond = 0, island_ontogeny = NA,
  tol = c(1e-04, 1e-05, 1e-07), maxiter = 1000 *
  round((1.25)^length(idparsopt)), methode = "lsodes",
  optimmethod = "subplex", CS_version = 1, verbose = 0,
  tolint = c(1e-16, 1e-10), distance_type = "continent",
  distance_dep = "power", parallel = "local", cpus = 3)
}
\arguments{
\item{datalist}{Data object containing information on colonisation and
branching times. This object can be generated using the DAISIE_dataprep
function, which converts a user-specified data table into a data object, but
the object can of course also be entered directly. It is an R list object
with the following elements.\cr The first element of the list has two three
components: \cr \cr \code{$island_age} - the island age \cr Then, depending
on whether a distinction between types is made, we have:\cr
\code{$not_present} - the number of mainland lineages that are not present
on the island \cr\cr The remaining elements of the list each contains
information on a single colonist lineage on the island and has 5
components:\cr \cr \code{$colonist_name} - the name of the species or clade
that colonized the island \cr \code{$branching_times} - island age and stem
age of the population/species in the case of Non-endemic, Non-endemic_MaxAge
and Endemic anagenetic species. For cladogenetic species these should be
island age and branching times of the radiation including the stem age of
the radiation.\cr \code{$stac} - the status of the colonist \cr \cr *
Non_endemic_MaxAge: 1 \cr * Endemic: 2 \cr * Endemic&Non_Endemic: 3 \cr *
Non_endemic: 4 \cr * Endemic_MaxAge: 5 \cr \cr \code{$missing_species} -
number of island species that were not sampled for particular clade (only
applicable for endemic clades)}

\item{initparsopt}{The initial values of the parameters that must be
optimized; they are all positive}

\item{idparsopt}{The ids of the parameters that must be optimized. The ids
are defined as follows: \cr \cr id = 1 corresponds to lambda^c0
(cladogenesis rate for unit area) \cr id = 2 corresponds to y (exponent of
area for cladogenesis rate) \cr id = 3 corresponds to mu0 (extinction rate
for unit area) \cr id = 4 corresponds to x (exponent of 1/area for
extinction rate) \cr id = 5 corresponds to K0 (clade-level carrying capacity
for unit area) \cr id = 6 corresponds to z (exponent of area for clade-level
carrying capacity) \cr id = 7 corresponds to gamma0 (immigration rate for
unit distance) \cr id = 8 corresponds to alpha (exponent of 1/distance for
immigration rate) \cr id = 9 corresponds to lambda^a0 (anagenesis rate for
unit distance) \cr id = 10 corresponds to beta (exponent of 1/distance for
anagenesis rate) \cr}

\item{parsfix}{The values of the parameters that should not be optimized}

\item{idparsfix}{The ids of the parameters that should not be optimized,
e.g. c(1,3) if lambda^c and K should not be optimized.}

\item{res}{Sets the maximum number of species for which a probability must
be computed, must be larger than the size of the largest clade}

\item{ddmodel}{Sets the model of diversity-dependence: \cr \cr ddmodel = 0 :
no diversity dependence \cr ddmodel = 1 : linear dependence in speciation
rate \cr ddmodel = 11: linear dependence in speciation rate and in
immigration rate \cr ddmodel = 2 : exponential dependence in speciation
rate\cr ddmodel = 21: exponential dependence in speciation rate and in
immigration rate\cr}

\item{cond}{cond = 0 : conditioning on island age \cr cond = 1 :
conditioning on island age and non-extinction of the island biota \cr}

\item{island_ontogeny}{type of island ontonogeny. If NA, then constant ontogeny is assumed}

\item{tol}{Sets the tolerances in the optimization. Consists of: \cr reltolx
= relative tolerance of parameter values in optimization \cr reltolf =
relative tolerance of function value in optimization \cr abstolx = absolute
tolerance of parameter values in optimization}

\item{maxiter}{Sets the maximum number of iterations in the optimization}

\item{methode}{Method of the ODE-solver. See package deSolve for details.
Default is "lsodes"}

\item{optimmethod}{Method used in likelihood optimization. Default is
"subplex" (see subplex package). Alternative is 'simplex' which was the
method in previous versions.}

\item{CS_version}{For internal testing purposes only. Default is 1, the
original DAISIE code.}

\item{verbose}{sets whether parameters and likelihood should be printed (1)
or not (0)}

\item{tolint}{Vector of two elements containing the absolute and relative
tolerance of the integration}

\item{distance_type}{Use 'continent' if the distance to the continent should
be used, use 'nearest_big' if the distance to the nearest big landmass
should be used, and use 'biologically_realistic' if the distance should take
into account some biologically realism, e.g. an average of the previous two
if both are thought to contribute.}

\item{distance_dep}{Sets what type of distance dependence should be used.
Default is a power law, denoted as 'power'. Alternatives are an exponantial
relationship denoted by 'exp' or sigmoids, either 'sigmoidal_col' for a
sigmoid in the colonization, 'sigmoidal_ana' for sigmoidal anagenesis,
'sigmoidal_clado' for sigmoidal cladogenesis, and 'sigmoidal_col_ana' for
signoids in both colonization and anagenesis.}

\item{parallel}{Sets whether parallel computation should be used. Use 'no'
if no parallel computing should be used, 'cluster' for parallel computing on
a unix/linux cluster, and 'local' for parallel computation on a local
machine.}

\item{cpus}{Number of cpus used in parallel computing. Default is 3. Will
not have an effect if parallel = 'no'.}
}
\value{
The output is a dataframe containing estimated parameters and
maximum loglikelihood.  \item{lambda_c0}{ gives the maximum likelihood
estimate of lambda^c, the rate of cladogenesis for unit area} \item{y}{
gives the maximum likelihood estimate of y, the exponent of area for the
rate of cladogenesis} \item{mu0}{ gives the maximum likelihood estimate of
mu0, the extinction rate} \item{x}{ gives the maximum likelihood estimate of
x, the exponent of 1/area for the extinction rate} \item{K0}{ gives the
maximum likelihood estimate of K0, the carrying-capacity for unit area}
\item{z}{ gives the maximum likelihood estimate of z, the exponent of area
for the carrying capacity} \item{gamma0}{ gives the maximum likelihood
estimate of gamma0, the immigration rate for unit distance} \item{y}{ gives
the maximum likelihood estimate of alpha, the exponent of 1/distance for the
rate of colonization} \item{lambda_a0}{ gives the maximum likelihood
estimate of lambda^a0, the rate of anagenesis for unit distance}
\item{beta}{ gives the maximum likelihood estimate of beta, the exponent of
1/distance for the rate of anagenesis} \item{loglik}{ gives the maximum
loglikelihood} \item{df}{ gives the number of estimated parameters, i.e.
degrees of feedom} \item{conv}{ gives a message on convergence of
optimization; conv = 0 means convergence}
}
\description{
This function computes the maximum likelihood estimates of the parameters of
the relationships between parameters of the DAISIE model with clade-specific
diversity-dependence and island area and distance of the island to the
mainlandor nearest landmass, for data from lineages colonizing several
islands/archipelagos. It also outputs the corresponding loglikelihood that
can be used in model comparisons.

A note on the sigmoidal functions used in distance_dep: For anagenesis and
cladogenesis, the functional relationship is k * (d/d0)^x/(1 + (d/d0)^x);
for colonization the relationship is: k - k * (d/d0)^x/(1 + (d/d0)^x). The
d0 parameter is the 11th parameter entered. In the of 'sigmoidal_col_ana',
the 11th parameter is the d0 for colonization and the 12th is the d0 for
anagenesis.
}
\examples{

cat('No examples')

}
\references{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015).
Equilibrium and non-equilibrium dynamics simultaneously operate in the
Galapagos islands. Ecology Letters 18: 844-852. <DOI:10.1111/ele.12461>.
}
\seealso{
\code{\link{DAISIE_ML_CS}},
}
\author{
Rampal S. Etienne
}
\keyword{models}
