% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DAISIE_odeint_cs}
\alias{DAISIE_odeint_cs}
\title{Driver for the boost::odeint solver}
\usage{
DAISIE_odeint_cs(runmod, y, times, lx, kk, par, Stepper, atol, rtol)
}
\description{
Driver for the boost::odeint solver
}
