% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daisie_odeint_iw}
\alias{daisie_odeint_iw}
\title{Driver for the boost::odeint solver for the IW model}
\value{
Object of type `state_type`, which itself is
`vector_t`<double>, with the result of the
integration depending on the runmod chosen.
}
\description{
Driver for the boost::odeint solver for the IW model
}
