% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DAISIE_odeint_iw}
\alias{DAISIE_odeint_iw}
\title{Driver for the boost::odeint solver}
\usage{
DAISIE_odeint_iw(y, times, pars, Stepper, atol, reltol)
}
\description{
Driver for the boost::odeint solver
}
