% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daisie_odeint_iw}
\alias{daisie_odeint_iw}
\title{odeint function made available to R}
\arguments{
\item{ry}{ry}

\item{rtimes}{rtimes}

\item{rpars}{rpars}

\item{Stepper}{Stepper}

\item{atolint}{atolint}

\item{reltolint}{reltolint}
}
\value{
output of the integration
}
\description{
odeint function made available to R
}
