% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core_1_4.R
\name{DAISIE_ONEcolonist}
\alias{DAISIE_ONEcolonist}
\title{Does something}
\usage{
DAISIE_ONEcolonist(time, island_spec, stt_table)
}
\arguments{
\item{time}{simulated amount of time}

\item{island_spec}{something}

\item{stt_table}{?Species-Through-Time table}
}
\description{
Does something
}
