% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ONEcolonist.R
\name{DAISIE_ONEcolonist}
\alias{DAISIE_ONEcolonist}
\title{Title}
\usage{
DAISIE_ONEcolonist(time, island_spec, stt_table)
}
\arguments{
\item{time}{Numeric defining the length of the simulation in time units.
For examples, if an island is known to be 4 million years old, setting
time = 4 will simulate the entire life span of the island; setting time = 2
will stop the simulation at the mid-life of the island.}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}

\item{stt_table}{Matrix with number of species at each time step.}
}
\value{
a list with these elements:
\itemize{
  item{[1]: stt_table, the same stt_table as put in}
  item{[2]: branching_times, branching times}
  item{[3]: stac, ?statuses}
  item{[4]: missing_species, ?the number of missing species}
  item{[5]: other_clades_same_ancestor, ?no idea}
  item{[6]: non-endemic species}
  item{[7]: endemic species}
  }
}
\description{
Title
}
