% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core_1_4.R
\name{DAISIE_ONEcolonist}
\alias{DAISIE_ONEcolonist}
\title{Does something}
\usage{
DAISIE_ONEcolonist(time, island_spec, stt_table,
  keep_final_state = FALSE)
}
\arguments{
\item{time}{simulated amount of time}

\item{island_spec}{matrix with current state of simulation}

\item{stt_table}{?Species-Through-Time table}

\item{keep_final_state}{logical indicating if final state of simulation 
should be returned. Default is \code{FALSE}}
}
\value{
a list with these elements:
\itemize{
  item{[1]: stt_table, the same stt_table as put in}
  item{[2]: branching_times, branching times}
  item{[3]: stac, ?statuses}
  item{[4]: missing_species, ?the number of missing species}
  item{[5]: other_clades_same_ancestor, ?no idea}
}
}
\description{
Does something
}
