% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_params.R
\name{DAISIE_plot_area}
\alias{DAISIE_plot_area}
\title{Plots island area function through time}
\usage{
DAISIE_plot_area(
  totaltime,
  area_pars,
  island_ontogeny = "beta",
  resolution,
  sea_level = "const"
)
}
\arguments{
\item{totaltime}{total time of simulation}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{NULL},
\code{"beta"} for a beta function describing area through time.}

\item{resolution}{numeric indicating resolution of plot. Should be < 0.}

\item{sea_level}{a string describing the type of sea level.
Can be \code{"const"} or \code{"sine"} for a sine function describing area}
}
\value{
a plot with the area size through time
}
\description{
Plots island area function through time
}
\seealso{
Other rate calculations: 
\code{\link{get_ana_rate}()},
\code{\link{get_ext_rate}()},
\code{\link{get_immig_rate}()},
\code{\link{island_area_vector}()},
\code{\link{island_area}()},
\code{\link{update_max_rates}()},
\code{\link{update_rates}()}
}
\concept{rate calculations}
