% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_area.R
\name{DAISIE_plot_area}
\alias{DAISIE_plot_area}
\title{Plots island area function through time}
\usage{
DAISIE_plot_area(totaltime, Apars, island_ontogeny = "quadratic",
  resolution)
}
\arguments{
\item{totaltime}{total time of simulation}

\item{Apars}{a named list containing area parameters:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{resolution}{numeric indicating resolution of plot. Should be < 0.}
}
\value{
a plot with the area size through time
}
\description{
Plots island area function through time
}
\seealso{
Other rates calculation: \code{\link{get_ana_rate}},
  \code{\link{get_ext_rate}}, \code{\link{get_immig_rate}},
  \code{\link{get_t_hor}}, \code{\link{island_area}},
  \code{\link{update_rates}}
}
\concept{rates calculation}
