% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_params.R
\name{DAISIE_plot_cladogenesis}
\alias{DAISIE_plot_cladogenesis}
\title{Plot cladogenesis rate through time}
\usage{
DAISIE_plot_cladogenesis(
  totaltime,
  K,
  area_pars,
  lac,
  island_ontogeny = "beta",
  sea_level = "const",
  hyper_pars = NULL,
  removed_timepoints,
  resolution
)
}
\arguments{
\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{K}{A numeric with carrying capacity.}

\item{area_pars}{A named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
  \item{[7]: angle of the slope of the island}
}}

\item{lac}{A numeric with the per capita cladogenesis rate.}

\item{island_ontogeny}{In \code{\link{DAISIE_sim_time_dependent}},
\code{\link{DAISIE_ML_CS}} and plotting a string describing the type of
island ontogeny. Can be \code{"const"}, \code{"beta"} for a beta function
describing area through time. String checked by
\code{\link{is_island_ontogeny_input}}. \cr In all other functions a numeric
describing the type of island ontogeny. Can be \code{0} for constant,
\code{1} for a beta function describing area through time. In ML functions
\code{island_ontogeny = NA} assumes constant ontogeny.}

\item{sea_level}{In \code{\link{DAISIE_sim_time_dependent}} and plotting a
string describing the type of sea level. Can be \code{"const"} or
\code{"sine"} for a sine function describing area through time. String
checked by \code{\link{is_sea_level_input}}.
\cr In all other functions a numeric describing the type of sea level. Can
be \code{0} for constant, \code{1} for a sine function describing area
through time.}

\item{hyper_pars}{A named list of numeric hyperparameters for the rate
calculations as returned by \code{\link{create_hyper_pars}}:
\itemize{
\item{[1]: is d the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: is x the exponent for calculating extinction rate}
}}

\item{removed_timepoints}{Positive integer with number of first datapoints
to be removed from rate plots (to prevent Inf)}

\item{resolution}{numeric indicating resolution of plot. Should be < 0.}
}
\value{
a plot with per capita cladogenesis rate through time and dataframe with immigration
at corresponding time
}
\description{
Plot cladogenesis rate through time
}
\author{
Pedro Neves
}
