% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_area.R
\name{DAISIE_plot_extinction}
\alias{DAISIE_plot_extinction}
\title{Plots extinction rate function through time}
\usage{
DAISIE_plot_extinction(island_area_time, totaltime, K, Apars, Epars,
  island_ontogeny = "quadratic", removed_timepoints)
}
\arguments{
\item{island_area_time}{something}

\item{totaltime}{something}

\item{K}{something}

\item{Apars}{something}

\item{Epars}{something}

\item{island_ontogeny}{something}

\item{removed_timepoints}{something}
}
\value{
extinction rate through time plot
}
\description{
Plots extinction rate function through time
}
