% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_area.R
\name{DAISIE_plot_immigration}
\alias{DAISIE_plot_immigration}
\title{Plot immigration rate through time}
\usage{
DAISIE_plot_immigration(island_area_time, totaltime, K, Apars, gam, mainland_n,
  island_ontogeny = "quadratic", removed_timepoints, immig_rate)
}
\arguments{
\item{island_area_time}{Something}

\item{totaltime}{Something}

\item{K}{Something}

\item{Apars}{Something}

\item{gam}{Something}

\item{mainland_n}{Something}

\item{island_ontogeny}{Something}

\item{removed_timepoints}{Something}

\item{immig_rate}{Something}
}
\value{
a plot with immigration rate through time
}
\description{
Plot immigration rate through time
}
