% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_area.R
\name{DAISIE_plot_immigration}
\alias{DAISIE_plot_immigration}
\title{Plot immigration rate through time}
\usage{
DAISIE_plot_immigration(totaltime, K, Apars, gam, mainland_n,
  island_ontogeny = "beta", removed_timepoints, resolution)
}
\arguments{
\item{totaltime}{total time of simulation}

\item{K}{K (clade-level carrying capacity)}

\item{Apars}{a named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{gam}{minimum per capita immigration rate}

\item{mainland_n}{number of mainland species. Set as 1 for clade-specific 
diversity dependence}

\item{island_ontogeny}{a string describing the type of island ontogeny. Can be \code{NULL},
\code{beta} for a beta function describing area through time,
 or \code{linear} for a linear function}

\item{removed_timepoints}{starting position of time vector}

\item{resolution}{resolution of time axis}
}
\value{
a plot with per capita immigration rate through time and dataframe with immigration 
at corresponding time
}
\description{
Plot immigration rate through time
}
\author{
Pedro Neves
}
