% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_sims.R
\name{DAISIE_plot_sims}
\alias{DAISIE_plot_sims}
\title{Plot island species-through-time (STT) plots}
\usage{
DAISIE_plot_sims(island_replicates, use_dev_new = TRUE,
  plot_plus_one = TRUE)
}
\arguments{
\item{island_replicates}{Island replicates in DAISIE format (produced in
\code{\link{DAISIE_sim}} with \code{format = TRUE} option). Minimally, this must be 
a list, that has as much elements as replicates. Each element must be a
list with the elements \code{island_age}, \code{not_present} 
and \code{stt_all}. \code{stt_all} must be a data frame with
the column names \code{Time}, \code{nI}, \code{nA}, \code{nC} 
and \code{present}.}

\item{use_dev_new}{Boolean to indicate to use \code{dev.new} before plotting. 
Set to \code{TRUE} for default behavior.
Set to \code{FALSE} when plotting within a vignette}

\item{plot_plus_one}{Boolean to indicate to plot all values plus one.
Set to \code{TRUE} for default behavior.
Set to \code{FALSE} to plot all values without adding one.
Only works when there is one type of species}
}
\value{
R plot.
}
\description{
Produces STT plots. If only one type of species is present in the simulated
islands, STT is plotted for all species. If two types are present, three
plots are produced: STT for all, STT for type 1 and STT for type 2.
}
\details{
R plots with number of total, endemic and non-endemic STTs for different
types of species for the entire time span the islands were simulated.
2.5-97.5th percentiles are plotted in light grey, 25-75th percentiles
plotted in dark grey.
}
\examples{
 

### Plot islands with single process (only one type of species)
data(islands_1type_1000reps)
DAISIE_plot_sims(island_replicates = islands_1type_1000reps)


### Plot island with type 1 and type 2
data(islands_2types_1000reps)
DAISIE_plot_sims(island_replicates = islands_2types_1000reps)


# Plot islands with single process
# Start counting from zero on the Y axis
DAISIE_plot_sims(
island_replicates = islands_1type_1000reps, 
  plot_plus_one = FALSE
)
}
\references{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015).
Equilibrium and non-equilibrium dynamics simultaneously operate in the
Galapagos islands. Ecology Letters 18: 844-852.
}
\seealso{
\code{\link{DAISIE_sim}} \code{\link{DAISIE_format_CS}}
}
\author{
Luis Valente
}
\keyword{models}
