% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_plot_stt.R
\name{DAISIE_plot_stt}
\alias{DAISIE_plot_stt}
\title{Create the Species-Through-Time plot. This is used to visualize
the output of \code{\link{DAISIE_sim}}}
\usage{
DAISIE_plot_stt(plot_plus_one = FALSE, time, stt_q0.025_all,
  stt_q0.25_all, stt_average_all, stt_q0.75_all, stt_q0.975_all)
}
\arguments{
\item{plot_plus_one}{Boolean to indicate to plot all values plus one.
Set to \code{TRUE} for default behavior.
Set to \code{FALSE} to plot all values without adding one.
Only works when there is one type of species}

\item{time}{the time span simulated}

\item{stt_q0.025_all}{STT 2.5\% quantile. 
Must be a data frame with columns name \code{Time}, \code{nI}, 
\code{Endemic} and \code{Total}}

\item{stt_q0.25_all}{STT 25\% quantile
Must be a data frame with columns name \code{Time}, \code{nI}, 
\code{Endemic} and \code{Total}}

\item{stt_average_all}{STT average
Must be a data frame with columns name \code{Time}, \code{nI}, 
\code{Endemic} and \code{Total}}

\item{stt_q0.75_all}{STT 75\% quantile
Must be a data frame with columns name \code{Time}, \code{nI}, 
\code{Endemic} and \code{Total}}

\item{stt_q0.975_all}{STT 97.5\% quantile
Must be a data frame with columns name \code{Time}, \code{nI}, 
\code{Endemic} and \code{Total}}
}
\description{
Create the Species-Through-Time plot. This is used to visualize
the output of \code{\link{DAISIE_sim}}
}
