% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{DAISIE_sim_core}
\alias{DAISIE_sim_core}
\title{Internal function of the DAISIE simulation}
\usage{
DAISIE_sim_core(time, mainland_n, pars)
}
\arguments{
\item{time}{Simulated amount of time}

\item{mainland_n}{A numeric stating the number of mainland species, that
is, the number of species that can potentially colonize the island.
If \code{\link{DAISIE_sim}} uses a clade-specific diversity dependence,
this value is set to 1. 
If \code{\link{DAISIE_sim}} uses an island-specific diversity dependence,
this value is set to the number of mainland species.}

\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}
}
\description{
Internal function of the DAISIE simulation
}
