% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{DAISIE_sim_core}
\alias{DAISIE_sim_core}
\title{Internal function of the DAISIE simulation}
\usage{
DAISIE_sim_core(time, mainland_n, pars, ddmodel_sim = 11,
  island_type = "oceanic", nonoceanic_pars = NULL,
  k_dist_pars = NULL, island_ontogeny = 0, sea_level = 0,
  hyper_pars = NULL, area_pars = NULL, dist_pars = NULL,
  ext_pars = NULL, pars_shift = FALSE, shift_times = NULL,
  ext_multiplier = 0.5)
}
\arguments{
\item{time}{Simulated amount of time}

\item{mainland_n}{A numeric stating the number of mainland species, that
is the number of species that can potentially colonize the island.
If \code{\link{DAISIE_sim}} uses a clade-specific diversity dependence,
this value is set to 1.
If \code{\link{DAISIE_sim}} uses an island-wide diversity dependence,
this value is set to the number of mainland species.}

\item{pars}{A numeric vector:
\itemize{
  \item{[1]: cladogenesis rate}
  \item{[2]: extinction rate}
  \item{[3]: carrying capacity}
  \item{[4]: immigration rate}
  \item{[5]: anagenesis rate}
}}

\item{ddmodel_sim}{The a numeric vector to determined which parameters should
be diversity dependent.}

\item{island_type}{Option island_type = 'oceanic' is a model equal to Valente
et al., 2015. island_type = 'nonoceanic' is a nonoceanic model where initial
species richness is non-zero determined by the nonoceanic parameters.}

\item{nonoceanic_pars}{A vector of length three with: the island area as a
proportion of the mainland, the probability of native species being
nonendemic and the size of the mainland pool.}

\item{k_dist_pars}{Vector of two elements to define gamma
distribution for sampling carrying capacities. \code{k_dist_pars[1]}
is the shape parameter, \code{k_dist_pars[2]} is the rate parameter}

\item{island_ontogeny}{a numeric describing the type of island ontogeny.
Can be \code{0} for constant, \code{1} for a beta function describing area.}

\item{sea_level}{a string describing the type of sea level.
Can be \code{"const"} or \code{"sine"} for a sine function describing area}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations, \code{hyper_pars[1]} is d_0 the scaling parameter for
exponent for calculating cladogenesis rate, \code{hyper_pars[2]}
is x the exponent for calculating extinction rate,
\code{hyper_pars[3]} is alpha the exponent for calculating the
immigration rate, \code{hyper_pars[4]} is beta the exponent for
calculating the anagenesis rate.}

\item{area_pars}{A named list containing area parameters as created by
create_area_pars:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{dist_pars}{a numeric for the distance from the mainland.}

\item{ext_pars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{pars_shift}{logical determining whether rate shift model runs.}

\item{shift_times}{a numeric vector specifying when the rate shifts occur
before the present.}

\item{ext_multiplier}{Numeric between 0 and 1 reducing effective extinction
rate for the calculation of t_hor. Default is 0.5.}
}
\description{
Internal function of the DAISIE simulation
}
