% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{DAISIE_sim_core_checked}
\alias{DAISIE_sim_core_checked}
\title{Runs one DAISIE simulation with a clade-specific carrying capacity.
Version of \code{DAISIE_sim_core} that checks all its inputs
and uses descriptively named arguments}
\usage{
DAISIE_sim_core_checked(sim_time, n_mainland_species, clado_rate, ext_rate,
  carr_cap, imm_rate, ana_rate)
}
\arguments{
\item{sim_time}{length of the simulated time}

\item{n_mainland_species}{number of mainland species}

\item{clado_rate}{cladogenesis rate}

\item{ext_rate}{extinction rate}

\item{carr_cap}{carrying capacity}

\item{imm_rate}{immigration rate}

\item{ana_rate}{anagenesis rate}
}
\value{
a list with these elements:
\describe{
  \item{stt_table}{a species-through-time table}
  \item{branching_times}{branching times}
  \item{stac}{
    status of the clade formed by the immigrant:
    \itemize{
      \item{0: immigrant is not present and has not formed an extant clade}
      \item{1: immigrant is present but has not formed an extant clade}
      \item{2: immigrant is not present but has formed an extant clade}
      \item{3: immigrant is present and has formed an extant clade}
      \item{4: immigrant is present but has not formed an extant clade, and it is known when it immigrated.}
      \item{5: immigrant is not present and has not formed an extant clade, but only an endemic species}
    }
  }
  \item{missing_species}{number of missing species}
  \item{other_clades_same_ancestor}{(not always present) ?no idea}
}
}
\description{
Runs one DAISIE simulation with a clade-specific carrying capacity.
Version of \code{DAISIE_sim_core} that checks all its inputs
and uses descriptively named arguments
}
\author{
Richel J.C. Bilderbeek
}
