% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_half_life.R
\name{DAISIE_sim_half_life}
\alias{DAISIE_sim_half_life}
\title{Simulates an island using DAISIE_sim_core and then calculates the half-life
and repeats this across a parameter space}
\usage{
DAISIE_sim_half_life(time, replicates, lac, mu, K, gam, laa, ssr,
  divdepmodel, ddmodel_sim, island_type, x_s, x_nonend)
}
\arguments{
\item{time}{Length of the simulation in time units. For examples, if an
island is known to be 4 million years old, setting time = 4 will simulate 
the entire life span of the island; setting time = 2 will stop the 
simulation at the mid-life of the island.}

\item{replicates}{Number of island replicates to be simulated.}

\item{lac}{numeric vector with all values of cladogenesis giving the 
parameter space of cladogenesis.}

\item{mu}{numeric vector with all values of extinction giving the parameter
space of extinction.}

\item{K}{numeric vector with all values of the carrying capacity giving the 
parameter space of the carrying capacity.}

\item{gam}{numeric vector with all values of immigration giving the parameter
space of immigration.}

\item{laa}{numeric vector with all values of anagenesis giving the parameter
space of anagenesis.}

\item{ssr}{numeric vector with all the supersaturation ratios (ssr) giving
the parameter space of the supersaturation ratio.}

\item{divdepmodel}{Option divdepmodel = 'CS' runs a model with clade-specific 
carrying capacity, where diversity-dependence operates only within single 
clades, i.e. only among species originating from the same mainland colonist.
Option divdepmodel = 'IW' runs a model with island-wide carrying capacity,
where diversity-dependence operates within and among clades.}

\item{ddmodel_sim}{numeric vector specifying which diversity-depedent models
should be used. \code{ddmodel_sim == 0} is a diversity-independent (constant
rate) model; \code{ddmodel_sim == 1} is a linear diversity-depedent 
cladogenesis model; \code{ddmodel_sim == 11} is a linear diversity-
dependent cladogenesis and immigration model; \code{ddmodel_sim == 2} is an
exponential cladogenesis model; \code{ddmodel_sim == 3} is a linear 
diversity-dependent extinction model.}

\item{island_type}{oceanic or non-oceanic island}

\item{x_s}{numeric vector with all the values of non-oceanic species sampling
giving the paramter space of non-oceanic species sampling (x_s is the
first element of the nonoceanic parameter vector from DAISIE_sim).}

\item{x_nonend}{numeric vector with all the values of the non-endemic
sampling giving the parameter space of the non-endemic sampling (x_nonend
is the second element of the nonoceanic parameter vector from DAISIE_sim).}
}
\value{
A list of numerics which are the relaxation half-lives.
}
\description{
Simulates an island using DAISIE_sim_core and then calculates the half-life
and repeats this across a parameter space
}
