% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_min_type2.R
\name{DAISIE_sim_min_type2}
\alias{DAISIE_sim_min_type2}
\title{Internal function of the DAISIE simulation}
\usage{
DAISIE_sim_min_type2(
  time,
  M,
  pars,
  replicates,
  prop_type2_pool,
  verbose = TRUE
)
}
\arguments{
\item{time}{Numeric defining the length of the simulation in time units.
For examples, if an island is known to be 4 million years old, setting
time = 4 will simulate the entire life span of the island; setting time = 2
will stop the simulation at the mid-life of the island.}

\item{M}{Numeric defining the size of mainland pool, i.e. the number of
species that can potentially colonize the island.}

\item{pars}{A numeric vector containing the model parameters:
\itemize{
  \item{pars[1]: lambda^c (cladogenesis rate)}
  \item{pars[2]: mu (extinction rate)}
  \item{pars[3]: K (carrying capacity), set K=Inf for diversity
  independence.}
  \item{pars[4]: gamma (immigration rate)}
  \item{pars[5]: lambda^a (anagenesis rate)}
  \item{pars[6]: lambda^c (cladogenesis rate) for either type 2 species
  or rate set 2 in rate shift model}
  \item{pars[7]: mu (extinction rate) for either type 2 species or rate
  set 2 in rate shift model}
  \item{pars[8]: K (carrying capacity) for either type 2 species or rate
  set 2 in rate shift model, set K=Inf for diversity independence.}
  \item{pars[9]: gamma (immigration rate) for either type 2 species or
  rate set 2 in rate shift model}
  \item{pars[10]: lambda^a (anagenesis rate) for either type 2 species
  or rate set 2 in rate shift model}
}
The elements 6:10 are required only when type 2 species are included
or in the rate shift model.}

\item{replicates}{Number of island replicates to be simulated.}

\item{prop_type2_pool}{Fraction of mainland species that belongs to the
second subset of species (type 2). Applies only when two types of species
are simulated (length(pars) = 10). For \code{\link{DAISIE_dataprep}} applies
only if number_clade_types = 2.  In \code{\link{DAISIE_dataprep}} the
default "proportional" sets the fraction to be proportional to the number
of clades of distinct macroevolutionary process that have colonised the
island.}

\item{verbose}{In simulation and dataprep functions a logical,
\code{Default = TRUE} gives intermediate output should be printed.
For ML functions a numeric determining if intermediate output should be
printed, \code{Default = 0} does not print, \code{verbose = 1} prints
intermediate output of the parameters and loglikelihood, \code{verbose = 2}
means also intermediate progress during loglikelihood computation is shown.}
}
\description{
Internal function of the DAISIE simulation
}
