% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{DAISIE_sim_update_state}
\alias{DAISIE_sim_update_state}
\title{Does something}
\usage{
DAISIE_sim_update_state(timeval, possible_event, maxspecID, mainland_spec,
  island_spec)
}
\arguments{
\item{timeval}{something}

\item{possible_event}{something}

\item{maxspecID}{something}

\item{mainland_spec}{something}

\item{island_spec}{something}
}
\description{
Does something
}
