\name{DAISIE_tutorial}                               
\alias{DAISIE_tutorial}
\title{Opens DAISIE tutorial}
\description{ This function opens a PDF file that contains a step-by-step tutorial on how to run DAISIE functions, using the Galapagos avian dataset from Valente, Phillimore & Etienne 2015 as an example}
\usage{
DAISIE_tutorial()
}
\details{
Opens a PDF file containing the tutorial}
\references{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015). Equilibrium and non-equilibrium dynamics simultaneously operate in the Galapagos islands. Ecology Letters 18: 844-852.}
\author{Luis Valente and Rampal S. Etienne}

\keyword{models}