% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ana_rate}
\alias{get_ana_rate}
\title{Function to calculate anagenesis rate given number of immigrant species}
\usage{
get_ana_rate(timeval, totaltime, laa, Apars, Epars, island_function_shape,
  extcutoff, island_spec, K)
}
\arguments{
\item{timeval}{something}

\item{totaltime}{something}

\item{laa}{something}

\item{Apars}{something}

\item{Epars}{something}

\item{island_function_shape}{something}

\item{extcutoff}{something}

\item{island_spec}{something}

\item{K}{something}
}
\description{
Function to calculate anagenesis rate given number of immigrant species
}
