% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ana_rate}
\alias{get_ana_rate}
\title{Calculate anagenesis rate}
\usage{
get_ana_rate(laa, hyper_pars, dist_pars, num_immigrants)
}
\arguments{
\item{laa}{per capita anagenesis rate}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations:
\itemize{
\item{[1]: is d_0 the scaling parameter for exponent for calculating
cladogenesis rate}
\item{[2]: is x the exponent for calculating extinction rate}
\item{[3]: is alpha, the exponent for calculating the immigration rate}
\item{[4]: is beta the exponent for calculating the anagenesis rate.}
}}

\item{dist_pars}{a numeric for the distance from the mainland.}

\item{num_immigrants}{a numeric with the current number of non-endemic
species (a.k.a non-endemic species).}
}
\description{
Internal function.
Calculates the anagenesis rate given the current number of
immigrant species and the per capita rate.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ana_rate}

Other rates calculation: 
\code{\link{DAISIE_plot_area}()},
\code{\link{get_ext_rate}()},
\code{\link{get_immig_rate}()},
\code{\link{island_area_vector}()},
\code{\link{island_area}()},
\code{\link{update_max_rates}()},
\code{\link{update_rates}()}
}
\author{
Pedro Neves, Joshua Lambert
}
\concept{rates calculation}
