% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ana_rate}
\alias{get_ana_rate}
\title{Calculate anagenesis rate}
\usage{
get_ana_rate(laa, hyper_pars, dist_pars, island_spec)
}
\arguments{
\item{laa}{per capita anagenesis rate}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations, \code{hyper_pars[1]} is d_0 the scaling parameter for
exponent for calculating cladogenesis rate, \code{hyper_pars[2]}
is x the exponent for calculating extinction rate,
\code{hyper_pars[3]} is alpha the exponent for calculating the
immigration rate, \code{hyper_pars[4]} is beta the exponent for
calculating the anagenesis rate.}

\item{dist_pars}{a numeric for the distance from the mainland.}

\item{island_spec}{matrix with current state of system}
}
\description{
Internal function.
Calculates the anagenesis rate given the current number of
immigrant species and the per capita rate.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ana_rate}

Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_dynamic_t_hor}},
  \code{\link{get_ext_rate}}, \code{\link{get_immig_rate}},
  \code{\link{get_t_hor}},
  \code{\link{island_area_vector}},
  \code{\link{island_area}}, \code{\link{update_rates}}
}
\author{
Pedro Neves
}
\concept{rates calculation}
