% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ana_rate}
\alias{get_ana_rate}
\title{Calculate anagenesis rate}
\usage{
get_ana_rate(laa, Tpars = NULL, island_spec)
}
\arguments{
\item{laa}{per capita anagenesis rate}

\item{Tpars}{A named list containing diversification rates considering two trait states:
\itemize{
  \item{[1]:A numeric with the per capita transition rate with state1}
  \item{[2]:A numeric with the per capita immigration rate with state2}
  \item{[3]:A numeric with the per capita extinction rate with state2}
  \item{[4]:A numeric with the per capita anagenesis rate with state2}
  \item{[5]:A numeric with the per capita cladogenesis rate with state2}
  \item{[6]:A numeric with the per capita transition rate with state2}
  \item{[7]:A numeric with the number of species with trait state 2 on mainland}
}}

\item{island_spec}{matrix with current state of system}
}
\description{
Internal function.
Calculates the anagenesis rate given the current number of
immigrant species and the per capita rate.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ana_rate}

Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ext_rate}}, \code{\link{get_immig_rate}},
  \code{\link{get_t_hor}}, \code{\link{get_trans_rate}},
  \code{\link{island_area_vector}},
  \code{\link{island_area}}, \code{\link{update_rates}}
}
\author{
Pedro Neves
}
\concept{rates calculation}
