% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ana_rate}
\alias{get_ana_rate}
\title{Calculate anagenesis rate}
\usage{
get_ana_rate(laa, island_spec)
}
\arguments{
\item{laa}{per capita anagenesis rate}

\item{island_spec}{matrix with current state of system}
}
\description{
Internal function. 
Calculates the anagenesis rate given the current number of
immigrant species and the per capita rate.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ana_rate}

Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ext_rate}}, \code{\link{get_immig_rate}},
  \code{\link{get_t_hor}}, \code{\link{island_area}},
  \code{\link{update_rates}}
}
\author{
Pedro Neves
}
\concept{rates calculation}
