% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_clado_rate}
\alias{get_clado_rate}
\title{Function to calculate cladogenesis rate given number of species on island}
\usage{
get_clado_rate(lac, island_ontogeny, island_spec, K)
}
\arguments{
\item{lac}{per capita cladogenesis rate}

\item{island_ontogeny}{a string describing the type of island ontogeny. 
Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{island_spec}{matrix with current state of system}

\item{K}{carrying capacity}
}
\description{
Function to calculate cladogenesis rate given number of species on island
}
\seealso{
Does the same as \link{DAISIE_calc_clade_clado_rate}
}
\author{
Pedro Neves
}
