% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_clado_rate}
\alias{get_clado_rate}
\title{Function to calculate cladogenesis rate given number of island species}
\usage{
get_clado_rate(timeval, totaltime, lac, Apars, Epars, island_function_shape,
  extcutoff, island_spec, K)
}
\arguments{
\item{timeval}{something}

\item{totaltime}{something}

\item{Apars}{something}

\item{Epars}{something}

\item{island_function_shape}{something}

\item{extcutoff}{something}

\item{island_spec}{something}

\item{K}{something}

\item{laa}{something}
}
\description{
Function to calculate cladogenesis rate given number of island species
}
