% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_clado_rate}
\alias{get_clado_rate}
\title{Calculate cladogenesis rate}
\usage{
get_clado_rate(lac, hyper_pars, num_spec, K, A)
}
\arguments{
\item{lac}{A numeric with the per capita cladogenesis rate.}

\item{hyper_pars}{A named list of numeric hyperparameters for the rate
calculations as returned by \code{\link{create_hyper_pars}()}:
\itemize{
  \item{[1]: is d the scaling parameter for exponent for calculating
  cladogenesis rate}
  \item{[2]: is x the exponent for calculating extinction rate}
}}

\item{num_spec}{A numeric with the current number of species.}

\item{K}{A numeric with carrying capacity.}

\item{A}{A numeric value for island area at a given point in time.}
}
\description{
Internal function.
Calculates the cladogenesis rate given the current number of
species in the system, the carrying capacity and the per capita cladogenesis
rate
}
\author{
Pedro Neves, Joshua Lambert
}
