% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_clado_rate}
\alias{get_clado_rate}
\title{Calculate cladogenesis rate}
\usage{
get_clado_rate(timeval, totaltime, lac, Apars, island_ontogeny,
  island_spec, K)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{total time of simulation}

\item{lac}{per capita cladogenesis rate}

\item{Apars}{a named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny. 
Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{island_spec}{matrix with current state of system}

\item{K}{carrying capacity}
}
\description{
Internal function. 
Calculates the cladogenesis rate given the current number of
species in the system, the carrying capacity and the per capita cladogenesis
rate
}
\references{
Valente, Luis M., Rampal S. Etienne, and Albert B. Phillimore. 
"The effects of island ontogeny on species diversity and phylogeny." 
Proceedings of the Royal Society of London B: Biological Sciences 281.1784 (2014): 20133227.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_clado_rate}
}
\author{
Pedro Neves
}
