% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_clado_rate_per_capita}
\alias{get_clado_rate_per_capita}
\title{Title}
\usage{
get_clado_rate_per_capita(lac, d, num_spec, K, A)
}
\arguments{
\item{lac}{A numeric with the per capita cladogenesis rate.}

\item{d}{Numeric defining the scaling parameter for exponent for
calculating cladogenesis rate.}

\item{num_spec}{A numeric with the current number of species.}

\item{K}{A numeric with carrying capacity.}

\item{A}{A numeric value for island area at a given point in time.}
}
\value{

}
\description{
Title
}
