% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ext_rate}
\alias{get_ext_rate}
\title{Function to describe changes in extinction rate through time. From
Valente et al 2014 ProcB}
\usage{
get_ext_rate(timeval, totaltime, mu, Apars, Epars, island_function_shape,
  extcutoff, island_spec, K)
}
\arguments{
\item{timeval}{something}

\item{totaltime}{something}

\item{mu}{something}

\item{Apars}{something}

\item{Epars}{something}

\item{island_function_shape}{something}

\item{extcutoff}{something}

\item{island_spec}{something}

\item{K}{something}
}
\description{
Function to describe changes in extinction rate through time. From
Valente et al 2014 ProcB
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ext_rate}
}
