% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ext_rate}
\alias{get_ext_rate}
\title{#' Function to describe per-capita changes in extinction rate through time
#'
#' This function is only called directly inside the RHS of the ontogeny
#' likelihood functions. In all other cases \code{\link{get_clado_rate}()} is to
#' ne called instead.
#'
#' @inheritParams default_params_doc
#'
#' @return Numeric with per capita extinction rate, given A(t), x, and mu0.
#' @keywords internal
#'
#' @examples
#' ext_rate_per_capita <- DAISIE:::get_ext_rate_per_capita(
#'   mu = 0.5,
#'   x = 1,
#'   A = 1000
#' )
get_ext_rate_per_capita <- function(mu,
                                    x,
                                    extcutoff = 1000,
                                    A = 1) {
  ext_rate_per_capita <- max(0, mu * (A ^ -x), na.rm = TRUE)
  ext_rate_per_capita <- min(ext_rate_per_capita, extcutoff, na.rm = TRUE)
  return(ext_rate_per_capita)
}}
\usage{
get_ext_rate(
  mu,
  hyper_pars,
  extcutoff = 1000,
  num_spec,
  A,
  trait_pars = NULL,
  island_spec = NULL
)
}
\description{

}
\details{

}
