% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ext_rate}
\alias{get_ext_rate}
\title{Function to describe changes in extinction rate through time. From
Valente et al 2014 ProcB}
\usage{
get_ext_rate(
  timeval,
  mu,
  ddmodel_sim = 11,
  hyper_pars = NULL,
  area_pars,
  ext_pars,
  island_ontogeny,
  sea_level = 0,
  extcutoff = 1100,
  num_spec,
  K
)
}
\arguments{
\item{timeval}{current time of simulation}

\item{mu}{per capita extinction rate in no ontogeny model}

\item{ddmodel_sim}{A numeric determining which parameters are diversity-
dependent.}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations, \code{hyper_pars[1]} is d_0 the scaling parameter for
exponent for calculating cladogenesis rate, \code{hyper_pars[2]}
is x the exponent for calculating extinction rate,
\code{hyper_pars[3]} is alpha the exponent for calculating the
immigration rate, \code{hyper_pars[4]} is beta the exponent for
calculating the anagenesis rate.}

\item{area_pars}{a named list containing area parameters as created by
create_area_pars:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{ext_pars}{a numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny.
Can be \code{NULL},
\code{"beta"} for a beta function describing area through time.}

\item{sea_level}{a numeric describing sea level can be \code{NULL}}

\item{extcutoff}{cutoff for extinction rate preventing it from being too
large and slowing down simulation. Default is 1100}

\item{K}{carrying capacity}

\item{island_spec}{matrix containing state of system}
}
\description{
Function to describe changes in extinction rate through time. From
Valente et al 2014 ProcB
}
\references{
Valente, Luis M., Rampal S. Etienne, and Albert B. Phillimore.
"The effects of island ontogeny on species diversity and phylogeny."
Proceedings of the Royal Society of London B: Biological Sciences 281.1784 (2014): 20133227.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ext_rate}

Other rates calculation: 
\code{\link{DAISIE_plot_area}()},
\code{\link{get_ana_rate}()},
\code{\link{get_global_max_area_time}()},
\code{\link{get_global_min_area_time}()},
\code{\link{get_immig_rate}()},
\code{\link{get_t_hor}()},
\code{\link{island_area_vector}()},
\code{\link{island_area}()},
\code{\link{update_max_rates}()},
\code{\link{update_rates}()}
}
\author{
Pedro Neves
}
\concept{rates calculation}
