% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_ext_rate}
\alias{get_ext_rate}
\title{Function to describe changes in extinction rate through time. From
Valente et al 2014 ProcB}
\usage{
get_ext_rate(timeval, mu, Apars, Epars, island_ontogeny, extcutoff,
  island_spec, K)
}
\arguments{
\item{timeval}{current time of simulation}

\item{mu}{per capita extinction rate in no ontogeny model}

\item{Apars}{a named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{a numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny. 
Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{extcutoff}{cutoff for extinction rate preventing it from being too 
large and slowing down simulation}

\item{island_spec}{matrix containing state of system}

\item{K}{carrying capacity}
}
\description{
Function to describe changes in extinction rate through time. From
Valente et al 2014 ProcB
}
\references{
Valente, Luis M., Rampal S. Etienne, and Albert B. Phillimore. 
"The effects of island ontogeny on species diversity and phylogeny." 
Proceedings of the Royal Society of London B: Biological Sciences 281.1784 (2014): 20133227.
}
\seealso{
Does the same as \link{DAISIE_calc_clade_ext_rate}

Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ana_rate}}, \code{\link{get_immig_rate}},
  \code{\link{get_t_hor}}, \code{\link{island_area}},
  \code{\link{update_rates}}
}
\author{
Pedro Neves
}
\concept{rates calculation}
