% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_max_rates.R
\name{get_global_max_area_time}
\alias{get_global_max_area_time}
\title{Get the time of maximum area}
\usage{
get_global_max_area_time(totaltime, area_pars, island_ontogeny, sea_level)
}
\arguments{
\item{totaltime}{stub}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{island_ontogeny}{a numeric describing the type of island ontogeny.
Can be \code{NULL}, \code{1} for a beta function describing
area through time.}

\item{sea_level}{stub}
}
\value{
Numeric with time at which area is maximum during the simulation.
}
\description{
Get the time of maximum area
}
\examples{
timeval <- 1
totaltime <- 10
area_pars <- DAISIE::create_area_pars(
  max_area = 5000,
  proportional_peak_t = 0.5,
  peak_sharpness = 1,
  total_island_age = 15,
  sea_level_amplitude = 0,
  sea_level_frequency = 0
)
island_ontogeny <- 1
sea_level <- 0

testthat::expect_silent(
  global_max_area_time <- DAISIE:::get_global_max_area_time(
    totaltime = totaltime,
    area_pars = area_pars,
    island_ontogeny = island_ontogeny,
    sea_level = sea_level
  )
)

}
\author{
Pedro Neves, Joshua Lambert, Shu Xie
}
