% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_max_rates.R
\name{get_global_min_area_time}
\alias{get_global_min_area_time}
\title{Get the time when area is minimum}
\usage{
get_global_min_area_time(totaltime, area_pars, island_ontogeny, sea_level)
}
\arguments{
\item{totaltime}{Numeric defining the length of the simulation in time units.}

\item{area_pars}{A named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
  \item{[7]: angle of the slope of the island}
}}

\item{island_ontogeny}{In \code{\link{DAISIE_sim_time_dependent}},
\code{\link{DAISIE_ML_CS}} and plotting a string describing the type of
island ontogeny. Can be \code{"const"}, \code{"beta"} for a beta function
describing area through time. String checked by
\code{\link{is_island_ontogeny_input}}. \cr In all other functions a numeric
describing the type of island ontogeny. Can be \code{0} for constant,
\code{1} for a beta function describing area through time. In ML functions
\code{island_ontogeny = NA} assumes constant ontogeny.}

\item{sea_level}{In \code{\link{DAISIE_sim_time_dependent}} and plotting a
string describing the type of sea level. Can be \code{"const"} or
\code{"sine"} for a sine function describing area through time. String
checked by \code{\link{is_sea_level_input}}.
\cr In all other functions a numeric describing the type of sea level. Can
be \code{0} for constant, \code{1} for a sine function describing area
through time.}
}
\value{
Numeric with time at which area is minimum during the simulation
}
\description{
Get the time when area is minimum
}
\examples{
timeval <- 1
totaltime <- 10
area_pars <- DAISIE::create_area_pars(
  max_area = 5000,
  proportional_peak_t = 0.5,
  peak_sharpness = 1,
  total_island_age = 15,
  sea_level_amplitude = 0,
  sea_level_frequency = 0,
   island_gradient_angle = 0
)
island_ontogeny <- 1
sea_level <- 0

testthat::expect_silent(
  DAISIE:::get_global_min_area_time(
    totaltime = totaltime,
    area_pars = area_pars,
    island_ontogeny = island_ontogeny,
    sea_level = sea_level
  )
)

}
\author{
Pedro Neves, Joshua Lambert, Shu Xie
}
