% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_immig_rate}
\alias{get_immig_rate}
\title{Something}
\usage{
get_immig_rate(timeval, totaltime, gam, Apars, Epars, island_function_shape,
  extcutoff, island_spec, K, mainland_n)
}
\arguments{
\item{timeval}{something}

\item{totaltime}{something}

\item{gam}{something}

\item{Apars}{something}

\item{Epars}{something}

\item{island_function_shape}{something}

\item{extcutoff}{something}

\item{island_spec}{something}

\item{K}{something}

\item{mainland_n}{something}
}
\description{
Something
}
