% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_immig_rate}
\alias{get_immig_rate}
\title{Function to calculate immigration rate given number of species on island
 and mainland}
\usage{
get_immig_rate(gam, island_spec, K, mainland_n)
}
\arguments{
\item{gam}{per capita immigration rate}

\item{island_spec}{matrix with current state of system}

\item{K}{carrying capacity}

\item{mainland_n}{total number of species present in the mainland}
}
\description{
Function to calculate immigration rate given number of species on island
 and mainland
}
\seealso{
Does the same as \link{DAISIE_calc_clade_imm_rate}

Other rates calculation: \code{\link{get_ana_rate}},
  \code{\link{get_ext_rate}}, \code{\link{get_thor}},
  \code{\link{island_area}}, \code{\link{update_rates}}
}
\author{
Pedro Neves
}
