% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_t_hor}
\alias{get_t_hor}
\title{Function to calculate and update horizon for maximum extinction rate}
\usage{
get_t_hor(timeval, totaltime, Tpars, Apars, ext, ext_multiplier,
  island_ontogeny, t_hor)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{total time of simulation}

\item{Tpars}{A named list containing diversification rates considering two trait states:
\itemize{
  \item{[1]:A numeric with the per capita transition rate with state1}
  \item{[2]:A numeric with the per capita immigration rate with state2}
  \item{[3]:A numeric with the per capita extinction rate with state2}
  \item{[4]:A numeric with the per capita anagenesis rate with state2}
  \item{[5]:A numeric with the per capita cladogenesis rate with state2}
  \item{[6]:A numeric with the per capita transition rate with state2}
  \item{[7]:A numeric with the number of species with trait state 2 on mainland}
}}

\item{Apars}{a named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{ext}{effective extinction rate at timeval}

\item{ext_multiplier}{reduces or increases distance of horizon to current
simulation time}

\item{island_ontogeny}{a string describing the type of island ontogeny.
Can be \code{NULL}, \code{"beta"} for a beta function
 describing area through time, or \code{"linear"} for a linear function}

\item{t_hor}{time of horizon for max extinction}
}
\description{
Internal function.
Calculates when the next horizon for maximum extinction will be in the
simulation
}
\seealso{
Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ana_rate}}, \code{\link{get_ext_rate}},
  \code{\link{get_immig_rate}},
  \code{\link{get_trans_rate}},
  \code{\link{island_area_vector}},
  \code{\link{island_area}}, \code{\link{update_rates}}
}
\author{
Pedro Neves
}
\concept{rates calculation}
