% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_thor}
\alias{get_thor}
\title{Function to calculate and update horizon for maximum extinction rate}
\usage{
get_thor(timeval, totaltime, Apars, ext_multiplier, island_ontogeny, thor)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{total time of simulation}

\item{Apars}{a numeric vector:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{ext_multiplier}{reduces or increases distance of horizon to current
simulation time}

\item{island_ontogeny}{a string describing the type of island ontogeny.
Can be \code{NULL}, \code{"quadratic"} for a beta function
 describing area through time, or \code{"linear"} for a linear function}

\item{thor}{time of horizon for max extinction}
}
\description{
Function to calculate and update horizon for maximum extinction rate
}
\seealso{
Other rates calculation: \code{\link{get_ana_rate}},
  \code{\link{get_ext_rate}}, \code{\link{get_immig_rate}},
  \code{\link{island_area}}, \code{\link{update_rates}}
}
\author{
Pedro Neves
}
