% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_thor}
\alias{get_thor}
\title{Something}
\usage{
get_thor(timeval, totaltime, Apars, ext_multiplier, island_ontogeny, thor)
}
\arguments{
\item{timeval}{something}

\item{totaltime}{something}

\item{Apars}{something}

\item{ext_multiplier}{something}

\item{island_ontogeny}{something}

\item{thor}{something}
}
\description{
Something
}
