% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{get_trans_rate}
\alias{get_trans_rate}
\title{Calculate transition rate}
\usage{
get_trans_rate(trait_pars, island_spec)
}
\arguments{
\item{trait_pars}{A named list containing diversification rates considering 
two trait states created by \code{\link{create_trait_pars}}:
\itemize{
  \item{[1]:A numeric with the per capita transition rate with state1}
  \item{[2]:A numeric with the per capita immigration rate with state2}
  \item{[3]:A numeric with the per capita extinction rate with state2}
  \item{[4]:A numeric with the per capita anagenesis rate with state2}
  \item{[5]:A numeric with the per capita cladogenesis rate with state2}
  \item{[6]:A numeric with the per capita transition rate with state2} 
  \item{[7]:A numeric with the number of species with trait state 2 on mainland} 
}}

\item{island_spec}{Matrix with current state of simulation containing number
of species.}
}
\description{
Internal function.
Calculates the transition rate given the current number of
immigrant species and the per capita rate.
}
\concept{rates calculation}
