% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_loglik_integrate.R
\name{integral_peak}
\alias{integral_peak}
\title{Computes integral of a very peaked function, modified from the SADISA package}
\usage{
integral_peak(
  logfun,
  optimmethod = "optimize",
  xx = seq(-20, 20, 2),
  xcutoff = 2,
  ymaxthreshold = 1e-12,
  ...
)
}
\arguments{
\item{logfun}{the logarithm of the function to integrate}

\item{optimmethod}{method used to find the optimum of the integrand. Options are 'subplex', 'optim' or 'Bart'.
Default is 'subplex'. If 'Bart' the following arguments apply.}

\item{xx}{the initial set of points on which to evaluate the function}

\item{xcutoff}{when the maximum has been found among the xx, this parameter sets the width of the interval to find the maximum in}

\item{ymaxthreshold}{sets the deviation allowed in finding the maximum among the xx}

\item{...}{any arguments of the function to optimize}
}
\value{
the result of the integration
}
\description{
# computes the logarithm of the integral of exp(logfun) from 0 to Inf under the following assumptions:
}
\references{
Haegeman, B. & R.S. Etienne (2017). A general sampling formula for community structure data. Methods in Ecology & Evolution. In press.
}
