% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{is.odd}
\alias{is.odd}
\title{Determine if the value is odd.}
\usage{
is.odd(x)
}
\arguments{
\item{x}{Object to determine}
}
\value{
Boolean indicating if object is odd
}
\description{
Determine if the value is odd.
}
\examples{
  testit::assert(
    DAISIE:::is.odd(
      x = 0
    ) == FALSE
  )
  
  testit::assert(
    DAISIE:::is.odd(
      x = 1
    ) == TRUE
  )
}
