% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_island_ontogeny.R
\name{is_island_ontogeny_input}
\alias{is_island_ontogeny_input}
\title{Check if island_ontogeny is correct after user input}
\usage{
is_island_ontogeny_input(island_ontogeny)
}
\arguments{
\item{island_ontogeny}{Character string that can be \code{"const"},
\code{"linear"} or \code{"beta"} depending on type of island ontogeny desired}
}
\value{
Boolean stating if island_ontogeny is correct.
}
\description{
Check if island_ontogeny is correct after user input
}
\seealso{
is_island_ontogeny_runtime
}
