% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{is_numeric_list}
\alias{is_numeric_list}
\title{Determine if list has only numerical values.}
\usage{
is_numeric_list(x)
}
\arguments{
\item{x}{Object to determine}
}
\description{
Determine if list has only numerical values.
}
\note{
do not forget: NAs are removed from a list!
}
