% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_check_format.R
\name{is_sea_level_input}
\alias{is_sea_level_input}
\title{Check if sea_level is correct after user input}
\usage{
is_sea_level_input(sea_level)
}
\arguments{
\item{sea_level}{In \code{\link{DAISIE_sim_time_dependent}} and plotting a
string describing the type of sea level. Can be \code{"const"} or
\code{"sine"} for a sine function describing area through time. String
checked by \code{\link{is_sea_level_input}}.
\cr In all other functions a numeric describing the type of sea level. Can
be \code{0} for constant, \code{1} for a sine function describing area
through time.}
}
\value{
Boolean stating if sea_level is correct.
}
\description{
Check if sea_level is correct after user input
}
\seealso{
is_sea_level_runtime
}
