% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_check_format.R
\name{is_simulation_outputs}
\alias{is_simulation_outputs}
\title{Measures if the input is a valid collection of simulation
outputs.}
\usage{
is_simulation_outputs(simulation_outputs)
}
\arguments{
\item{simulation_outputs}{A list with matrices and vectors of simulation 
produced by DAISIE_sim.}
}
\value{
TRUE if the input is a valid collection of simulation
outputs.
}
\description{
Measures if the input is a valid collection of simulation
outputs.
}
\examples{
library(testthat)
 
expect_false(is_simulation_outputs("nonsense")) 

simulation_outputs <- create_test_simulation_outputs() 
expect_true(is_simulation_outputs(simulation_outputs))
}
\author{
Richel J.C Bilderbeek, Pedro Neves
}
