% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{island_area}
\alias{island_area}
\title{Function to describe changes in area through time. Adapted from
Valente et al 2014 ProcB}
\usage{
island_area(timeval, totaltime, Apars, island_function_shape)
}
\arguments{
\item{timeval}{something}

\item{totaltime}{something}

\item{Apars}{something}

\item{island_function_shape}{something}
}
\description{
Function to describe changes in area through time. Adapted from
Valente et al 2014 ProcB
}
