% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{island_area}
\alias{island_area}
\title{Function to describe changes in area through time. Adapted from
Valente et al 2014 ProcB}
\usage{
island_area(timeval, area_pars, island_ontogeny, sea_level)
}
\arguments{
\item{timeval}{current time of simulation}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny.
Can be \code{NULL}, \code{"beta"} for a beta function describing area
through time.}

\item{sea_level}{a numeric describing the type of sea level.}
}
\description{
Function to describe changes in area through time. Adapted from
Valente et al 2014 ProcB
}
\references{
Valente, Luis M., Rampal S. Etienne, and Albert B. Phillimore.
"The effects of island ontogeny on species diversity and phylogeny."
Proceedings of the Royal Society of London B: Biological
Sciences 281.1784 (2014): 20133227.
}
\seealso{
Other rate calculations: 
\code{\link{DAISIE_plot_area}()},
\code{\link{get_ana_rate}()},
\code{\link{get_ext_rate}()},
\code{\link{get_immig_rate}()},
\code{\link{island_area_vector}()},
\code{\link{update_max_rates}()},
\code{\link{update_rates}()}
}
\author{
Pedro Neves, Joshua Lambert
}
\concept{rate calculations}
