% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{land_bridge_periods}
\alias{land_bridge_periods}
\title{Determines which rate set to use in the shift-rates simulation}
\usage{
land_bridge_periods(timeval, totaltime, shift_times)
}
\arguments{
\item{timeval}{numeric current time during a simulation.}

\item{shift_times}{a vector of numerics with times of shifts,
times are time back from the present.}
}
\value{
Boolean
}
\description{
Determines which rate set to use in the shift-rates simulation
}
\examples{
periods <- DAISIE:::land_bridge_periods(timeval = 0.5,
                              totaltime = 10,
                              shift_times = c(3, 6))
}
