% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_data.R
\docType{data}
\name{made_up_datalist}
\alias{made_up_datalist}
\title{Colonization and branching times of artificial data in list format
accepted by DAISIE_ML and DAISIE_loglik_all}
\format{
A list with 11 elements.
}
\source{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015). Equilibrium
and non-equilibrium dynamics simultaneously operate in the Galapagos
islands. Ecology Letters 18: 844-852.
}
\description{
A list containing artificial data in the same format as the
empirical data.
}
\keyword{datasets}
