% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_ext_test.R, R/DAISIE_sim_core.R
\name{pick_timeval}
\alias{pick_timeval}
\alias{pick_timeval}
\title{Something}
\usage{
pick_timeval(rates, timeval)

pick_timeval(rates, timeval)
}
\arguments{
\item{rates}{something}

\item{timeval}{something}

\item{rates}{something}

\item{timeval}{something}
}
\description{
Something

Something
}
