% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{sample_relaxed_rate}
\alias{sample_relaxed_rate}
\title{Samples from distribution when parameter is relaxed}
\usage{
sample_relaxed_rate(pars, relaxed_par)
}
\arguments{
\item{pars}{A vector of 5 elements with the model parameters}

\item{relaxed_par}{A string determining which parameter is relaxed}
}
\value{
A vector of parameters.
}
\description{
Samples from distribution when parameter is relaxed
}
\keyword{internal}
