% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAISIE_loglik_tree.R
\name{TRAISIE_loglik_cpp_tree}
\alias{TRAISIE_loglik_cpp_tree}
\title{Using hidden traits}
\usage{
TRAISIE_loglik_cpp_tree(
  parameter,
  phy,
  traits,
  sampling_fraction,
  num_hidden_states,
  mainland = FALSE,
  trait_mainland_ancestor = NULL,
  atol = 1e-15,
  rtol = 1e-15,
  method = "odeint::runge_kutta_cash_karp54",
  use_normalization = TRUE,
  num_threads = 1
)
}
\description{
Likelihood calculation including hidden traits
}
