% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAISIE_loglik_tree.R
\name{TRAISIE_loglik_cpp_tree}
\alias{TRAISIE_loglik_cpp_tree}
\title{Using hidden traits}
\usage{
TRAISIE_loglik_cpp_tree(
  parameter,
  phy,
  traits,
  sampling_fraction,
  num_hidden_states,
  mainland = FALSE,
  trait_mainland_ancestor = NULL,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "odeint::runge_kutta_cash_karp54",
  use_normalization = TRUE,
  num_threads = 1
)
}
\arguments{
\item{parameter}{vector of parameters in the following order: lambda_c, mu, gamma,
lambda_a, q and p.}

\item{traits}{vector with trait states for each tip in the phylogeny. The
order of the states must be the same as the tree tips.}

\item{sampling_fraction}{vector that states the sampling proportion per
observed trait state. It must have as many elements as there are trait states.}

\item{num_hidden_states}{The number of hidden trait states.}

\item{trait_mainland_ancestor}{This is the trait of the mainland ancestor of
a given lineage, that is the trait of the species at the stem of the tree.
In the data, this is given as a vector. For example, for data with a trait
with three observed states, it the trait of the mainland ancestor of one of
the lineages is 0,  trait_mainland_ancestor <- c(1,0,0) for that lineage}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{Method of the ODE-solver. Supported Boost \code{ODEINT}
solvers (steppers) are:
\code{"odeint::runge_kutta_cash_karp54"}
\code{"odeint::runge_kutta_fehlberg78"}
\code{"odeint::runge_kutta_dopri5"}
\code{"odeint::bulirsch_stoer"}
without \code{odeint::}-prefix, \code{\link[deSolve]{ode}} method is
assumed. The default method overall is
\code{"lsodes"} for \code{\link{DAISIE_ML_CS}()}
and \code{"ode45"} from \code{\link[deSolve]{ode}()} for
\code{\link{DAISIE_ML_IW}()}.}

\item{num_threads}{number of threads to be used. Default is one thread.}
}
\description{
Likelihood calculation including hidden traits
}
\keyword{internal}
