% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAISIE_loglik_tree.R
\name{TRAISIE_loglik_R_tree}
\alias{TRAISIE_loglik_R_tree}
\title{Using hidden traits}
\usage{
TRAISIE_loglik_R_tree(
  parameter,
  phy,
  traits,
  sampling_fraction,
  num_hidden_states,
  mainland = FALSE,
  trait_mainland_ancestor = NULL,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rhs_func = loglik_hidden_rhs
)
}
\arguments{
\item{num_hidden_states}{The number of hidden trait states.}

\item{atol}{A numeric specifying the absolute tolerance of integration.}

\item{rtol}{A numeric specifying the relative tolerance of integration.}

\item{methode}{Method of the ODE-solver. Supported Boost \code{ODEINT}
solvers (steppers) are:
\code{"odeint::runge_kutta_cash_karp54"}
\code{"odeint::runge_kutta_fehlberg78"}
\code{"odeint::runge_kutta_dopri5"}
\code{"odeint::bulirsch_stoer"}
without \code{odeint::}-prefix, \code{\link[deSolve]{ode}} method is
assumed. The default method overall is
\code{"lsodes"} for \code{\link{DAISIE_ML_CS}()}
and \code{"ode45"} from \code{\link[deSolve]{ode}()} for
\code{\link{DAISIE_ML_IW}()}.}

\item{rhs_func}{ll function}
}
\description{
Likelihood calculation including hidden traits
}
\keyword{internal}
