% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAISIE_logpEC.R
\name{TRAISIE_logpEC}
\alias{TRAISIE_logpEC}
\title{Testing function for comparison with DAISIE}
\usage{
TRAISIE_logpEC(
  datalist,
  brts,
  parameter,
  phy,
  traits,
  num_observed_states,
  num_hidden_states,
  trait_mainland_ancestor = NA,
  status,
  sampling_fraction,
  num_threads = 5,
  atol = 1e-15,
  rtol = 1e-15,
  methode = "ode45",
  rcpp_methode = "odeint::runge_kutta_cash_karp54",
  use_Rcpp = 2
)
}
\arguments{
\item{datalist}{Data object containing information on colonisation and
  branching times. This object can be generated using the DAISIE_dataprep
  function, which converts a user-specified data table into a data object,
  but the object can of course also be entered directly.
  It is an R list object with the following elements.\cr
  The first element of the list has two or three components: \cr \cr
  \code{$island_age} - the island age \cr
  Then, depending on whether a distinction between types is
  made, we have:\cr
 \code{$not_present} - the number of mainland lineages
  that are not present on the island \cr
  or:\cr
 \code{$not_present_type1} -
  the number of mainland lineages of type 1 that are not present on the
  island \cr
 \code{$not_present_type2} - the number of mainland lineages of
  type 2 that are not present on the island \cr \cr
 The remaining elements of
  the list each contains information on a single colonist lineage on the
  island and has 5 components:\cr \cr
 \code{$colonist_name} - the name of the
  species or clade that colonized the island \cr
 \code{$branching_times} -
  island age followed by stem age of the population/species in the case of
  Non-endemic, Non-endemic_MaxAge species and Endemic species with no close relatives
  on the island. For endemic clades with more than one species on the island
  (cladogenetic clades/ radiations) these should be island age followed by the
  branching times of the island clade including the stem age of the clade\cr
 \code{$stac} - the status of the colonist \cr \cr
- Non_endemic_MaxAge: 1 \cr
- Endemic: 2 \cr
- Endemic&Non_Endemic: 3 \cr
- Non_Endemic: 4 \cr
- Endemic_Singleton_MaxAge: 5 \cr
- Endemic_Clade_MaxAge: 6 \cr
- Endemic&Non_Endemic_Clade_MaxAge: 7 \cr
- Non_endemic_MaxAge_MinAge: 8 \cr
- Endemic_Singleton_MaxAge_MinAge: 9 \cr \cr
 \code{$missing_species} - number of island species that were not
  sampled for particular clade (only applicable for endemic clades) \cr
 \code{$type1or2} - whether the colonist belongs to type 1 or type 2 \cr}

\item{brts}{Numeric vector of branching times}

\item{methode}{Method of the ODE-solver. Supported Boost \code{ODEINT}
solvers (steppers) are:
\code{"odeint::runge_kutta_cash_karp54"}
\code{"odeint::runge_kutta_fehlberg78"}
\code{"odeint::runge_kutta_dopri5"}
\code{"odeint::bulirsch_stoer"}
without \code{odeint::}-prefix, \code{\link[deSolve]{ode}} method is
assumed. The default method overall is
\code{"lsodes"} for \code{\link{DAISIE_ML_CS}()}
and \code{"ode45"} from \code{\link[deSolve]{ode}()} for
\code{\link{DAISIE_ML_IW}()}.}
}
\description{
This function calculates the likelihood of observing a clade with specified species trait states,
given known colonization time. It is designed for comparison with DAISIE-based models.
}
\examples{
library(DAISIE)
data("Galapagos_datalist")
datalist <- Galapagos_datalist
i <- 5
phy <- DDD::brts2phylo(datalist[[i]]$branching_times[-c(1, 2)])
brts <- datalist[[i]]$branching_times
traits <- sample(c(1, 0), length(brts), replace = TRUE)
num_observed_states     =  2
num_hidden_states       =  2
sampling_fraction       =  sample(c(1, 1), num_observed_states, replace = TRUE)
trait_mainland_ancestor = c(1, 0)
datalist[[1]]$Mainland_pool_sizes <- c(550, 250)
datalist[[1]]$M <- 1000

parameter <- list(
  c(2.546591, 1.2, 1, 0.2),
  c(2.678781, 2, 1.9, 3),
  c(0.009326754, 0.003, 0.002, 0.2),
  c(1.008583, 1, 2, 1.5),
  matrix(c(
    0,    .001,    0.005,  0,
    .001,    0,    0.000,0.005,
    0.015,    000,    0,  0.005,
    0,   0.0025,  0.005,0.00
  ), nrow = 4, byrow = TRUE),
  1
)
  status                  = 2
TRAISIE_logpEC(
  datalist              = datalist,
  brts                    = brts,
  phy                     = phy,
  traits                  = traits,
  status                  = 2,
  sampling_fraction       = sampling_fraction,
  parameter               = parameter,
  trait_mainland_ancestor = c(1,0),
  num_observed_states     = 2,
  num_hidden_states       = 2,
  atol                    = 1e-15,
  rtol                    = 1e-15,
  methode                 = "ode45",
  num_threads             = 8,
  use_Rcpp                = 2)
}
