% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRAISIE_solve_branch.R
\name{TRAISIE_solve_branch}
\alias{TRAISIE_solve_branch}
\title{branch solving for TRAISIE}
\usage{
TRAISIE_solve_branch(
  interval_func,
  initial_conditions,
  time,
  parameter,
  trait_mainland_ancestor,
  methode = "ode45",
  rcpp_methode = "odeint::runge_kutta_cash_karp54",
  atol,
  rtol,
  use_Rcpp = 0
)
}
\arguments{
\item{interval_func}{chosen function for interval, can also be string if
using Rcpp}

\item{initial_conditions}{vector of initial conditions}

\item{time}{vector with two time points}

\item{methode}{Method of the ODE-solver. Supported Boost \code{ODEINT}
solvers (steppers) are:
\code{"odeint::runge_kutta_cash_karp54"}
\code{"odeint::runge_kutta_fehlberg78"}
\code{"odeint::runge_kutta_dopri5"}
\code{"odeint::bulirsch_stoer"}
without \code{odeint::}-prefix, \code{\link[deSolve]{ode}} method is
assumed. The default method overall is
\code{"lsodes"} for \code{\link{DAISIE_ML_CS}()}
and \code{"ode45"} from \code{\link[deSolve]{ode}()} for
\code{\link{DAISIE_ML_IW}()}.}
}
\description{
solve along branch
}
