% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{translate_island_ontogeny}
\alias{translate_island_ontogeny}
\title{Translate user-friendly ontogeny codes to numerics}
\usage{
translate_island_ontogeny(island_ontogeny)
}
\arguments{
\item{island_ontogeny}{a string describing the type of island ontogeny.
Can be \code{"const"}, \code{"beta"} for a beta function describing area
through time.}
}
\value{
Numeric, 0 for null-ontogeny, 1 for beta function
}
\description{
This function simulates islands with given cladogenesis,
extinction, Kprime, immigration and anagenesis parameters. If a single
parameter set is provided (5 parameters) it simulates islands where all
species have the same macro-evolutionary process. If two paramater sets
(10 parameters) are provided, it simulates islands where two different
macro-evolutionary processes operate, one applying to type 1 species and
other to type 2 species. If two parameter sets and a time shift (11
parameters) are provided, it simulates islands where at the given time
a shift between the parameter sets will occur.

Returns R list object that contains the simulated islands
}
\examples{
translate_island_ontogeny("const")
}
\references{
Valente, L.M., A.B. Phillimore and R.S. Etienne (2015).
Equilibrium and non-equilibrium dynamics simultaneously operate in the
Galapagos islands. Ecology Letters 18: 844-852.
Hauffe, T., D. Delicado, R.S. Etienne and L. Valente (submitted).
Lake expansion increases equilibrium diversity via the target effect of
island biogeography.
}
\seealso{
\code{\link{DAISIE_format_CS}} \code{\link{DAISIE_plot_sims}}
}
\author{
Luis Valente and Albert Phillimore
}
