% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{translate_island_ontogeny}
\alias{translate_island_ontogeny}
\title{Translate user-friendly ontogeny codes to numerics}
\usage{
translate_island_ontogeny(island_ontogeny)
}
\value{
Numeric, 0 for null-ontogeny, 1 for beta function
}
\description{
Translate user-friendly ontogeny codes to numerics
}
\examples{
translate_island_ontogeny("const")
}
