% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_utils.R
\name{translate_sea_level}
\alias{translate_sea_level}
\title{Translate user-friendly sea-level codes to numerics}
\usage{
translate_sea_level(sea_level)
}
\arguments{
\item{sea_level}{In \code{\link{DAISIE_sim_time_dep}()} and plotting a
string describing the type of sea level. Can be \code{"const"} or
\code{"sine"} for a sine function describing area through time.
\cr In all other functions a numeric describing the type of sea level. Can
be \code{0} for constant, \code{1} for a sine function describing area
through time.}
}
\value{
Numeric, 0 for null-sea-level, 1 for sine function
}
\description{
Translate user-friendly sea-level codes to numerics
}
\examples{
translated_sea_level <- DAISIE:::translate_sea_level("const")
}
\keyword{internal}
