% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_max_rates.R
\name{update_max_rates}
\alias{update_max_rates}
\title{Title}
\usage{
update_max_rates(
  timeval,
  totaltime,
  gam,
  mu,
  laa,
  lac,
  ddmodel_sim,
  hyper_pars = NULL,
  area_pars,
  dist_pars = NULL,
  ext_pars = NULL,
  island_ontogeny = NULL,
  sea_level = NULL,
  extcutoff,
  K,
  num_spec,
  num_immigrants,
  mainland_n
)
}
\arguments{
\item{mainland_n}{}
}
\value{

}
\description{
Title
}
