% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_max_rates.R
\name{update_max_rates}
\alias{update_max_rates}
\title{Calculates the maximum rates for a Gillespie simulation}
\usage{
update_max_rates(
  timeval,
  totaltime,
  gam,
  mu,
  laa,
  lac,
  ddmodel_sim,
  hyper_pars = NULL,
  area_pars,
  dist_pars = NULL,
  ext_pars = NULL,
  island_ontogeny = NULL,
  sea_level = NULL,
  extcutoff,
  K,
  num_spec,
  num_immigrants,
  mainland_n
)
}
\arguments{
\item{timeval}{A numeric with the current time of simulation}

\item{totaltime}{A numeric with the total time of simulation}

\item{gam}{A numeric with the per capita immigration rate}

\item{mu}{A numeric with the per capita extinction rate}

\item{laa}{A numeric with the per capita anagenesis rate}

\item{lac}{A numeric with the per capita cladogenesis rate}

\item{ddmodel_sim}{A numeric determining which parameters are diversity-
dependent.}

\item{hyper_pars}{A numeric vector for hyperparameters for the rate
calculations, \code{hyper_pars[1]} is d_0 the scaling parameter for
exponent for calculating cladogenesis rate, \code{hyper_pars[2]}
is x the exponent for calculating extinction rate,
\code{hyper_pars[3]} is alpha the exponent for calculating the
immigration rate, \code{hyper_pars[4]} is beta the exponent for
calculating the anagenesis rate.}

\item{area_pars}{a named list containing area and sea level parameters as
created by \code{\link{create_area_pars}}:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
  \item{[5]: amplitude of area fluctuation from sea level}
  \item{[6]: frequency of sine wave of area change from sea level}
}}

\item{dist_pars}{stub}

\item{ext_pars}{stub}

\item{island_ontogeny}{stub}

\item{sea_level}{stub}

\item{extcutoff}{stub}

\item{K}{a numeric with carrying capacity}

\item{num_spec}{a numeric with the current number of species.}

\item{num_immigrants}{a numeric with the current number of non-endemic
species (a.k.a non-endemic species).}

\item{mainland_n}{stub}
}
\value{
stub
}
\description{
Internal function that updates the all the max rates at time t.
}
\seealso{
Other rates calculation: 
\code{\link{DAISIE_plot_area}()},
\code{\link{get_ana_rate}()},
\code{\link{get_ext_rate}()},
\code{\link{get_immig_rate}()},
\code{\link{island_area_vector}()},
\code{\link{island_area}()},
\code{\link{update_rates}()}
}
\concept{rates calculation}
