% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{update_rates}
\alias{update_rates}
\title{Calculates algorithm rates}
\usage{
update_rates(timeval, totaltime, gam, mu, laa, lac, Apars, Epars,
  island_ontogeny, extcutoff, K, island_spec, mainland_n, thor_ext, thor_c_i)
}
\arguments{
\item{timeval}{current time of simulation}

\item{totaltime}{total time of simulation}

\item{gam}{per capita immigration rate}

\item{mu}{per capita extinction rate in no ontogeny model}

\item{laa}{per capita anagenesis rate}

\item{lac}{per capita cladogenesis rate}

\item{Apars}{a numeric vector:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{a numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{island_ontogeny}{a string describing the type of island ontogeny. 
Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{extcutoff}{cutoff for extinction rate preventing it from being too 
large and slowing down simulation. Should be big}

\item{K}{carrying capacity}

\item{island_spec}{matrix containing state of system}

\item{mainland_n}{total number of species present in the mainland}

\item{thor_ext}{time of horizon for max extinction}

\item{thor_c_i}{time of horizon for max cladogenesis and immigration}
}
\description{
Internal function that updates the all the rates and 
max extinction horizon at time t.
}
\seealso{
Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ana_rate}}, \code{\link{get_ext_rate}},
  \code{\link{get_immig_rate}}, \code{\link{get_thor}},
  \code{\link{island_area}}
}
