% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_sim_core.R
\name{update_rates}
\alias{update_rates}
\title{Something}
\usage{
update_rates(timeval, totaltime, gam, mu, laa, lac, Apars, Epars,
  island_ontogeny, extcutoff, K, island_spec, mainland_n, thor)
}
\arguments{
\item{timeval}{something}

\item{totaltime}{something}

\item{gam}{something}

\item{mu}{something}

\item{laa}{something}

\item{lac}{something}

\item{Apars}{something}

\item{Epars}{something}

\item{island_ontogeny}{something}

\item{extcutoff}{something}

\item{K}{something}

\item{island_spec}{something}

\item{mainland_n}{something}

\item{thor}{something}
}
\description{
Something
}
