% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAISIE_rates.R
\name{update_rates}
\alias{update_rates}
\title{Calculates algorithm rates}
\usage{
update_rates(timeval, totaltime, gam, mu, laa, lac, Apars = NULL,
  Epars = NULL, island_ontogeny = NULL, extcutoff, K, island_spec,
  mainland_n, t_hor = NULL)
}
\arguments{
\item{timeval}{A numeric with the current time of simulation}

\item{totaltime}{A numeric with the total time of simulation}

\item{gam}{A numeric with the per capita immigration rate}

\item{mu}{A numeric with the per capita extinction rate in no ontogeny model}

\item{laa}{A numeric with the per capita anagenesis rate}

\item{lac}{A numeric with the per capita cladogenesis rate}

\item{Apars}{A named list containing area parameters as created by create_area_params:
\itemize{
  \item{[1]: maximum area}
  \item{[2]: value from 0 to 1 indicating where in the island's history the 
  peak area is achieved}
  \item{[3]: sharpness of peak}
  \item{[4]: total island age}
}}

\item{Epars}{A numeric vector:
\itemize{
  \item{[1]: minimum extinction when area is at peak}
  \item{[2]: extinction rate when current area is 0.10 of maximum area}
}}

\item{island_ontogeny}{A string describing the type of island ontogeny. 
Can be \code{NULL},
\code{"quadratic"} for a beta function describing area through time,
 or \code{"linear"} for a linear function}

\item{extcutoff}{A numeric with the cutoff for extinction rate preventing it from being too 
large and slowing down simulation. Should be big.}

\item{K}{A numeric with K (clade-specific carrying capacity)}

\item{island_spec}{A matrix containing state of system}

\item{mainland_n}{A numeirc with the total number of species present in the mainland}

\item{t_hor}{A numeric with the time of horizon for max cladogenesis, immigration and minimum extinction}
}
\description{
Internal function that updates the all the rates and 
max extinction horizon at time t.
}
\seealso{
Other rates calculation: \code{\link{DAISIE_plot_area}},
  \code{\link{get_ana_rate}}, \code{\link{get_ext_rate}},
  \code{\link{get_immig_rate}}, \code{\link{get_t_hor}},
  \code{\link{island_area}}
}
\concept{rates calculation}
